\name{adp-class}

\docType{class}

\alias{adp-class}
\alias{[[,adp-method}

\title{Class to store acoustic Doppler profiler data}

\description{Class to store acoustic Doppler proviler data, with standard
    slots \code{metadata}, \code{data} and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{adp[["u1"]]}, where the
    string could also be \code{"u2"} or \code{"u3"}, (FILL IN FOR MORE).}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

