\name{air.hal}
\docType{data}
\alias{air.hal}
\title{Time-series of wind and temperature at Halifax, Canada.}
\description{A sample air dataset in the form of a data frame with columns
  containing \code{t}, 
  containing \code{u}, 
  containing \code{v}, and
  containing \code{temperature}.}
\usage{data(air.hal)}
\author{Dan Kelley}
\keyword{misc}
