\name{airHalifax}
\docType{data}
\alias{airHalifax}
\title{Time-series of wind and temperature at Halifax, Canada.}
\description{A sample air dataset in the form of a data frame with columns
  containing \code{t}, 
  containing \code{u}, 
  containing \code{v}, and
  containing \code{temperature}.}
\usage{data(airHalifax)}
\author{Dan Kelley}
\keyword{datasets}
