% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.R
\name{applyMagneticDeclination,adv-method}
\alias{applyMagneticDeclination,adv-method}
\title{Alter an adv-class object to account for magnetic declination}
\usage{
\S4method{applyMagneticDeclination}{adv}(
  object = "oce",
  declination = 0,
  debug = getOption("oceDebug")
)
}
\arguments{
\item{object}{an \linkS4class{adv} object.}

\item{declination}{numeric value holding magnetic declination in degrees,
positive for clockwise from north.}

\item{debug}{a debugging flag, set to a positive value to get debugging.}
}
\value{
An \linkS4class{adv} object, modified as explained in \sQuote{Description}.
}
\description{
Acoustic-Doppler velocimetry instruments that infer direction using magnetic
compasses need to have a correction applied for magnetic declination, if the
goal is to infer currents with x and y oriented eastward and northward,
respectively.  This function, which may be called directly or indirectly via
\code{\link[=applyMagneticDeclination]{applyMagneticDeclination()}}, handles this task by altering velocity
components (and heading values, if they exist), and also by setting the
\code{north} item of the \code{metadata} slot to \code{"geographic"}.  (If \code{north} already
has this value, the declination is still applied, but with a warning.)
}
\seealso{
Use \code{\link[=magneticField]{magneticField()}} to determine the declination,
inclination and intensity at a given spot on the world, at a given time.

Other things related to magnetism: 
\code{\link{applyMagneticDeclination,adp-method}},
\code{\link{applyMagneticDeclination,cm-method}},
\code{\link{applyMagneticDeclination,oce-method}},
\code{\link{applyMagneticDeclination}()},
\code{\link{magneticField}()}

Other things related to adv data: 
\code{\link{[[,adv-method}},
\code{\link{[[<-,adv-method}},
\code{\link{adv-class}},
\code{\link{adv}},
\code{\link{beamName}()},
\code{\link{beamToXyz}()},
\code{\link{enuToOtherAdv}()},
\code{\link{enuToOther}()},
\code{\link{plot,adv-method}},
\code{\link{read.adv.nortek}()},
\code{\link{read.adv.sontek.adr}()},
\code{\link{read.adv.sontek.serial}()},
\code{\link{read.adv.sontek.text}()},
\code{\link{read.adv}()},
\code{\link{rotateAboutZ}()},
\code{\link{subset,adv-method}},
\code{\link{summary,adv-method}},
\code{\link{toEnuAdv}()},
\code{\link{toEnu}()},
\code{\link{velocityStatistics}()},
\code{\link{xyzToEnuAdv}()},
\code{\link{xyzToEnu}()}
}
\author{
Dan Kelley, aided by Clark Richards and Jaimie Harbin.
}
\concept{things related to adv data}
\concept{things related to magnetism}
