% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{applyMagneticDeclination,oce-method}
\alias{applyMagneticDeclination,oce-method}
\title{Apply magnetic declination}
\usage{
\S4method{applyMagneticDeclination}{oce}(
  object = "oce",
  declination = 0,
  debug = getOption("oceDebug")
)
}
\arguments{
\item{object}{an object of same type as the input, with velocities
(and headings, if they exist) modified to account for the magnetic
declination, and with \verb{@metadata$north} being set to \code{"geographic"}.}

\item{declination}{magnetic declination, in degrees.}
}
\description{
Velocities (and headings, if they also exist) are modified to take into
account the supplied declination.  To signal that this has been done, the
\code{metadata} slot is altered by setting an item name \code{north} to \code{"geographic"}.
(Note that the Most reading functions set this to \code{"magnetic"} by default.)
If the \code{north} item is already \code{"geographic"} then a warning is issued,
but the transformations are still done; this is to permit rectifying errors
without the user having to modify the data directly.
}
\concept{functions that handle magnetic declination}
