% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{applyMagneticDeclination}
\alias{applyMagneticDeclination}
\title{Alter an object to account for magnetic declination}
\usage{
applyMagneticDeclination(object = "oce", declination = "ANY", debug = "ANY")
}
\arguments{
\item{object}{an object of \linkS4class{cm}, \linkS4class{adp}, or \linkS4class{adv} class.}

\item{declination}{numeric value holding magnetic declination in degrees,
positive for clockwise from north.}

\item{debug}{a debugging flag, set to a positive value to get debugging.}
}
\value{
an object of the same class as \code{object}, in which the velocities
(and headings, if they exist)
have been rotated to the \code{v} coordinate is aligned to true (geographic)
north.
Also, the item named \code{north} in the \code{metadata}
slot will be set to \code{geographic}.  If \code{x} already has this setting, then
a warning is issued, so that users can alter the declination setting, but
will have been made aware of a previously-set value.  Use e.g.
\code{x[["north"]]} to see the existing setting, if it exists.  (Objects
created with previous versions of oce will not have this setting.)
}
\description{
Current-measuring instruments that infer flow direction using magnetic
compasses require a correction for magnetic declination, in order to infer
currents with x and y oriented eastward and northward, respectively.
\code{\link[=applyMagneticDeclination]{applyMagneticDeclination()}} is a generic function that handles this task by
altering velocity components (and heading values, if they exist).  It works
for objects of the \linkS4class{cm}, \linkS4class{adp} and \linkS4class{adv} and \linkS4class{cm}
classes by calling specialized methods for those
classes.
}
\seealso{
Use \code{\link[=magneticField]{magneticField()}} to determine the declination,
inclination and intensity at a given spot on the world, at a given time.

Other things related to magnetism: 
\code{\link{applyMagneticDeclination,adp-method}},
\code{\link{applyMagneticDeclination,adv-method}},
\code{\link{applyMagneticDeclination,cm-method}},
\code{\link{applyMagneticDeclination,oce-method}},
\code{\link{magneticField}()}
}
\author{
Dan Kelley, aided, for the \linkS4class{adp} and \linkS4class{adv} variants,
by Clark Richards and Jaimie Harbin.
}
\concept{things related to magnetism}
