% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\name{applyMagneticDeclinationAdp}
\alias{applyMagneticDeclinationAdp}
\title{Alter an adp-class object to account for magnetic declination}
\usage{
applyMagneticDeclinationAdp(x, declination = 0, debug = getOption("oceDebug"))
}
\arguments{
\item{x}{an \linkS4class{adp} object.}

\item{declination}{magnetic declination in degrees.}

\item{debug}{a debugging flag, set to a positive value to get debugging.}
}
\value{
An \linkS4class{adp} object, modified as explained in \sQuote{Description}.
}
\description{
Acoustic-Doppler profiling instruments that infer direction using magnetic
compasses to determine current direction need to have a correction applied
for magnetic declination, if the goal is to infer currents with x and y
oriented eastward and northward, respectively.  This function, which
may be called directly or indirectly via \code{\link[=handleMagneticDeclination]{handleMagneticDeclination()}},
handles this task by altering velocity components (and heading values,
if they exist), and also by setting the \code{north} item of the \code{metadata}
slot to \code{"geographic"}.  (If \code{north} already has this value, the declination
is still applied, but with a warning.)
}
\seealso{
Use \code{\link[=magneticField]{magneticField()}} to determine the declination,
inclination and intensity at a given spot on the world, at a given time.

Other things related to magnetism: 
\code{\link{applyMagneticDeclinationAdv}()},
\code{\link{applyMagneticDeclinationCm}()},
\code{\link{applyMagneticDeclinationFIXME}()},
\code{\link{magneticField}()}
}
\author{
Dan Kelley, with help from Clark Richards and Jaimie Harbin.
}
\concept{things related to magnetism}
