% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.R
\name{applyMagneticDeclinationCm}
\alias{applyMagneticDeclinationCm}
\title{Alter a cm-class object to account for magnetic declination}
\usage{
applyMagneticDeclinationCm(x, declination = 0, debug = getOption("oceDebug"))
}
\arguments{
\item{x}{a \linkS4class{cm} object.}

\item{declination}{magnetic declination in degrees.}

\item{debug}{a debugging flag, set to a positive value to get debugging.}
}
\value{
Object, with velocity components (and heading data, if they exist)
adjusted as outlined in \sQuote{Description}.
}
\description{
Current-meter (\code{cm}) instruments determine directions from onboard compasses,
so interpreting velocity components in geographical coordinates requires that
magnetic declination be taken into account.  This is what the present
function does.  It rotates the \code{u} and \code{v} velocity components so that \code{u}
points eastward and \code{v} points northwards, and it also changes the
\code{Hdg} (heading) vectors in the same way. To signal that the work has been
done, the \code{north} item in the \code{metadata} is set to \code{"geographic"}. (This is
set to \code{"magnetic"} by \code{\link[=read,cm-method]{read,cm-method()}}.)  A warning is issued if the data
are already in geographic coordinates, but the \code{declination} is applied
nevertheless, so that a user can correct previous values if necessary.
}
\seealso{
Use \code{\link[=magneticField]{magneticField()}} to determine the declination,
inclination and intensity at a given spot on the world, at a given time.

Other things related to magnetism: 
\code{\link{applyMagneticDeclinationAdp}()},
\code{\link{applyMagneticDeclinationAdv}()},
\code{\link{applyMagneticDeclinationFIXME}()},
\code{\link{magneticField}()}
}
\author{
Dan Kelley
}
\concept{things related to magnetism}
