% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{applyMagneticDeclinationFIXME}
\alias{applyMagneticDeclinationFIXME}
\title{Alter an object to account for magnetic declination}
\usage{
applyMagneticDeclinationFIXME(
  x,
  declination = 0,
  debug = getOption("oceDebug")
)
}
\arguments{
\item{x}{an \linkS4class{oce} object of \linkS4class{cm}, \linkS4class{adp}, or \linkS4class{adv}.}

\item{declination}{magnetic declination in degrees.}

\item{debug}{a debugging flag, set to a positive value to get debugging.}
}
\value{
Object, with velocity components (and heading data, if they exist)
that have been adjusted to be aligned with
geographic north and east.  Also, the item named \code{north} in the \code{metadata}
slot will be set to \code{geographic}.  If \code{x} already has this setting, then
a warning is issued, so that users can alter the declination setting, but
will have been made aware of a previously-set value.  Use e.g.
\code{x[["north"]]} to see the existing setting, if it exists.  (Objects
created with previous versions of oce will not have this setting.)
}
\description{
Instruments that infer direction using magnetic compasses to determine
current direction need to have a correction applied for magnetic declination,
if the goal is to infer currents with x and y oriented eastward and
northward, respectively.  This function handles this task, by altering
velocity components (and heading values, if they exist) for objects of
\linkS4class{cm}, \linkS4class{adp} and \linkS4class{adv} and \linkS4class{cm}. It is a generic
function that works by \code{\link[=applyMagneticDeclinationAdp]{applyMagneticDeclinationAdp()}},
\code{\link[=applyMagneticDeclinationAdv]{applyMagneticDeclinationAdv()}}, or \code{\link[=applyMagneticDeclinationCm]{applyMagneticDeclinationCm()}}, as
appropriate.  (The user may also call these functions directly.)
}
\references{
\enumerate{
\item \samp{https://www.ngdc.noaa.gov/IAGA/vmod/igrf.html}
}
}
\seealso{
Use \code{\link[=magneticField]{magneticField()}} to determine the declination,
inclination and intensity at a given spot on the world, at a given time.

Other things related to magnetism: 
\code{\link{applyMagneticDeclinationAdp}()},
\code{\link{applyMagneticDeclinationAdv}()},
\code{\link{applyMagneticDeclinationCm}()},
\code{\link{magneticField}()}
}
\author{
Dan Kelley, with help from Clark Richards and Jaimie Harbin.
}
\concept{things related to magnetism}
