% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{argShow}
\alias{argShow}
\title{Show an argument to a function, e.g. for debugging}
\usage{
argShow(x, nshow = 2, last = FALSE, sep = "=")
}
\arguments{
\item{x}{the argument}

\item{nshow}{number of values to show at first (if length(x)> 1)}

\item{last}{indicates whether this is the final argument to the function}

\item{sep}{the separator between name and value}
}
\description{
Show an argument to a function, e.g. for debugging
}

