% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bound125}
\alias{bound125}
\title{Calculate a rounded bound, rounded up to matissa 1, 2, or 5}
\usage{
bound125(x)
}
\arguments{
\item{x}{a single positive number}
}
\value{
for positive x, a value exceeding x that has mantissa 1, 2, or 5; otherwise, x
}
\description{
Calculate a rounded bound, rounded up to matissa 1, 2, or 5
}
