\name{cm-class}

\docType{class}

\alias{cm-class}
\alias{[[,cm-method}

\title{Class to store current meter data}

\description{Class to store current meter data, with standard slots
    \code{metadata}, \code{data} and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{cm[["u"]]}, where
    the string could also be \code{"v"}.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

