\name{cm}

\docType{data}

\alias{cm}

\title{Current-meter record}

\description{This is sample current meter record from an InterOcean S4 devide.}

\usage{data(cm)}

\examples{
\dontrun{
library(oce)
data(cm)
summary(cm)
plot(cm)
}
}

\source{This is a snippet from a time series recorded as part of the SLEIWEX
    experiment (\url{http://myweb.dal.ca/kelley/SLEIWEX/index.php}.}

\seealso{The documentation for \code{cm-class} in the Oce package explains the
    structure of cm objects, and also outlines the other functions dealing with
    them.}

\keyword{datasets}

