\name{coastline.hal}
\docType{data}
\alias{coastline.hal}
\title{Sample coastline data set (Halifax Harbour)}
\description{A sample coastline dataset, showing Halifax Harbour.
The little island at about 44.64N is visible outside the author's window.}
\usage{data(coastline.hal)}
\source{}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/}}
\seealso{Coastlines may be read with
\code{\link{read.coastline}}, or created with
\code{\link{as.coastline}}.  Another
sample coastline file is \code{\link{coastline.maritimes}}; see these for details of the data format.   Other coastline datasets provided with
oce are \code{\link{coastline.maritimes}} and \code{\link{coastline.world}}.}
\author{Dan Kelley}
\keyword{misc}
