\name{coastline.maritimes}
\docType{data}
\alias{coastline.maritimes}
\title{Sample coastline data set, showing Maritime provinces of Canada}
\description{A sample coastline dataset, showing Nova Scotia, 
Prince Edward Island, and part of New Brunswick.
This is the oce author's stomping ground, and so he feels compelled
to apologize for the error at the Strait of Canso.}
\usage{data(coastline.maritimes)}
\source{The data were extracted from the NOAA site, specifying the Splus
data type, and read with \code{\link{read.coastline}}.}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/}}
\seealso{Coastlines may be read with
\code{\link{read.coastline}}, or created with
\code{\link{as.coastline}}.  Another
sample coastline file is \code{\link{coastline.halifax}}.}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
