\name{coastline}
\docType{data}
\alias{coastline}
\title{Sample coastline data set}
\description{A sample coastline dataset, showing Nova Scotia, 
Prince Edward Island, and part of New Brunswick.
This is the oce author's stomping ground, and so he feels compelled
to apologize for the error at the Strait of Canso.}
\usage{data(coastline)}
\source{The data were extracted from the NOAA site, specifying the Splus
data type, and read with \code{\link{read.coastline}}.}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/}}
\seealso{\code{\link{read.coastline}}}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
