\name{coastlineHalifax}
\docType{data}
\alias{coastlineHalifax}
\title{Coastline data set, showing Halifax Harbour}
\description{A sample coastline dataset, showing Halifax Harbour.
The little island at about 44.64N is visible outside the author's window.}
\usage{data(coastlineHalifax)}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/}}
\seealso{Coastlines may be read with
\code{\link{read.coastline}}, or created with
\code{\link{as.coastline}}.  Another
sample coastline file is \code{\link{coastlineMaritimes}}; see these for details of the data format.   Other coastline datasets provided with
oce are \code{\link{coastlineMaritimes}} and \code{\link{coastlineWorld}}.}
\author{Dan Kelley}
\keyword{misc}
