\name{coastlineMaritimes}

\docType{data}

\alias{coastlineMaritimes}

\title{Coastline data set, showing Maritime provinces of Canada}

\description{A sample coastline dataset, showing Nova Scotia, Prince Edward
    Island, and part of New Brunswick.  This is the oce author's stomping
    ground, and so he feels compelled to apologize for the error at the Strait
    of Canso.}

\usage{data(coastlineMaritimes)}

\source{The data were extracted from the the 1:2,000,000 ``World Data Bank II''
    dataset at the NOAA site, with a few ``fake'' points added to create a closed
    geometry that can be drawn as a filled polygon.}

\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/index.jsp?llat=43.0&ulat=47.2&llon=-66.8&rlon=-59.5}}

\seealso{The documentation for \code{\link{coastline-class}} explains the
    structure of coastline objects, and also outlines the other functions
    dealing with them.}

\author{Dan Kelley}

\keyword{datasets}
