\name{coastlineSLE}
\docType{data}
\alias{coastlineSLE}
\title{Coastline of St Lawrence Estuary}
\description{A sample coastline dataset, showing the St Lawrence Estuary.}
\usage{data(coastlineSLE)}
\source{The data were extracted from the the 1:250,000 ``World Vector Shoreline''
dataset at the NOAA site.}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline}}
\seealso{Coastlines may be read with
\code{\link{read.coastline}}, or created with
\code{\link{as.coastline}}; see these for details of the data format.  Other coastline datasets provided with
oce are \code{\link{coastlineHalifax}} and \code{\link{coastlineWorld}}.}
\author{Dan Kelley}
\keyword{misc}
