% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\name{coastlineTrim}
\alias{coastlineTrim}
\title{Trim a Coastline to a Longitude-Latitude Box}
\usage{
coastlineTrim(x, west = -180, east = 180, south = -90, north = 90,
  debug = getOption("oceDebug"))
}
\arguments{
\item{x}{A \code{coastline} object, i.e. one inheriting from \code{\link{coastline-class}}.}

\item{west}{Longitude of the western boundary of the retained region.}

\item{east}{Longitude of the eastern boundary of the retained region.}

\item{south}{Latitude of the southern boundary of the retained region.}

\item{north}{Latitude of the northrn boundary of the retained region.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
A \code{coastline} object.
}
\description{
Subset a coastline file to a specified sub-region, adding fake
points along the intersection boundary with the box. This can speed
up repeated plots of sub-domains.
}
\details{
Fake points are added along the boundary
of any coastline segments that are only partly contained within the box,
which permits accurat filled coastline plots of the subdomain. It
makes sense to set the focus box to be larger than will be required
in plotting, so that \code{\link{plot,coastline-method}} will
not leave white bands at the top/bottom or left/right sides of
the plot panel, depending on the geometry of the plot device. (The
code in \dQuote{Examples} produces such artifacts on the left
and right sides for wide plot devices, and at top and bottom
for tall devices.)

Note that the method uses the \code{\link[raster]{intersect}} function
in the \CRANpkg{raster} package, so that package must be available
for \code{coastlineTrim} to work.
}
\examples{
\donttest{
library(oce)
data(coastlineWorldFine, package="ocedata")
## Region centred on New Brunswick, in eastern Canada
EC <- coastlineTrim(coastlineWorldFine, west=-80, east=-50, south=40, north=55)
plot(EC)
}
}
\seealso{
Other things related to \code{coastline} data: \code{\link{[[,coastline-method}},
  \code{\link{[[<-,coastline-method}},
  \code{\link{as.coastline}},
  \code{\link{coastline-class}},
  \code{\link{coastlineBest}}, \code{\link{coastlineCut}},
  \code{\link{coastlineWorld}},
  \code{\link{download.coastline}},
  \code{\link{plot,coastline-method}},
  \code{\link{read.coastline.openstreetmap}},
  \code{\link{read.coastline.shapefile}},
  \code{\link{summary,coastline-method}}
}
\author{
Dan Kelley
}
\concept{things related to \code{coastline} data}
