% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\docType{methods}
\name{composite,amsr-method}
\alias{composite,amsr-method}
\title{Average objects to create a composite}
\usage{
\S4method{composite}{amsr}(object, ...)
}
\details{
Form averages for each item in the \code{data} slot of the supplied objects,
taking into account the bad-data codes. If none of the objects has good
data at any particular pixel (i.e. particular latitude and longitude),
the resultant will have the bad-data code of the last item in the argument
list.
The metadata in the result are taken directly from the metadata of the
final argument, except that the filename is set to a comma-separated list
of the component filenames.
}
\seealso{
Other functions that create composite objects: \code{\link{composite}}

Other things related to \code{amsr} data: \code{\link{[[<-,amsr-method}},
  \code{\link{amsr-class}}, \code{\link{plot,amsr-method}},
  \code{\link{read.amsr}},
  \code{\link{subset,amsr-method}},
  \code{\link{summary,amsr-method}}
}

