% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{composite,ANY-method}
\alias{composite,ANY-method}
\title{Combine several oce objects into a composite}
\usage{
\S4method{composite}{ANY}(object, ...)
}
\arguments{
\item{object}{Any object.}
}
\description{
Combine several oce objects into a composite
}

