% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{composite}
\alias{composite}
\title{Create a composite object by averaging across good data}
\usage{
composite(object, ...)
}
\arguments{
\item{object}{An object inheriting from \link{oce-class}.}

\item{...}{Other oce objects of the same inheritance as \code{object}.}
}
\description{
Items within the \code{data} slots of the
objects that are supplied as arguments are averaged in a way that
makes sense for the object class, i.e. taking into account
the particular bad-data codes of that particular class.
}
\details{
This only works for objects inheriting from \code{\link{amsr-class}}.
}
\seealso{
Other functions that create composite objects: \code{\link{composite,amsr-method}}
}

