% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{concatenate,list-method}
\alias{concatenate,list-method}
\title{Concatenate a List of oce Objects}
\usage{
\S4method{concatenate}{list}(object, debug = getOption("oceDebug"))
}
\arguments{
\item{object}{a \link{list} of \linkS4class{oce} objects, all of which must have the
same sub-class (e.g. all of \linkS4class{ctd}, or \linkS4class{adp}, etc).}

\item{debug}{a debugging flag, set to a positive value to get debugging. Note
that \code{debug-1} is passed to the other \code{concatenate()} functions that are
called by the present function.}
}
\value{
An object of class corresponding to that in the elements of \code{object}.
}
\description{
Concatenate a List of oce Objects
}
\seealso{
Other functions for concatenating oce objects: 
\code{\link{concatenate}()},
\code{\link{concatenate,adp-method}}
}
\concept{functions for concatenating oce objects}
