% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{concatenate,oce-method}
\alias{concatenate,oce-method}
\title{Concatenate oce objects}
\usage{
\S4method{concatenate}{oce}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{oce-class}}, or a list containing such
objects (in which case the remaining arguments are ignored).}

\item{...}{Optional additional objects of \code{\link{oce-class}}.}
}
\description{
Concatenate oce objects

This function concatenates oce objects.
}
\examples{
## 1. Split, then recombine, a ctd object.
data(ctd)
ctd1 <- subset(ctd, scan <= median(ctd[["scan"]]))
ctd2 <- subset(ctd, scan > median(ctd[["scan"]]))
CTD <- concatenate(ctd1, ctd2)

## 2. Split, then recombine, an adp object.
data(adp)
midtime <- median(adp[["time"]])
adp1 <- subset(adp, time <= midtime)
adp2 <- subset(adp, time > midtime)
ADP <- concatenate(adp1, adp2)

\dontrun{
## 3. Download two met files and combine them.
met1 <- read.met(download.met(id=6358, year=2003, month=8))
met2 <- read.met(download.met(id=6358, year=2003, month=9))
MET <- concatenate(met1, met2)
}

}
\seealso{
Other functions that concatenate \code{oce} objects.: \code{\link{concatenate,adp-method}}
}
\author{
Dan Kelley
}
