% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate oce Objects (Generic)}
\usage{
concatenate(object, ..., debug = getOption("oceDebug"))
}
\arguments{
\item{object}{an \linkS4class{oce} object.}

\item{...}{optional additional \linkS4class{oce} objects.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
An object of class corresponding to that of \code{object}.
}
\description{
Concatenate oce Objects (Generic)
}
\seealso{
Other functions that concatenate oce objects: 
\code{\link{concatenate,adp-method}},
\code{\link{concatenate,list-method}},
\code{\link{concatenate,oce-method}}
}
\concept{functions that concatenate oce objects}
