\name{ctd-class}

\docType{class}

\alias{ctd-class}
\alias{[[,ctd-method}

\title{Class to store hydrographic data, e.g. from a CTD}

\description{Class to store hydrographic data, with standard slots
    \code{metadata}, \code{data} and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{ctd[["salinity"]]}, where
    the string could also be \code{"temperature"} or \code{"pressure"}, or
    abbreviations matching uniquely to any of these.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

