\name{ctd}
\docType{data}
\alias{ctd}
\title{Sample seawater (CTD) profile}
\description{This is sample CTD profile provided for testing.
It has been pruned by selecting only the subsequence during
which the instrument was being lowered through the water column.}
\usage{data(ctd)}
\format{
  This file is from a \emph{Sea-Bird SBE 19} CTD instrument.
  \emph{(There should be a reference to a URL with the data format here,
    but I have yet to find one.  --DEK)}}
}
\source{Peter Galbraith acquired these data during the 2002 SLEIWEX
  field program, part of a Dalhousie/IML cooperative study of
  internal-wave mixing in the St Lawrence estuary.} 
\references{
  The Seabird CTD instrument is described at
  \url{http://www.seabird.com/products/spec_sheets/19plusdata.htm}.
}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
