% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{D4902337_219.nc}
\alias{D4902337_219.nc}
\title{Sample Argo Profile File}
\description{
Cycle 219 of the Argo float with ID 4902337, holding
delayed-mode hydrographic data acquired in July 2021,
over the continental slope, approximately 650km east
of Halifax, Nova Scotia, and 300km south of St. Pierre
and Miquelon (near the south coast of Newfoundland).
This file is used in the documentation for \code{\link[=argo2ctd]{argo2ctd()}}.
}
\examples{
argo <- read.argo(system.file("extdata", "D4902337_219.nc", package = "oce"))
summary(argo)

}
\seealso{
Other raw datasets: 
\code{\link{CTD_BCD2014666_008_1_DN.ODF.gz}},
\code{\link{adp_rdi.000}},
\code{\link{ctd.cnv.gz}},
\code{\link{ctd_aml_type1.csv.gz}},
\code{\link{ctd_aml_type3.csv.gz}},
\code{\link{d200321-001.ctd.gz}},
\code{\link{d201211_0011.cnv.gz}},
\code{\link{xbt.edf}},
\code{\link{xbt2.edf}}

Other things related to argo data: 
\code{\link{[[,argo-method}},
\code{\link{[[<-,argo-method}},
\code{\link{argo}},
\code{\link{argo-class}},
\code{\link{argo2ctd}()},
\code{\link{argoGrid}()},
\code{\link{argoNames2oceNames}()},
\code{\link{as.argo}()},
\code{\link{handleFlags,argo-method}},
\code{\link{plot,argo-method}},
\code{\link{read.argo}()},
\code{\link{read.argo.copernicus}()},
\code{\link{subset,argo-method}},
\code{\link{summary,argo-method}}
}
\concept{raw datasets}
\concept{things related to argo data}
