% Generated by roxygen2 (4.0.1): do not edit by hand
\name{define_ODF_header}
\alias{define_ODF_header}
\title{Define the header section of an ODF file.}
\usage{
define_ODF_header()
}
\description{
Defines and returns the header section of an ODF file.
Output:
  ODF_HEADER is a cell array containing the definition. It also
  outputs a mat file named "definition_ODF_header.mat'.
}
\details{
This tool usually is called by other tools such as ODF read,
  write and edit etc. It is one of the core tools a user shouldn't
  make any changes to it.

A new block can be added easily to this definition to extend an
  ODF header structure. The new block should be considered as
  'optional', otherwise all existing ODF files will have to be
  modified as well.

Report any bugs to DataServicesDonnees@dfo-mpo.gc.ca

ODSToolbox Version: 2.0

Last Updated: June 17, 2014

Source:
  Ocean Data and Information Services,
  Bedford Institute of Oceanography, DFO, Canada.
  DataServicesDonnees@dfo-mpo.gc.ca
}
\examples{
ODF_header <- define_ODF_header()
}
\author{
Yongcun Hu, Patrick Upson
}

