% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{firstFinite}
\alias{firstFinite}
\title{Get First Finite Value in a Vector or Array.}
\usage{
firstFinite(v)
}
\arguments{
\item{v}{A numerical vector or array.}
}
\value{
The first finite value, or NULL if there are no
finite values.
}
\description{
If x is a vector, this is straightforward.  If x is
anything else, it is first converted to a vector with
\code{\link[=as.vector]{as.vector()}}, so the first value will be with
respect to storage by columns, for a matrix, etc.
}
