\name{full.filename}
\alias{full.filename}
\title{full name of file, including path}
\description{Return full name of file, including path}

\usage{full.filename(filename)}

\arguments{
  \item{filename}{name of file}
}

\details{Determines the full name of a file, including the path.  Used
  by many \code{read.X} routines, where \code{X} is the name of a class
  of object.}

\value{Full file name}

\author{Dan Kelley}

\keyword{misc}
