% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{handleFlags,vector-method}
\alias{handleFlags,vector-method}
\title{Signal Erroneous Application to non-oce Objects}
\usage{
\S4method{handleFlags}{vector}(
  object = "oce",
  flags = list(),
  actions = list(),
  where = list(),
  debug = getOption("oceDebug")
)
}
\arguments{
\item{object}{A vector, which cannot be the case for \code{oce} objects.}

\item{flags}{Ignored.}

\item{actions}{Ignored.}

\item{where}{Ignored.}

\item{debug}{Ignored.}
}
\description{
Signal Erroneous Application to non-oce Objects
}
