% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{head.oce}
\alias{head.oce}
\title{Extract the start of an \code{oce} object}
\usage{
\method{head}{oce}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{oce} object of a suitable class (presently only \code{adp} is
permitted).}

\item{n}{Number of elements to extract.}

\item{...}{ignored}
}
\description{
Extract the start of an \code{oce} object
}
\seealso{
\code{\link{tail.oce}}, which yields the end of an \code{oce} object.
}

