\name{historyItem}

\alias{historyItem}

\title{Create a history item.}

\description{Create a history item.}

\usage{historyItem(action="")}

\arguments{
  \item{action}{A character string describing the action.}
}

\details{This is used, mainly by internal oce functions, to create a
  single-entry history}

\value{An object of class \code{"history"}, suitable for
  inclusion in any \code{"oce"} object.}

\seealso{\code{\link{historyAdd}}}

\author{Dan Kelley}

\keyword{misc}
