% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{inferUnits}
\alias{inferUnits}
\title{Infer units and scales from character strings}
\usage{
inferUnits(strings, dictionary = "default", debug = 0)
}
\arguments{
\item{dictionary}{either a data frame or a string indicating a csv
file that holds a data frame.  If a string, and if it ends in \code{".csv"},
then that file is used, but if it has another ending, it is taken
as an indicator of a built-in file that is provided with
the package, named \code{units_}, followed by the supplied \code{dictionary}
value, and ending with \code{.csv}.}

\item{s}{vector of character values.}
}
\value{
if the unit is recognized, a list is returned, with the first
column, named \code{unit}, being an \code{\link[=expression]{expression()}} and the second,
named \code{scale}, being an optional descriptor of the unit.
}
\description{
Infer units and scales from character strings
}
\author{
Dan Kelley
}
