% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{initializeFlagScheme,oce-method}
\alias{initializeFlagScheme,oce-method}
\title{Establish a data-quality scheme for a}
\usage{
\S4method{initializeFlagScheme}{oce}(object, name = NULL, mapping = NULL,
  debug = getOption("oceDebug"))
}
