% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{initializeFlagScheme}
\alias{initializeFlagScheme}
\title{Establish a data-quality scheme for a oce object}
\usage{
initializeFlagScheme(object, scheme = NULL, debug = 0)
}
\arguments{
\item{object}{An oce object.}

\item{scheme}{A list describing the flag scheme, e.g \code{list(good=1, bad=2)}
might be used for a hypothetical class. Some classes also permit \code{scheme}
to be a character string that names a built-in scheme; see \dQuote{Details}.}

\item{debug}{Integer set to 0 for quiet action or to 1 for some debugging.}
}
\value{
An object with the \code{metadata} slot set to the indicated
\code{scheme} (expanded to a list, if the argument was a recognized
character string).
}
\description{
This function stores adds an item called \code{flagScheme} to the
\code{metadata} slot. The purpose is both to document a flag scheme
and to make it so that \code{\link{initializeFlags}}
and \code{\link{setFlags}} can specify flags by
name, as opposed to number. A generic function, it is
specialized for some classes via interpretation of the
\code{scheme} argument (see \dQuote{Details}, for those
object classes that have such specializations).
}
\details{
There are no pre-defined \code{scheme}s for this object class.
}
\section{Caution}{

This function was added in early May, 2018, and is likely to undergo
changes until the mid-summer of that year.  Use with caution.
}

\seealso{
Other functions relating to data-quality flags: \code{\link{handleFlags,adp-method}},
  \code{\link{handleFlags,argo-method}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{handleFlags,section-method}},
  \code{\link{handleFlags}},
  \code{\link{initializeFlagScheme,ctd-method}},
  \code{\link{initializeFlagScheme,oce-method}},
  \code{\link{initializeFlags,adp-method}},
  \code{\link{initializeFlags,oce-method}},
  \code{\link{initializeFlags}},
  \code{\link{setFlags,adp-method}},
  \code{\link{setFlags,ctd-method}},
  \code{\link{setFlags,oce-method}}, \code{\link{setFlags}}
}
