% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{integerToAscii}
\alias{integerToAscii}
\title{Infer ASCII Code From an Integer Value}
\usage{
integerToAscii(i)
}
\arguments{
\item{i}{an integer, or integer vector.}
}
\value{
A character, or character vector.
}
\description{
Infer ASCII Code From an Integer Value
}
\examples{
library(oce)
A <- integerToAscii(65)
cat("A=", A, "\n")
}
\author{
Dan Kelley
}
