% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{IPTS68}
\alias{IPTS68}
\title{Convert from ITS90 temperature to IPTS68 temperature using Saunders' (1990) formula}
\usage{
IPTS68(T90)
}
\arguments{
\item{T90}{temperature expressed in the ITS90 scale}
}
\value{
Temperature expressed in the IPTS68 scale
}
\description{
Convert from ITS90 temperature to IPTS68 temperature using Saunders' (1990) formula
}

