% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{ITS90}
\alias{ITS90}
\title{Convert from IPTS68 temperature to ITS90temperature using Saunders' (1990) formula}
\usage{
ITS90(T68)
}
\arguments{
\item{T68}{temperature expressed in the IPTS68 scale}
}
\value{
Temperature expressed in the ITS90 scale
}
\description{
Convert from IPTS68 temperature to ITS90temperature using Saunders' (1990) formula
}

