\name{landsat-class}

\docType{class}

\alias{landsat-class}
\alias{[[,landsat-method}
\alias{[[<-,landsat-method}
\alias{show,landsat-method}

\title{Class to store landsat data}

\description{Class to store landsat data, with standard slots
    \code{metadata}, \code{data} and \code{processingLog}.}

\section{Methods}{

    Consider a landsat object named \code{landsat}.

    \emph{Accessing data values.}
   
    (fill in later)
    
    \emph{Accessing scalar values.}
  
    Various scalar quantities are also available, e.g.
    \code{landsat[["time"]]}, etc.

    \emph{Accessing derived values.}
  
    (fill in later)

    \emph{Assigning values.}
   
    (fill in later)

    \emph{Overview of contents.}
   
    The \code{show} method (e.g.  \code{show(landsat)}) displays information
    about the object.

}

\seealso{A file containing Landsat data may be read with
    \code{\link{read.landsat}}.
 
}

\author{Dan Kelley}

\keyword{classes}

\keyword{oce}

