\name{latlon.format}

\alias{latlon.format}

\title{Format a latitude-longitude pair}

\description{Format a latitude-longitude pair, using "S" for negative latitudes, etc.}

\usage{latlon <- latlon.format(lat, lon)}

\arguments{
  \item{lat}{latitude in \eqn{^\circ}{deg}N north of the equator.}
  \item{lon}{longitude in \eqn{^\circ}{deg}N east of Greenwich.}
}

\value{A character string.}

\seealso{\code{\link{lat.format}} and \code{\link{lon.format}}.}

\author{Dan Kelley}

\keyword{misc}
