% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\docType{data}
\name{lisst}
\alias{lisst}
\title{LISST dataset}
\source{
This was constructed artifically using \code{\link{as.lisst}},
}
\usage{
data(lisst)
}
\description{
LISST (Laser in-situ scattering and transmissometry) dataset, constructed
artificially.
}
\author{
Dan Kelley
}
\seealso{
Other datasets provided with \code{oce}: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{lobo}},
  \code{\link{section}}, \code{\link{topoWorld}}
}

