\name{lobo-class}

\docType{class}

\alias{lobo-class}
\alias{[[,lobo-method}

\title{Class to store LOBO data}

\description{Class to store LOBO data, with standard slots
    \code{metadata}, \code{data} and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{lobo[["salinity"]]}, where
    the string could also be \code{"temperature"} or \code{"pressure"}, or
    abbreviations matching uniquely to any of these.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

