% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lowpass}
\alias{lowpass}
\title{Perform lowpass digital filtering}
\usage{
lowpass(x, filter = "hamming", n, replace = TRUE)
}
\arguments{
\item{x}{a vector to be smoothed}

\item{filter}{name of filter; at present, only \code{"hamming"} is permitted.}

\item{n}{length of filter (must be an odd integer exceeding 1)}

\item{replace}{a logical value indicating whether points near the
ends of \code{x} should be copied into the end regions, replacing
the \code{NA} values that would otherwise be placed there by
\link[stats]{filter}.}
}
\value{
\code{lowpass} returns a filtered version of \code{x}.
}
\description{
Filter coefficients are constructed, and then \link[stats]{filter}
in the \pkg{stats} package is used to filter the data. This leaves \code{NA}
values within half the filter length of the ends of the time series, but
these may be replaced with the original \code{x} values, if the argument
\code{replace} is set to \code{TRUE}.
}
\examples{

library(oce)
x <- seq(-5, 5) + rnorm(11)
plot(1:11, x, type='o')
X <- lowpass(x, n=5)
lines(1:11, X, col=2)
points(1:11, X, col=2)
}
\author{
Dan Kelley
}
