% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{NEWhandleFlags,adp-method}
\alias{NEWhandleFlags,adp-method}
\title{handle flags}
\usage{
\S4method{NEWhandleFlags}{adp}(object, flags = NULL, action = "NA")
}
\arguments{
\item{object}{FILL IN}

\item{flags}{FILL IN}

\item{action}{FILL IN}
}
\description{
handle flags
}
\seealso{
Other functions that handle flags: \code{\link{NEWhandleFlags,ctd-method}}
}

