% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oce.R
\name{oce.grid}
\alias{oce.grid}
\title{Add a grid to an existing plot, with special abilities for those created by oce}
\usage{
oce.grid(xat, yat, col = "lightgray", lty = "dotted", lwd = par("lwd"))
}
\arguments{
\item{xat}{either a list of x values at which to draw the grid, or the return value from an oce plotting function}

\item{yat}{a list of y values at which to plot the grid (ignored if \code{gx} was a return value from an oce plotting function)}

\item{col}{colour of grid lines (see \code{\link{par}})}

\item{lty}{type for grid lines (see \code{\link{par}})}

\item{lwd}{width for grid lines (see \code{\link{par}})}
}
\value{
nothing
}
\description{
Add a grid to an existing plot, with special abilities for those created by oce
}
\details{
Add a grid to an existing plot, with special abilities for those created by oce.
}
\examples{
library(oce)
i <- imagep(volcano)
oce.grid(i)
}

