\name{oceans}

\docType{data}

\alias{oceans}

\title{Geometries of the five world oceans.}

\description{Surface areas, average depths, and maximum depths for five world
    oceans, calculated from the ETOPO-1 dataset.  Area is in square metres
    (converted from square kilometres on the NOAA site), and depths are in
    metres.}

\usage{data(oceans)}

\examples{
data(oceans)
}

\source{Data were downloaded from the NOAA website
    \url{http://ngdc.noaa.gov/mgg/global/etopo1_ocean_volumes.html} in October
    2012.}

\keyword{datasets}
