% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oceColors9A}
\alias{oceColors9A}
\alias{oce.colors9A}
\title{Create colours in a red-white-blue colour scheme}
\usage{
oceColors9A(n)
}
\arguments{
\item{n}{number of colours}
}
\description{
Create colours in a red-white-blue colour scheme
}
\seealso{
Other things related to colors: \code{\link{colormap}},
  \code{\link{colors}}, \code{\link{oce.colorsGebco}},
  \code{\link{oceColors9B}}, \code{\link{oceColorsCDOM}},
  \code{\link{oceColorsChlorophyll}},
  \code{\link{oceColorsDensity}},
  \code{\link{oceColorsFreesurface}},
  \code{\link{oceColorsJet}},
  \code{\link{oceColorsOxygen}},
  \code{\link{oceColorsPAR}},
  \code{\link{oceColorsPalette}},
  \code{\link{oceColorsPhase}},
  \code{\link{oceColorsSalinity}},
  \code{\link{oceColorsTemperature}},
  \code{\link{oceColorsTurbidity}},
  \code{\link{oceColorsTwo}},
  \code{\link{oceColorsVelocity}},
  \code{\link{oceColorsViridis}},
  \code{\link{oceColorsVorticity}}
}
