% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{oceColorsCubeHelix}
\alias{oceColorsCubeHelix}
\alias{oce.colorsCubeHelix}
\title{Create Colors in the Cube Helix Style}
\usage{
oceColorsCubeHelix(
  n = 256,
  start = 0.5,
  rotations = -1.5,
  hue = 1.2,
  gamma = 1
)
}
\arguments{
\item{n}{number of colors to create, with default 256.}

\item{start}{numeric value indicating the starting colour.
The default, 0.5, corresponds to purpose, as in Green (2011).}

\item{rotations}{numeric value for number of rotations of the
helix, with default -1.5 as in Green (2011).}

\item{hue}{numeric value with default 1.2, as in Green (2011).}

\item{gamma}{numeric value, with default 1.}
}
\description{
This is a colour scheme that renders well in black and white,
that has colours that are reasonably distinct for many
viewers, and that may be good for persons with colour
deficiencies.  For more on such matters, the underlying
theory, and the meanings of the parameter values, please
see Green (2011). That R code for the present function
is a rewriting of Fortran code provided in an appendix
of that paper.
}
\examples{
library(oce)
# Example 1: normal (dark for low values)
imagep(volcano, col = oceColorsCubeHelix)
# Example 2: reversed (dark for high values)
imagep(volcano, col = \(n) rev(oceColorsCubeHelix(n)))

}
\references{
\enumerate{
\item Green, D.A. “A Colour Scheme for the Display of Astronomical
Intensity Images.” Bulletin of the Astronomical Society of India,
39 (2011): 289–95.
}
}
\seealso{
Other things related to colors: 
\code{\link{colormap}()},
\code{\link{colormapGMT}()},
\code{\link{oceColors9B}()},
\code{\link{oceColorsCDOM}()},
\code{\link{oceColorsChlorophyll}()},
\code{\link{oceColorsClosure}()},
\code{\link{oceColorsDensity}()},
\code{\link{oceColorsFreesurface}()},
\code{\link{oceColorsGebco}()},
\code{\link{oceColorsJet}()},
\code{\link{oceColorsOxygen}()},
\code{\link{oceColorsPAR}()},
\code{\link{oceColorsPalette}()},
\code{\link{oceColorsPhase}()},
\code{\link{oceColorsSalinity}()},
\code{\link{oceColorsTemperature}()},
\code{\link{oceColorsTurbidity}()},
\code{\link{oceColorsTurbo}()},
\code{\link{oceColorsTwo}()},
\code{\link{oceColorsVelocity}()},
\code{\link{oceColorsViridis}()},
\code{\link{oceColorsVorticity}()},
\code{\link{ocecolors}}
}
\author{
Dan Kelley based on Fortran code by D.A. Green (2011).
}
\concept{things related to colors}
