% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{oceGetData}
\alias{oceGetData}
\title{Get data from an \code{oce} object}
\usage{
oceGetData(object, name)
}
\arguments{
\item{object}{an \code{oce} object}

\item{name}{String indicating the name of the item to be found.}
}
\description{
In contrast to the various \code{[[} functions, this is
guaranteed to look only within the \code{data} slot. If
the named item is not found, \code{NULL} is returned.
}
