% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accessors.R
\name{oceMetadata}
\alias{oceMetadata}
\title{Get a metadata item from an oce object}
\usage{
oceMetadata(object, name, default = NA)
}
\arguments{
\item{object}{An oce object.}

\item{name}{Name of the desired item.}

\item{default}{A default value to be returned if \code{name} is not found}
}
\value{
The desired item, or the default, if the item is not present in the metadata slot.
}
\description{
Get a metadata item from an oce object
}
\details{
Get a named item from the object, returning a provided default if not found.
}

