% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{oceSetData}
\alias{oceSetData}
\title{Set something in the \code{data} slot of an \code{oce} object}
\usage{
oceSetData(object, name, value, units, note = "")
}
\arguments{
\item{object}{an \code{oce} object}

\item{name}{String indicating the name of the item to be set.}

\item{value}{Value for the item.}

\item{units}{An optional list specifying units for the item, containing an element
named \code{units} that is an \code{\link{expression}}, and a string named
\code{scale} that describes the scale used. For example, modern temperatures
have \code{unit=list(unit=expression(degree*C), scale="ITS-90")}.}

\item{note}{A note to be stored in the processing log.}
}
\description{
Set something in the \code{data} slot of an \code{oce} object
}

