% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\docType{class}
\name{odf-class}
\alias{odf-class}
\title{Class to Store ODF data}
\description{
Class for data stored in a format used at Canadian Department of Fisheries
and Oceans laboratories. This is somewhat unusual amongst \code{oce}
classes, in that it does not map to a particular instrument, but rather to a
storage type; in that sense, it is similar to the \code{\link{bremen-class}}.
}
\section{Slots}{

\describe{
\item{\code{data}}{As with all \code{oce} objects, the \code{data} slot is a
\code{\link{list}} containing the main data for the object.}

\item{\code{metadata}}{As with all \code{oce} objects, the \code{metadata} slot is
a \code{\link{list}} containing information about the \code{data} or about the
object itself.  An example of the former might be the location at which a
measurement was made, and of the latter might be the name of the experiment in
which the data were collected.}

\item{\code{processingLog}}{As with all \code{oce} objects, the \code{processingLog}
slot is a \code{\link{list}} with entries describing the creation and evolution
of the object. The contents are updated by various \code{oce} functions to
keep a record of processing steps.  Object summaries and
\code{\link{processingLogShow}} both display the log.}
}}

\section{Retrieving slot contents (general)}{


The full contents of the \code{data} and \code{metadata} slots of a \code{oce}
object named \code{obj} may be retrieved in the standard R way, with
\code{slot(obj, "data")} and \code{slot(obj, "metadata")}, or with
\code{obj[["data"]]} and \code{obj[["metadata"]]}, where the latter uses an
overloading of the \code{[[} operator for \code{oce} objects.

The \code{[[} mechanism can also be used to look for named items within
the slots. It first checks the \code{metadata} slot, returning the value
if it is there, and then tries in the \code{data} slot if not. If this
order of precedence is unsuitable, \code{\link{oceGetData}} and
\code{\link{oceGetMetadata}} can be used.

In addition to looking up stored quantities, the \code{[[} operator
has other useful capabilities. If an object holds data-quality flags, then
\code{obj[["flags"]]} will retrieve them, as a list with names matching
the variable names.  Similarly, \code{obj[["units"]]} return a list of units.
Individual items may be found in the usual R way within these lists, but there
is a simpler way: use e.g. \code{obj[["temperatureFlag"]]} or
\code{obj[["temperatureUnit"]]}. In such cases, \code{[[} is returning
information that is stored somewhere in the object, but it can also return
wholly \strong{derived} quantities, depending on the object type. For
example, \code{obj[["N2"]]} will return the square of the buoyancy
frequency, if \code{obj} is an object from which that can be calculated
(such as a \code{ctd} object). Information about accessing such derived
quantities may be found in the \dQuote{Special access methods}
section of the documentation for those classes that support it.
}

\section{Modifying slot contents (general)}{


Although there is a \code{[[<-} notation for modifying the contents
of \code{oce} objects, it is wiser to use \code{\link{oceSetData}}
\code{\link{oceSetMetadata}}, respectively, to be sure where the change will be
made, and to record the change in the \code{processingLog}.
}

\references{
[1] Anthony W. Isenor and David Kellow, 2011. ODF Format Specification
Version 2.0. (This is a .doc file downloaded from a now-forgotten URL by Dan Kelley,
in June 2011.)

[2] The St Lawrence Global Observatory website has information on ODF format at
\url{https://slgo.ca/app-sgdo/en/docs_reference/format_odf.html}

[3] List of variable codes:
\url{https://slgo.ca/app-sgdo/en/docs_reference/code_parametre_odf.html}
(checked 2018-02-11); only a subset are handled.
}
\seealso{
Other things related to \code{odf} data: \code{\link{ODF2oce}},
  \code{\link{ODFNames2oceNames}},
  \code{\link{[[,odf-method}},
  \code{\link{[[<-,odf-method}},
  \code{\link{plot,odf-method}},
  \code{\link{read.ctd.odf}}, \code{\link{read.odf}},
  \code{\link{subset,odf-method}},
  \code{\link{summary,odf-method}}

Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{coastline-class}}, \code{\link{ctd-class}},
  \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{met-class}}, \code{\link{oce-class}},
  \code{\link{rsk-class}}, \code{\link{sealevel-class}},
  \code{\link{section-class}}, \code{\link{topo-class}},
  \code{\link{windrose-class}}
}
\author{
Dan Kelley
}
