% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\name{ODFListFromHeader}
\alias{ODFListFromHeader}
\title{Create a list of ODF header metadata}
\usage{
ODFListFromHeader(header)
}
\arguments{
\item{header}{Vector of character strings, holding the header}
}
\value{
A list holding the metadata, with item names matching
those in the ODF header, except that duplicates are transformed
through the use of \code{\link{unduplicateNames}}.
}
\description{
Create a list of ODF header metadata
}
\seealso{
Other things related to \code{odf} data: \code{\link{ODF2oce}},
  \code{\link{ODFNames2oceNames}},
  \code{\link{[[,odf-method}},
  \code{\link{[[<-,odf-method}}, \code{\link{odf-class}},
  \code{\link{plot,odf-method}},
  \code{\link{read.ctd.odf}}, \code{\link{read.odf}},
  \code{\link{subset,odf-method}},
  \code{\link{summary,odf-method}}
}
