% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\name{ODFNames2oceNames}
\alias{ODFNames2oceNames}
\title{Translate ODF CODE strings to Oce Variable Names}
\usage{
ODFNames2oceNames(
  ODFnames,
  ODFunits = NULL,
  columns = NULL,
  PARAMETER_HEADER = NULL,
  debug = getOption("oceDebug")
)
}
\arguments{
\item{ODFnames}{Vector of strings holding ODF names.}

\item{ODFunits}{Vector of strings holding ODF units.}

\item{columns}{Optional list containing name correspondances, as described for
\code{\link[=read.ctd.odf]{read.ctd.odf()}}.}

\item{PARAMETER_HEADER}{Optional list containing information on the data variables.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
A vector of strings.
}
\description{
Translate ODF CODE strings to Oce Variable Names. In the BIO dialect, these
CODE strings are often identical to the NAME strings, but in the IML dialect,
this seems never to be the case.
}
\details{
The following table gives the recognized ODF code names for variables,
along with the translated names as used in oce objects. Note that the
code names are appended with strings such as \code{"_01"}, \code{"_02"}, etc,
for repeats. The converted name for an \code{"_01"} item is as shown below,
and for e.g. \code{"_02"} a suffix 2 is added to the oce name, etc.
Note that quality-control items (with names starting with \code{"QQQQ"})
are given matching names.
\tabular{lll}{
\strong{ODF name} \tab \strong{Oce name}      \tab \strong{Notes}                                    \cr
\code{ACO2} \tab \code{CO2Atmosphere}         \tab                                              \cr
\code{ALTB} \tab \code{altimeter}             \tab                                              \cr
\code{ALKW} \tab \code{alkalinity}            \tab                                              \cr
\code{AMON} \tab \code{ammonium}              \tab                                              \cr
\code{ATMP} \tab \code{pressureAtmosphere}    \tab                                              \cr
\code{ATTU} \tab \code{attenuation}           \tab                                              \cr
\code{AUTH} \tab \code{authority}             \tab                                              \cr
\code{BATH} \tab \code{barometricDepth}       \tab                                              \cr
\code{BEAM} \tab \code{a}                     \tab                                              \cr
\code{BNO7} \tab \code{bestNODC7Number}       \tab that is an "oh" letter, not a zero           \cr
\code{CNTR} \tab \code{scan}                  \tab                                              \cr
\code{CPHL} \tab \code{chlorophyll}           \tab                                              \cr
\code{CRAT} \tab \code{conductivity}          \tab Conductivity ratio (may have spurious unit)  \cr
\code{CMNT} \tab \code{comment}               \tab                                              \cr
\code{CNDC} \tab \code{conductivity}          \tab Conductivity in mS/cm or S/m                 \cr
\code{COND} \tab \code{conductivity}          \tab Conductivity in mS/cm or S/m                 \cr
\code{DCHG} \tab \code{discharge}             \tab                                              \cr
\code{DENS} \tab \code{density}               \tab                                              \cr
\code{DEPH} \tab \code{pressure}              \tab                                              \cr
\code{DOXY} \tab \code{oxygen}                \tab                                              \cr
\code{DPDT} \tab \code{dpdt}                  \tab                                              \cr
\code{DPWT} \tab \code{dryWeight}             \tab                                              \cr
\code{DRYT} \tab \code{temperatureDryBulb}    \tab                                              \cr
\code{ERRV} \tab \code{error}                 \tab                                              \cr
\code{EWCT} \tab \code{u}                     \tab                                              \cr
\code{FFFF} \tab \code{overall(FFFF)}         \tab Archaic overall flag, replaced by \code{QCFF}     \cr
\code{FLOR} \tab \code{fluorometer}           \tab                                              \cr
\code{GEOP} \tab \code{geopotential}          \tab                                              \cr
\code{GSPD} \tab \code{windSpeedGust}         \tab                                              \cr
\code{HCDM} \tab \code{directionMagnetic}     \tab                                              \cr
\code{HCDT} \tab \code{directionTrue}         \tab                                              \cr
\code{HCSP} \tab \code{speedHorizontal}       \tab                                              \cr
\code{HEAD} \tab \code{heading}               \tab                                              \cr
\code{IDEN} \tab \code{sampleNumber}          \tab                                              \cr
\code{LABT} \tab \code{temperatureLaboratory} \tab                                              \cr
\code{LATD} \tab \code{latitude}              \tab                                              \cr
\code{LHIS} \tab \code{lifeHistory}           \tab                                              \cr
\code{LOND} \tab \code{longitude}             \tab                                              \cr
\code{LPHT} \tab \code{pHLaboratory}          \tab                                              \cr
\code{NSCT} \tab \code{v}                     \tab                                              \cr
\code{NUM_} \tab \code{scansPerAverage}       \tab                                              \cr
\code{MNSV} \tab \code{retentionFilterSize}   \tab                                              \cr
\code{MNSZ} \tab \code{organismSizeMinimum}   \tab                                              \cr
\code{MODF} \tab \code{additionalTaxonomicInformation} \tab                                     \cr
\code{MXSZ} \tab \code{organismSizeMaximum}   \tab                                              \cr
\code{NONE} \tab \code{noWMOcode}             \tab                                              \cr
\code{NTRI} \tab \code{nitrite}               \tab                                              \cr
\code{NTRZ} \tab \code{nitrite+nitrate}       \tab                                              \cr
\code{OCUR} \tab \code{oxygenCurrent}         \tab                                              \cr
\code{OSAT} \tab \code{oxygenSaturation}      \tab                                              \cr
\code{OTMP} \tab \code{oxygenTemperature}     \tab                                              \cr
\code{OXYV} \tab \code{oxygenVoltage}         \tab                                              \cr
\code{OXV_} \tab \code{oxygenVoltageRaw}      \tab                                              \cr
\code{PCO2} \tab \code{CO2}                   \tab                                              \cr
\code{PHA_} \tab \code{phaeopigment}          \tab                                              \cr
\code{PHOS} \tab \code{phosphate}             \tab                                              \cr
\code{PHPH} \tab \code{pH}                    \tab                                              \cr
\code{PHT_} \tab \code{pHTotal}               \tab                                              \cr
\code{PHY_} \tab \code{phytoplanktonCount}    \tab                                              \cr
\code{POTM} \tab \code{theta}                 \tab                                              \cr
\code{PRES} \tab \code{pressure}              \tab                                              \cr
\code{PSAL} \tab \code{salinity}              \tab                                              \cr
\code{PSAR} \tab \code{par}                   \tab                                              \cr
\code{PTCH} \tab \code{pitch}                 \tab                                              \cr
\code{QCFF} \tab \code{overall(QCFF)}         \tab Overall flag (see also archaic FFFF)         \cr
\code{REFR} \tab \code{reference}             \tab                                              \cr
\code{RELH} \tab \code{humidityRelative}      \tab                                              \cr
\code{ROLL} \tab \code{roll}                  \tab                                              \cr
\code{SDEV} \tab \code{pressureStdDev}        \tab                                              \cr
\code{SECC} \tab \code{SecchiDepth}           \tab                                              \cr
\code{SEX_} \tab \code{sex}                   \tab                                              \cr
\code{SIGP} \tab \code{sigmaTheta}            \tab                                              \cr
\code{SIGT} \tab \code{sigmat}                \tab                                              \cr
\code{SLCA} \tab \code{silicate}              \tab                                              \cr
\code{SNCN} \tab \code{scanCounter}           \tab                                              \cr
\code{SPAR} \tab \code{SPAR}                  \tab                                              \cr
\code{SPFR} \tab \code{sampleFraction}        \tab                                              \cr
\code{SPVA} \tab \code{specificVolumeAnomaly} \tab                                              \cr
\code{SYTM} \tab \code{time}                  \tab                                              \cr
\code{TAXN} \tab \code{taxonomicName}         \tab                                              \cr
\code{TE90} \tab \code{temperature}           \tab                                              \cr
\code{TEMP} \tab \code{temperature}           \tab                                              \cr
\code{TICW} \tab \code{totalInorganicCarbon}  \tab                                              \cr
\code{TILT} \tab \code{tilt}                  \tab                                              \cr
\code{TOTP} \tab \code{pressureAbsolute}      \tab                                              \cr
\code{TRB_} \tab \code{turbidity}             \tab                                              \cr
\code{TRBH} \tab \code{trophicDescriptor}     \tab                                              \cr
\code{TSN_} \tab \code{taxonomicSerialNumber} \tab                                              \cr
\code{UNKN} \tab \code{-}                     \tab                                              \cr
\code{VAIS} \tab \code{BVFrequency}           \tab                                              \cr
\code{VCSP} \tab \code{w}                     \tab                                              \cr
\code{VMXL} \tab \code{waveHeightMaximum}     \tab                                              \cr
\code{VRMS} \tab \code{waveHeightMean}        \tab                                              \cr
\code{VTCA} \tab \code{wavePeriod}            \tab                                              \cr
\code{WDIR} \tab \code{windDirection}         \tab                                              \cr
\code{WSPD} \tab \code{windSpeed}             \tab                                              \cr
\code{WTWT} \tab \code{wetWeight}             \tab                                              \cr
\code{ZOO_} \tab \code{zooplanktonCount}      \tab                                              \cr
}
Any code not shown in the list is transferred to the oce object without renaming, apart from
the adjustment of suffix numbers. The following code have been seen in data files from
the Bedford Institute of Oceanography: \code{ALTB}, \code{PHPH} and \code{QCFF}.
}
\section{A note on unit conventions}{

Some older ODF files contain non-standard units for conductivity,
including \code{mho/m}, \code{mmho/cm}, and \code{mmHo}. As the
units for conductivity are important for derived quantities
(e.g. salinity), such units are converted to standard units
(e.g. \code{S/m} and \code{mS/cm}).  (This was once done with a warning,
but on 2020-02-07 the warning was removed, since it did not
indicate a problem with the file or the data scanning; rather,
it was a simple matter of nudging towards uniformity in a way
that ought to confuse no users, akin to converting \code{m**3} to
\code{m^3}, which is also done here without warning.)
}

\references{
For sources that describe the ODF format, see the documentation
for the \linkS4class{odf}.
}
\seealso{
Other functions that interpret variable names and units from headers: 
\code{\link{cnvName2oceName}()},
\code{\link{oceNames2whpNames}()},
\code{\link{oceUnits2whpUnits}()},
\code{\link{unitFromStringRsk}()},
\code{\link{unitFromString}()},
\code{\link{woceNames2oceNames}()},
\code{\link{woceUnit2oceUnit}()}

Other things related to odf data: 
\code{\link{CTD_BCD2014666_008_1_DN.ODF.gz}},
\code{\link{ODF2oce}()},
\code{\link{ODFListFromHeader}()},
\code{\link{[[,odf-method}},
\code{\link{[[<-,odf-method}},
\code{\link{odf-class}},
\code{\link{plot,odf-method}},
\code{\link{read.ctd.odf}()},
\code{\link{read.odf}()},
\code{\link{subset,odf-method}},
\code{\link{summary,odf-method}}
}
\author{
Dan Kelley
}
\concept{functions that interpret variable names and units from headers}
\concept{things related to odf data}
