\name{plot.nctd}
\docType{methods}
\alias{plot.nctd}
\alias{plot,nctd-method}

\title{Plot seawater CTD data}

\description{Plot a summary diagram for CTD data.}

\usage{\S4method{plot}{nctd}(x, which=1:4)}

\arguments{
  \item{x}{...}
  \item{which}{...}
}

\details{...}

\seealso{...}

\author{Dan Kelley}

\keyword{hplot}
