% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.nortek.ad2cp.R
\name{plotAD2CP}
\alias{plotAD2CP}
\title{Plot an AD2CP Object}
\usage{
plotAD2CP(x, which, col, ...)
}
\arguments{
\item{x}{an AD2CP object, as created with \code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}} or by
\code{\link[=read.oce]{read.oce()}} on a file of the AD2CP type.}

\item{which}{a character value indicating what to plot.  See
\dQuote{Details} and \dQuote{Examples}.}

\item{col}{passed to either \code{\link[=imagep]{imagep()}} if the graph is an image,
or to \code{\link[=oce.plot.ts]{oce.plot.ts()}} if the plot shows variation over time, or
otherwise to \code{\link[=plot]{plot()}}.}

\item{...}{optional other arguments, passed to the lower-level plotting
commands.}
}
\description{
Used by \code{\link{plot,adp-method}} or called directly, this function
provides a rough overview of some aspects of AD2CP data.  Given the
complexity of these data objects, and no attempt is made by the present
function to offer nuanced plot control, or even to offer the ability to plot
the full suite of data. Users are encouraged to use standard plotting tools,
if they need anything more than the default.  Two methods are permitted,
as explained in the \sQuote{Details}.
}
\details{
\emph{Method 1}: \code{which} is the name of one of the elements in \code{x[["data"]]}.
This must be one of the following: "burst", "average", "bottomTrack",
"interleavedBurst", "burstAltimeterRaw", "DVLBottomTrack", "echosounder",
"DVLWaterTrack", "altimeter", or "averageAltimeter".  Based on the contents
of that element, a choice is made as to what to plot.  See Example 1.

\emph{Method 2}: \code{which} is a string containing the character "/", with
the name of an element of \code{x[["data"]]} to the left, and the name
of a subelement to the right.  See Example 2.
}
\examples{
library(oce)
# This example will only work for the author, because it uses a
# private file.  The file contains 'burst' and 'average' data.
f <- "~/Dropbox/oce_secret_data/ad2cp/secret1_trimmed.ad2cp"
if (file.exists(f)) {
    d <- read.oce(f)
    # Example 1 (plots a 4-panel velocity plot)
    plot(d, which="average", col=oceColorsVelocity)
    # Example 2 (plots amplitude for beam 1)
    plot(d, which="average/a/1")
}

}
\author{
Dan Kelley
}
