% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{preferAdjusted}
\alias{preferAdjusted}
\title{Set Preference for Adjusted Values}
\usage{
preferAdjusted(argo, which = "all", fallback = TRUE)
}
\arguments{
\item{argo}{An \linkS4class{argo} object.}

\item{which}{A character vector naming the data items for which
(depending also on the value of \code{fallback}) adjusted values
are to be sought by future calls to \code{\link{[[,argo-method}}.
The short names are used, e.g. \code{which="oxygen"} means that
adjusted oxygen is to be returned in future calls
such as \code{argo[["oxygen"]]}.  The default,
\code{"all"}, means to  use adjusted values for any variable that
has such values.}

\item{fallback}{A logical value indicating whether to fall back
to unadjusted values for any data field in which the
adjusted values are all \code{NA}.  The default value, \code{TRUE},
makes it more likely that the results will be useful.}
}
\value{
An \linkS4class{argo} object its \code{metadata} slot altered
so that the \code{adjustedWhich} and \code{adjustedFallback} elements
are as set by the \code{which} and \code{fallback} arguments.
}
\description{
\code{preferAdjusted} sets \code{metadata} items named \code{"adjustedWhich"} and
\code{"adjustedFallback"} so that future data accesses with
\code{\link{[[,argo-method}} will return
adjusted data (and associated flags and units), if
indeed adjusted data exist.  Flags and units will also be
chosen to match.  Such a preference be registered for all
the data stored the object, or for a subset of them,
The preference carries through to computed variables, such as \code{SA},
that depend on the actual data.  See \dQuote{Details}
for the handling of cases that lack adjusted data, or for which
all the adjusted values for a given variable are \code{NA}.
}
\details{
Suppose that \code{which="all"}, or that \code{which} is a character
including the string \code{"salinity"}.  This indicates that
\code{\link{[[,argo-method}} should return adjusted
for the salinity data, and for its associated flags
and units.  Thus, \code{argo[["salinity"]]}
would return \code{argo@data$salinityAdjusted},
instead of \code{argo@data$salinity},
\code{argo[["salinityFlag"]]} would return
\code{argo@metadata$flags$salinityAdjusted}
instead of \code{argo@metadata$flags$salinity}, and
\code{argo[["salinityUnits"]]} would
\code{argo@metadata$units$salinityAdjusted}
instead of \code{argo@metadata$units$salinity}.

The default value \code{which="all"} indicates that later calls to
\code{\link{[[,argo-method}} should return adjusted
values for all data that have such values (returning the
unadjusted data, otherwise). This can be handy for quick
work, but analysts may also choose to restrict their use of
adjusted values to a subset of variables, e.g. if they wish
to adjust some things by themselves, based on unadjusted values.

The default value \code{fallback=TRUE} indicates that later calls to
\code{\link{[[,argo-method}} should return unadjusted values for any
data items that have \code{NA} for all the adjusted values.  This
condition is rare for core variables (salinity, temperature and
pressure) but is common for biogeochemical variables; see
e.g. Section 2.2.5 of reference 1 for a discussion of
the conditions under which Argo netcdf files contain
adjusted values. Setting \code{fallback=FALSE} means that adjusted
values (if they exist) will always be returned, even if they
are a useless collection of \code{NA} values.

Regardless of whether \code{selectAdjusted} has been used on an \linkS4class{argo}
object, it is possible to access either unadjusted or adjusted
data directly, using the original variable names stored in the
source netcdf file.  For example, \code{argo[["DOXY"]]}
yields unadjusted oxygen values, and
\code{argo[["DOXY_ADJUSTED"]]} yields adjusted values (if they exist, or
\code{NULL} if they do not).
Similarly, adjusted value can always be obtained by using a form
like \code{argo[["oxygenAdjusted"]]}.  In other words, this function
only has an affect on base names, such as \code{"oxygen"}.

Presently, this function only works for \linkS4class{argo} objects,
but it might be extended in future if other types of data
are seen to employ a similar scheme of supplying adjusted
values.
}
\references{
\enumerate{
\item Argo Data Management Team. "Argo User’s Manual V3.3." Ifremer,
November 28, 2019. \url{https://doi.org/10.13155/29825}.
}
}
\author{
Dan Kelley, based on discussions with Jaimie Harbin (with
respect to the \code{\link{[[,argo-method}} interface) and Clark Richards
(with respect to storing the preference in the \code{metadata} slot).
}
