% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{presentTime}
\alias{presentTime}
\title{Get the present time, in a stated timezone}
\usage{
presentTime(tz = "UTC")
}
\arguments{
\item{tz}{String indicating the desired timezone. The default is
to use UTC, which is used very commonly in oceanographic work. To
get the local time, use \code{tz=""} or \code{tz=NULL},}
}
\value{
A \code{\link{POSIXct}} object holding the present time, in the
indicated timezone.
}
\description{
Get the present time, in a stated timezone
}
\examples{
presentTime() # UTC
presentTime("") # the local timezone

}
