\name{processing.log.append}
\alias{processing.log.append}
\title{Append item to the processing log of object}
\description{
Appends a time-stamped item to the processing log of an object.
}
\usage{processing.log.append(x, action="")}
\arguments{
  \item{x}{An \code{oce} object.}
  \item{action}{A character string describing the action.}
}
\details{
This is used to add a time-stamped item to the processing log of the object.
}
\value{
A copy of the object, with the updated processing log.
}
\seealso{
}
\examples{
}
\references{
}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
