\name{log.summary}
\alias{log.summary}
\title{Summarize the processing log of an object}
\description{
Summarize the processing log of an object.
}
\usage{
log.summary(object)
}
\arguments{
\item{object}{An \code{oce} object.}
}
\details{
Prints log entries with UTC timestamps.
}
\value{
None.
}
\seealso{
}
\examples{
library(oce)
data(ctd)
log.summary(ctd)
}
\references{
}
\author{Dan Kelley}
\keyword{misc}
