% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingLog.R
\name{processingLogShow}
\alias{processingLogShow}
\title{Show the Processing Log of an oce Object}
\usage{
processingLogShow(x)
}
\arguments{
\item{x}{an \linkS4class{oce} object.}
}
\description{
Show the Processing Log of an oce Object
}
\seealso{
Other things related to processing logs: 
\code{\link{processingLog<-}()},
\code{\link{processingLogAppend}()},
\code{\link{processingLogItem}()}
}
\concept{things related to processing logs}
