\name{pt-class}

\docType{class}

\alias{pt-class}
\alias{[[,pt-method}

\title{Class to store pressure-temperature recorder data}

\description{Class to store pressure-temperature recorder data, with standard
    slots \code{metadata}, \code{data} and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{pt[["time"]]}, where the
    string could also be \code{"pressure"} or \code{"temperature"}, or
    abbreviations matching uniquely to any of these.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

