\name{read.bremen}

\alias{read.bremen}

\title{Read a Bremen data file (TEMPORARY CODE)}

\description{Read a Bremen data file (TEMPORARY CODE)}

\description{Read a CTD data file, producing an object of type \code{bremen}.}

\usage{read.bremen(file)}

\arguments{

  \item{file}{a connection or a character string giving the name of the file to
      load.}
}

\section{Issues}{This may be renamed (or removed) without notice.}

\value{An object of \code{\link[base]{class}} \code{"bremen"}.}

\author{Dan Kelley}

\keyword{misc}
