% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odf.R
\name{read.odf}
\alias{read.odf}
\title{Read an ODF file}
\usage{
read.odf(file, debug = getOption("oceDebug"))
}
\arguments{
\item{file}{the file containing the data.}

\item{debug}{a debugging flag, 0 for none, 1 for some debugging}
}
\value{
an object of class \code{oce}. It is up to a calling function to determine what to do with this object.
}
\description{
Read an ODF file
}
\details{
ODF (Ocean Data Format) is a
format developed at the Bedford Institute of Oceanography and also used
at other Department of Fisheries and Oceans (Canada) facilities.
It can hold various types of time-series data, which includes a variety
of instrument types. Thus, \code{read.odf()}
is used by \code{read.ctd.odf} for CTD data, etc.
}
\references{
Anthony W. Isenor and David Kellow, 2011. ODF Format Specification Version 2.0. (A .doc file downloaded from a now-forgotten URL by Dan Kelley, in June 2011.)
}

