% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel_gc_2026.R
\name{read.sealevel.gc2026}
\alias{read.sealevel.gc2026}
\title{Read a sealevel File in Government of Canada format 2026}
\usage{
read.sealevel.gc2026(file, debug = 0)
}
\arguments{
\item{file}{string vector of length 2, holding the names of the metadata file
and the data file.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
A \linkS4class{sealevel} object.
}
\description{
This is somewhat provisional code, to handle what seems to be
a new format as of the year 2026.  In this format, files come
in pairs, and so this function takes just the station identifier
and seeks files with names related to that, according to
a pattern observed on 2026-02-17. Whether this pattern will hold
is unclear.
}
\details{
The two input files are constructed from the value of the \code{stn} parameter.
For example, with \verb{stn="00491", the constructed filenames will be }"00491_metadata.csv"\code{and}"00491_data.csv"\verb{. Note that the first of these is read, with contents inserted into the }metadata` of the returned object,
but the items contained therein are not otherwise examined by this function.
Instead, the function determines station location from a header contained
within the "data" file.
}
\seealso{
Other things related to sealevel data: 
\code{\link{[[,sealevel-method}},
\code{\link{[[<-,sealevel-method}},
\code{\link{as.sealevel}()},
\code{\link{plot,sealevel-method}},
\code{\link{read.sealevel}()},
\code{\link{sealevel}},
\code{\link{sealevel-class}},
\code{\link{sealevelTuktoyaktuk}},
\code{\link{subset,sealevel-method}},
\code{\link{summary,sealevel-method}}
}
\author{
Dan Kelley and Chantelle Layton
}
\concept{things related to sealevel data}
