% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbt.R
\name{read.xbt.edf2}
\alias{read.xbt.edf2}
\title{Read an xbt File in Sippican Format Type 2}
\usage{
read.xbt.edf2(
  file,
  longitude = NA,
  latitude = NA,
  encoding = "latin1",
  debug = getOption("oceDebug"),
  processingLog
)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to
load.}

\item{longitude}{optional signed number indicating the longitude in degrees
East.}

\item{latitude}{optional signed number indicating the latitude in degrees North.}

\item{encoding}{a character value that indicates the encoding to be used for
this data file, if it is textual.  The default value for most functions is
\code{"latin1"}, which seems to be suitable for files containing text written in
English and French.}

\item{debug}{a flag that turns on debugging.  The value indicates the depth
within the call stack to which debugging applies.}

\item{processingLog}{if provided, the action item to be stored in the log.  This
parameter is typically only provided for internal calls; the default that it
provides is better for normal calls by a user.}
}
\value{
An \linkS4class{xbt} object.
}
\description{
The function was written by inspection of a particular file provided by a
user in late 2024. The format has been referred to as MK21 export format, and
a key difference to the format handled by \code{\link[=read.xbt.edf]{read.xbt.edf()}} is that data
columns are separated by tab characters, not spaces. The reading of header
data is more rudimentary than is the case for \code{\link[=read.xbt.edf]{read.xbt.edf()}}, because the
sample data file made available to the author did not seem to have much
useful information in its header.
}
\author{
Dan Kelley
}
