% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbt.R
\name{read.xbt.noaa2}
\alias{read.xbt.noaa2}
\title{Read an xbt File in UBT (Universal BathyThermograph) Format}
\usage{
read.xbt.noaa2(
  file,
  debug = getOption("oceDebug"),
  missingValue = -9.99,
  encoding = "latin1",
  processingLog
)
}
\arguments{
\item{file}{character value naming a file, or a file connection, containing the data.}

\item{debug}{a flag that turns on debugging.  The value indicates the depth
within the call stack to which debugging applies.}

\item{missingValue}{numerical value that is to be interpreted as \code{NA}}

\item{encoding}{a character value that indicates the encoding to be used for
this data file, if it is textual.  The default value for most functions is
\code{"latin1"}, which seems to be suitable for files containing text written in
English and French.}

\item{processingLog}{if provided, the action item to be stored in the log.  This
parameter is typically only provided for internal calls; the default that it
provides is better for normal calls by a user.}
}
\value{
A list containing \linkS4class{xbt} objects, one per line in \code{file}.
}
\description{
This file format, described at Reference 1, is used in NOAA archives, e.g.
the one at Reference 2. Note that, of the 4 subtypes that may
be indicated in the first 2 characters, only type \code{X } is handled. (FIXME?)
}
\details{
A sample file has the following as its first 3 lines.

\if{html}{\out{<div class="sourceCode">}}\preformatted{X 74335 N107715 W1197207110302131  446870001311075 3  11          +0016716B020104                12    01510   51490   81290  121050  17 710  18 650  22 610  25 520  30 490  48 430  88 400 136 380
X 74329 N107715 W1197207110412131  446870002311075 3  11          +0016616B020104                11    01460   51420   81120  10 610  13 540  14 500  20 460  26 390  59 390 180 400 187 390
X 74323 N107715 W1197207110427131  446870003311075 3  11          +0016616B020104                10    01590   41500   81250  12 580  15 490  20 410  82 400 121 400 166 400 194 400
}\if{html}{\out{</div>}}
}
\references{
\enumerate{
\item https://www.ncei.noaa.gov/data/oceans/nodc/formats/UBT_Universal_Bathythermograph.html
\item https://data.noaa.gov/onestop/collections/details/22c9b8d0-c32b-4824-815b-04ce80078d10
}
}
\seealso{
Other things related to xbt data: 
\code{\link{[[,xbt-method}},
\code{\link{[[<-,xbt-method}},
\code{\link{as.xbt}()},
\code{\link{plot,xbt-method}},
\code{\link{read.xbt}()},
\code{\link{read.xbt.noaa1}()},
\code{\link{subset,xbt-method}},
\code{\link{summary,xbt-method}},
\code{\link{xbt}},
\code{\link{xbt-class}},
\code{\link{xbt.edf}},
\code{\link{xbt2.edf}}
}
\author{
Dan Kelley
}
\concept{things related to xbt data}
