% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbt.R
\name{read.xbt}
\alias{read.xbt}
\title{Read an xbt file}
\usage{
read.xbt(file, type = "sippican", longitude = NA, latitude = NA,
  debug = getOption("oceDebug"), processingLog)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to
load.}

\item{type}{character string indicating type of file (only \code{"sippican"} is permitted).}

\item{longitude}{optional signed number indicating the longitude in degrees
East.}

\item{latitude}{optional signed number indicating the latitude in degrees North.}

\item{debug}{a flag that turns on debugging.  The value indicates the depth
within the call stack to which debugging applies.}

\item{processingLog}{if provided, the action item to be stored in the log.  This
parameter is typically only provided for internal calls; the default that it
provides is better for normal calls by a user.}
}
\value{
An \linkS4class{xbt} object.
}
\description{
In this version, only one type of file is handled, namely the \code{.edf} format, used
for Sippican XBT files.  This is handled by calling \code{\link[=read.xbt.edf]{read.xbt.edf()}}.
}
\examples{
library(oce)
xbt <- read.oce(system.file("extdata", "xbt.edf", package="oce"))
summary(xbt)
plot(xbt)

}
\references{
\enumerate{
\item Sippican, Inc. “Bathythermograph Data Acquisition System: Installation, Operation and Maintenance
Manual (P/N 308195, Rev. A),” 2003.
https://pages.uoregon.edu/drt/MGL0910_Science_Report/attachments/MK21_ISA_Manual_Rev_A.pdf.
}
}
\seealso{
Other things related to xbt data: \code{\link{[[<-,xbt-method}},
  \code{\link{plot,xbt-method}},
  \code{\link{subset,xbt-method}},
  \code{\link{summary,xbt-method}}, \code{\link{xbt-class}}
}
\author{
Dan Kelley
}
\concept{things related to xbt data}
