% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_odf}
\alias{read_odf}
\title{READ_ODF: Read in an ODF file.}
\usage{
read_odf(filename)
}
\arguments{
\item{filename}{location and name of the ODF file to be processed}
}
\description{
Description:
  Read in an ODF file.
}
\details{
ODSToolbox Version: 2.0

Last Updated: June 25, 2014.

Source:
  Ocean Data and Information Services,
  Bedford Institute of Oceanography, DFO, Canada.
  DataServicesDonnees@dfo-mpo.gc.ca

Notes:
  This program was totally re-designed and re-written for
  ODSToolbox Version 2.0. It is not based on Version 1.0.

  While this new version of read_odf corrects many errors in
  Version 1.0, and includes some new functionalities such as
  checking if all mandatory header blocks and all mandatory
  fields are presented in the input ODF file etc., it is possible
  that this program may have some conflicts with other tools in
  ODSToolbox, please email yongcun.hu@dfo-mpo.gc.ca if the
  user find any problems.

See also \code{\link{write_odf}}.

Copyright 2006-2010, DFO, Bedford Institute of Oceanography, Canada.
All Rights Reserved.
}
\author{
Yongcun Hu, Patrick Upson
}

