% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{renameData}
\alias{renameData}
\title{Rename items in the data slot of an oce object (defunct)}
\usage{
renameData(x, old = NULL, new = NULL)
}
\arguments{
\item{x}{Ignored, since this function is defunct.}

\item{old}{Ignored, since this function is defunct.}

\item{new}{Ignored, since this function is defunct.}
}
\description{
\strong{WARNING:} This function will be removed in oce 1.6.0; see \link{oce-defunct}.
}
\author{
Dan Kelley
}
\concept{deprecated functions}
