% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repositoryURL}
\alias{repositoryURL}
\title{Calculate a web link to download Oceanographic data}
\usage{
repositoryURL(domain, subdomain, item, window, id)
}
\arguments{
\item{domain}{The base-level domain (more information to be added later).}

\item{subdomain}{The second-level domain (more information to be added later).}

\item{item}{A string indicating the item sought; at present only \code{"argo"} is allowed.}

\item{window}{An indication of a window to select data (ignored at present).}

\item{id}{A string indicating the ID of the sought item.}
}
\description{
This function is able to construct universal resource identifiers (URIs)
that can be used with \code{\link{download.file}} to download data from
several common Oceanographic repositories.
}
\details{
The goal is to spare users the task of finding such URIs for casual use
or one-off applications.  There is no intention of trying to reproduce
the web tools provided with many repositories.

This function will be most useful if users contribute to it, both by 
pointing out changes to URLs and by suggesting additions for useful
datasets.
}
\examples{
file <- repositoryURL(item="argo", id=list(id="6900388"))
\dontrun{
download.file(file, "argo.nc")
argo <- read.oce("argo.nc")
summary(argo)
plot(argo)
}
}
\author{
Dan Kelley
}
\seealso{
\code{\link{download.file}} for downloading the datafiles found by \code{repositoryURI}
}

