% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\docType{data}
\name{section}
\alias{section}
\title{Sample section Data}
\source{
The following code was used to download the datafile and create the \code{section}
object.

\if{html}{\out{<div class="sourceCode">}}\preformatted{download.file("https://cchdo.ucsd.edu/data/7872/a03_hy1.csv",
    "a03_hy1.csv")
section <- read.section("a03_hy1.csv",
    sectionId = "a03", institute = "SIO",
    ship = "R/V Professor Multanovskiy",
    scientist = "Vladimir Tereschenkov")
}\if{html}{\out{</div>}}
}
\usage{
data(section)
}
\description{
This is from a 1993 occupation of WOCE line A03 that ran westward across the
Atlantic at approximately 36 N, ending with a turn to the northwest in the
Gulf Stream region.
}
\details{
The sampling times in this dataset are not fully sequential. For example,
Station 41 is reported to be at \verb{1993-10-03T00:06:00}, which, going by
station numbers, suggests an error in the reported day. Station 45 seems to
have a similar problem. However, the goal here is to represent the data as
archived, so no changes are made to the times.
}
\examples{
library(oce)
# Gulf Stream
data(section)
GS <- subset(section, 113 <= stationId & stationId <= 129)
GSg <- sectionGrid(GS, p = seq(0, 5000, 100))
plot(GSg, span = 1500) # increase span to show more coastline

}
\seealso{
Other datasets provided with oce: 
\code{\link{adp}},
\code{\link{adv}},
\code{\link{amsr}},
\code{\link{argo}},
\code{\link{cm}},
\code{\link{coastlineWorld}},
\code{\link{ctd}},
\code{\link{ctdRaw}},
\code{\link{echosounder}},
\code{\link{landsat}},
\code{\link{lisst}},
\code{\link{lobo}},
\code{\link{met}},
\code{\link{ocecolors}},
\code{\link{rsk}},
\code{\link{sealevel}},
\code{\link{sealevelTuktoyaktuk}},
\code{\link{topoWorld}},
\code{\link{wind}},
\code{\link{xbt}}

Other things related to section data: 
\code{\link{[[,section-method}},
\code{\link{[[<-,section-method}},
\code{\link{as.section}()},
\code{\link{handleFlags,section-method}},
\code{\link{initializeFlagScheme,section-method}},
\code{\link{plot,section-method}},
\code{\link{read.section}()},
\code{\link{section-class}},
\code{\link{sectionAddStation}()},
\code{\link{sectionGrid}()},
\code{\link{sectionSmooth}()},
\code{\link{sectionSort}()},
\code{\link{subset,section-method}},
\code{\link{summary,section-method}}
}
\concept{datasets provided with oce}
\concept{things related to section data}
