% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{setFlagScheme,oce-method}
\alias{setFlagScheme,oce-method}
\title{Establish a data-quality scheme for a oce object}
\usage{
\S4method{setFlagScheme}{oce}(object, scheme = scheme,
  debug = getOption("oceDebug"))
}
\arguments{
\item{object}{An oce object.}

\item{scheme}{A list describing the flag scheme, e.g \code{list(good=1, bad=2)}
might be used for a hypothetical class. Some classes also permit \code{scheme}
to be a character string that names a built-in scheme; see \dQuote{Details}.}

\item{debug}{Integer set to 0 for quiet action or to 1 for some debugging.}
}
\value{
An object with the \code{metadata} slot set to the indicated
\code{scheme} (expanded to a list, if the argument was a recognized
character string).
}
\description{
This function stores adds an item called \code{flagScheme} to the
\code{metadata} slot. The purpose is both to document a flag scheme
and to make it so that \code{setFlags} can specify flags by
name, in addition to number. A generic function, it is
specialized for some classes via interpretation of the
\code{scheme} argument (see \dQuote{Details}, for those
object classes that have such specializations).
}
\details{
There are no pre-defined \code{scheme}s for this object class.
}
\section{Caution}{

This function was added in early May, 2018, and is likely to undergo
changes until the mid-summer of that year.  Use with caution.
}

\seealso{
Other functions relating to data-quality flags: \code{\link{handleFlags,adp-method}},
  \code{\link{handleFlags,argo-method}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{handleFlags,section-method}},
  \code{\link{handleFlags}},
  \code{\link{setFlagScheme,ctd-method}},
  \code{\link{setFlagScheme}},
  \code{\link{setFlags,adp-method}},
  \code{\link{setFlags,ctd-method}},
  \code{\link{setFlags,oce-method}}, \code{\link{setFlags}}
}
