% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{snakeToCamel}
\alias{snakeToCamel}
\title{Convert each of a vector of strings from SNAKE_CASE to camelCase}
\usage{
snakeToCamel(s)
}
\arguments{
\item{s}{A vector of character values.}
}
\value{
A vector of character values
}
\description{
This function is mainly for use by \code{\link[=read.argo]{read.argo()}}, but it may be helpful
in other contexts as well.
}
\details{
The basic procedure is to chop the string up into substrings separated by
the underline character, then to upper-case the first letter of
all substrings except the first, and then to paste the substrings
together.

However, there are exceptions.  First, any upper-case string that contains no
underlines is converted to lower case, but any mixed-case string with no
underlines is returned as-is (see the second example). Second,
\code{QC} is passed through directly (since it is an acronym) and
if the first letter of remaining text is upper-cased (see the last
two examples).
}
\examples{
library(oce)
snakeToCamel("PARAMETER_DATA_MODE") # "parameterDataMode"
snakeToCamel("PARAMETER")           # "parameter"
snakeToCamel("HISTORY_QCTEST")      # "historyQCTest"
snakeToCamel("PROFILE_DOXY_QC")     # "profileDoxyQC"
}
\author{
Dan Kelley
}
