\name{standardDepths}

\alias{standardDepths}

\title{Standard oceanographic depths}

\description{Standard oceanographic depths}

\usage{standardDepths()}

\details{This returns so-called standard deptsh 0m, 10m, etc. below the sea
    surface.}

\value{A vector of depths, c(0, 10, ...).}

\author{Dan Kelley}

\keyword{misc}
