% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{standardizeLongitude}
\alias{standardizeLongitude}
\title{Put longitude in the range from -180 to 180}
\usage{
standardizeLongitude(longitude)
}
\arguments{
\item{longitude}{in degrees East, possibly exceeding 180}
}
\value{
longitude in signed degrees East
}
\description{
Put longitude in the range from -180 to 180
}
\seealso{
\code{\link{matrixShiftLongitude}} and \code{\link{shiftLongitude}} are more
powerful relatives to \code{standardizeLongitude}.
}
