% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\docType{methods}
\name{[[,adp-method}
\alias{[[,adp-method}
\title{Extract Parts of an \code{adp} Object}
\usage{
\S4method{[[}{adp}(x, i, j, ...)
}
\arguments{
\item{x}{An adp object, i.e. one inheriting from \code{\link{adp-class}}.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
In addition to the usual extraction of elements by name, some shortcuts
are also provided, e.g. \code{u1} retrieves \code{v[,1]}, and similarly
for the other velocity components. The \code{a} and \code{q}
data can be retrived in \code{\link{raw}} form or numeric
form; see examples.
}
\examples{
data(adp)
# Tests for beam 1, distance bin 1, first 5 observation times
adp[["v"]][1:5,1,1]
adp[["a"]][1:5,1,1]
adp[["a", "numeric"]][1:5,1,1]
as.numeric(adp[["a"]][1:5,1,1]) # same as above

}
\author{
Dan Kelley
}
\seealso{
Other functions that extract parts of oce objects: \code{\link{[[,ladp-method}}

Other things related to \code{adp} data: \code{\link{[[<-,adp-method}},
  \code{\link{adp-class}}, \code{\link{adp}},
  \code{\link{as.adp}}, \code{\link{plot,adp-method}},
  \code{\link{read.adp.nortek}},
  \code{\link{read.adp.rdi}},
  \code{\link{read.adp.sontek.serial}},
  \code{\link{read.adp.sontek}}, \code{\link{read.adp}},
  \code{\link{read.aquadoppHR}},
  \code{\link{read.aquadoppProfiler}},
  \code{\link{read.aquadopp}},
  \code{\link{subset,adp-method}},
  \code{\link{summary,adp-method}},
  \code{\link{xyzToEnuAdp}}
}

