% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amsr.R
\docType{methods}
\name{[[,amsr-method}
\alias{[[,amsr-method}
\title{Extract something from an amsr object}
\usage{
\S4method{[[}{amsr}(x, i, j, ...)
}
\arguments{
\item{x}{An amsr object, i.e. one inheriting from \code{\link{amsr-class}}.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item (ignored).}

\item{...}{Optional additional information (ignored).}
}
\description{
Extract something from an amsr object
}
\author{
Dan Kelley
}

