% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{[[,ctd-method}
\alias{[[,ctd-method}
\title{Extract Something From a CTD Object}
\usage{
\S4method{[[}{ctd}(x, i, j, ...)
}
\arguments{
\item{x}{A ctd object, i.e. one inheriting from \code{\link{ctd-class}}.}

\item{i}{Character string indicating the name of item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
The \code{[[} notation works for all \code{oce} objects. It looks up
information in a two-step process. First, a class-specific function 
is used to try to access the requested information (see
\dQuote{Details of the specialized method}). Then, if no match is found, a
general function is used (see \sQuote{Details of the general
method}). If neither function is successful in locating
the desired item, then \code{NULL} is returned.
}
\section{Details of the general method}{


If the specialized method produces no matches, the following generalized
method is applied. As with the specialized method, the procedure hinges
first on the value of \code{i}.

First, a check is made as to whether \code{i} names one of the standard
\code{oce} slots, and returns the slot contents if so. Thus, 
\code{x[["metadata"]]} will retrieve the \code{metadata} slot, 
while \code{x[["data"]]} and \code{x[["processingLog"]]} return
those slots.

Then, a search is made of the items in the \code{metadata} slot. A direct
match to the item named in \code{i} will retrieve the corresponding item.
For example, a common use for \code{\link{ctd-class}} data is
\code{x[["station"]]}, which looks up the station ID.

Finally, a partial-match search is done of the \code{data} slot, using
\code{\link{pmatch}}.

If none of these three conditions holds, \code{NULL} is returned.
}

\section{Details of the specialized ctd method}{

As noted above, this method is applied before the general method. The
purpose is to permit calculation of secondary data that are not
actually stored in the object.  For example, ctd data conventionally
store in-situ temperature, which is retrieved by \code{x[["temperature"]]}
which is a direct lookup of the data variable of that name. However,
\code{x[["t"]]} can be used as a synonym for this, and \code{x[["temperature68"]]}
can be used to convert to the IPTS-1968 temperature scale. Potential temperature
in the UNESCO formulation may be retrived with \code{x[["theta"]]} or
\code{x[["potential temperature"]]}, while the Gibbs Seawater (GSW) temperature
parameter called Conservative Temperature can be retrieved with 
\code{x[["CT"]]} or \code{x[["conservative temperature"]]}. Similarly,
the GSW reference salinity and preformed salinity are retrieve with
\code{x[["SR"]]} and \code{x[["Sstar"]]}.

The z coordinate is retrieved with \code{x[["z"]]} and the depth
with \code{x[["depth"]]}. These are calculated with 
with \code{\link{swZ}}) and \code{\link{swDepth}}, respectively.

In addition, some specialized quantities are also provided, e.g. the
density ratio \code{x[["Rrho"]]} (computed with \code{\link{swRrho}}),
spice \code{x[["spice"]]} (computed with \code{\link{swSpice}}).
}
\examples{
data(ctd)
head(ctd[["temperature"]])

}
\author{
Dan Kelley
}
\seealso{
Other functions that access data within oce objects: \code{\link{[[,section-method}}

Other functions that handle CTD data: \code{\link{as.ctd}},
  \code{\link{ctd-class}}, \code{\link{ctdAddColumn}},
  \code{\link{ctdDecimate}}, \code{\link{ctdFindProfiles}},
  \code{\link{ctdTrim}}, \code{\link{ctdUpdateHeader}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.odv}}, \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}}, \code{\link{write.ctd}}
}

