% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladp.R
\docType{methods}
\name{[[,ladp-method}
\alias{[[,ladp-method}
\title{Extract Something From an ladp Object}
\usage{
\S4method{[[}{ladp}(x, i, j, ...)
}
\arguments{
\item{x}{A ctd object, i.e. one inheriting from \code{\link{ladp-class}}.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
Extract Something From an ladp Object
}
\examples{
data(ctd)
head(ctd[["temperature"]])

}
\author{
Dan Kelley
}
\seealso{
Other functions that access oce data and metadata: \code{\link{[[,adp-method}},
  \code{\link{[[,adv-method}},
  \code{\link{[[,argo-method}},
  \code{\link{[[,section-method}}

Other functions that handle ladp data: \code{\link{as.ladp}},
  \code{\link{ladp-class}}, \code{\link{plot,ladp-method}},
  \code{\link{summary,ladp-method}}
}

