% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat.R
\name{[[,landsat-method}
\alias{[[,landsat-method}
\title{Extract Something From a landsat Object}
\usage{
\S4method{[[}{landsat}(x, i, j, ...)
}
\arguments{
\item{x}{a \linkS4class{landsat} object.}

\item{i}{character value indicating the name of an item to extract.}

\item{j}{optional additional information on the \code{i} item.}

\item{...}{ignored.}
}
\description{
Generally, the \code{[[} method lets users extract information from \code{oce}
objects, without having to know the details of the internal storage.  For
many \code{oce} sub-classes, \code{[[} can also return quantities that are computed
from the object's contents.
}
\details{
A two-step process is used to try to find the requested information. First, a
class-specific function is used (see \dQuote{Details of the Specialized
Method}).  If this yields nothing, then a general method is used (see
\dQuote{Details of the General Method}).  If both methods fail, then \code{[[}
returns NULL.

Some understanding of the subclass is required to know what can be retrieved
with \code{[[}. When dealing with an unfamiliar subclass, it can be useful to
first use \code{x[["?"]]} to get a listing of the retrievable items. See
\dQuote{Details of the Specialized Method} for more information.
}
\section{Details of the Specialized Method}{


If \code{i} is \code{"?"}, then the return value is a list
containing four items, each of which is a character vector
holding the names of things that can be accessed with \code{[[}.
The \code{data} and \code{metadata} items hold the names of
entries in the object's data and metadata
slots, respectively. The \code{data} entries are difficult
to deal with directly, and so users are advised to
use \code{dataDerived} instead.

\emph{Accessing band data.}  The data may be accessed with e.g.
\code{landsat[["panchromatic"]]}, for the panchromatic band.  If a new
``band'' is added with \code{\link[=landsatAdd]{landsatAdd()}}, it may be referred by
name.  In all cases, a second argument can be provided, to govern
decimation.  If this is missing, all the relevant data are returned.  If
this is present and equal to \code{TRUE}, then the data will be
automatically decimated (subsampled) to give approximately 800 elements in
the longest side of the matrix.  If this is present and numerical, then its
value governs decimation.  For example,
\code{landsat[["panchromatic",TRUE]]} will auto-decimate, typically
reducing the grid width and height from 16000 to about 800.  Similarly,
\code{landsat[["panchromatic",10]]} will reduce width and height to about
1600.  On machines with limited RAM (e.g. under about 6GB), decimation is a
good idea in almost all processing steps.  It also makes sense for
plotting, and in fact is done through the `decimate` argument of
\code{\link[=plot,landsat-method]{plot,landsat-method()}}.

\emph{Accessing derived data.}  One may retrieve several derived quantities
that are calculated from data stored in the object:
\code{landsat[["longitude"]]} and \code{landsat[["latitude"]]} give pixel
locations.  Accessing \code{landsat[["temperature"]]} creates an estimate
of ground temperature as follows (see reference 4).  First, the ``count value'' in
band 10, denoted \eqn{b_{10}}{b_10} say, is scaled with coefficients stored
in the image metadata using
\eqn{\lambda_L=b_{10}M_L+A_L}{lambda_L=b_10*M_L+A_L} where \eqn{M_L}{M_L}
and \eqn{A_L}{A_L} are values stored in the metadata (e.g.  the first in
\code{landsat@metadata$header$radiance_mult_band_10}) Then the result is
used, again with coefficients in the metadata, to compute Celcius
temperature \eqn{T=K_2/ln(\epsilon
    K_1/\lambda_L+1)-273.15}{T=K_2/ln(epsilon*K_1/\lambda_L+1)-273.15}.
The value of the emissivity \eqn{\epsilon}{epsilon} is set to unity by
\code{\link[=read.landsat]{read.landsat()}}, although it can be changed easily later, by
assigning a new value to `landsat@metadata$emissivity`. The default
emissivity value set by \code{\link[=read.landsat]{read.landsat()}} is from reference 11, and is
within the oceanic range suggested by reference 5. Adjustment is as simple as
altering `landsat@metadata$emissivity`. This value can be a single
number meant to apply for the whole image, or a matrix with dimensions
matching those of band 10.  The matrix case is probably more useful for
images of land, where one might wish to account for the different
emissivities of soil and vegetation, etc.; for example, Table 4 of
reference 9 lists 0.9668 for soil and 0.9863 for vegetation,
while Table 5 of reference 10 lists 0.971 and 0.987 for the same quantities.

\emph{Accessing metadata.} Anything in the metadata can be accessed by
name, e.g. \code{landsat[["time"]]}.  Note that some items are simply
copied over from the source data file and are not altered by e.g.
decimation.  An exception is the lat-lon box, which is altered by
\code{\link[=landsatTrim]{landsatTrim()}}.
}

\section{Details of the General Method}{


Note: the text of this section is identical for all \code{oce} subclasses, and so
some of what you read here may not be relevant to the class being described
in this help page.

If the specialized method produces no matches, the following generalized
method is applied. As with the specialized method, the procedure hinges first
on the values of \code{i} and, optionally, \code{j}. The work proceeds in steps, by
testing a sequence of possible conditions in sequence.
\enumerate{
\item A check is made as to whether \code{i} names one of the standard \code{oce} slots.
If so, \code{[[} returns the slot contents of that slot.  Thus, \code{x[["metadata"]]}
will retrieve the \code{metadata} slot, while \code{x[["data"]]} and
\code{x[["processingLog"]]} return those slots.
\item If \code{i} is a string ending in the \code{"Unit"}, then the characters preceding
that string are taken to be the name of an item in the data object, and a
list containing the unit is returned (or \code{NULL} if there is no such unit).
This list consists of an item named \code{unit}, which is an \code{\link[=expression]{expression()}}, and
an item named \code{scale}, which is a string describing the measurement scale.
If the string ends in \code{" unit"}, e.g.  \code{x[["temperature unit"]]} (note the
space), then just the expression is returned, and if it ends in \code{" scale"},
then just the scale is returned.
\item If \code{i} is a string ending in \code{"Flag"}, then the corresponding data-quality
flag is returned (or \code{NULL} if there is no such flag).
\item If the object holds hydrographic information (pressure, salinity,
temperature, longitude and latitude) then another set of possibilities
arises.  If \code{i} is \code{"sigmaTheta"}, then the value of \code{\link[=swSigmaTheta]{swSigmaTheta()}} is
called with \code{x} as the sole argument, and the results are returned.
Similarly, \code{\link[=swSigma0]{swSigma0()}} is used if \code{i="sigma0"}, and \code{\link[=swSpice]{swSpice()}} is used if
\code{i="spice"}. Of course, these actions only make sense for objects that
contain the relevant items within their \code{data} slot.
\item After these possibilities are eliminated, the action depends on whether
\code{j} has been provided.  If \code{j} is not provided, or is the string \code{""}, then
\code{i} is sought in the \code{metadata} slot, and then in the \code{data} slot, returning
whichever is found first.  In other words, if \code{j} is not provided, the
\code{metadata} slot takes preference over the \code{data} slot. However, if \code{j} is
provided, then it must be either the string \code{"metadata"} or \code{"data"}, and it
directs where to look.
\item If none of the above-listed conditions holds, then \code{NULL} is returned.
}
}

\seealso{
Other functions that extract parts of oce objects: 
\code{\link{[[,adp-method}},
\code{\link{[[,adv-method}},
\code{\link{[[,amsr-method}},
\code{\link{[[,argo-method}},
\code{\link{[[,bremen-method}},
\code{\link{[[,cm-method}},
\code{\link{[[,coastline-method}},
\code{\link{[[,ctd-method}},
\code{\link{[[,echosounder-method}},
\code{\link{[[,g1sst-method}},
\code{\link{[[,gps-method}},
\code{\link{[[,ladp-method}},
\code{\link{[[,lisst-method}},
\code{\link{[[,lobo-method}},
\code{\link{[[,met-method}},
\code{\link{[[,oce-method}},
\code{\link{[[,odf-method}},
\code{\link{[[,rsk-method}},
\code{\link{[[,sealevel-method}},
\code{\link{[[,section-method}},
\code{\link{[[,tidem-method}},
\code{\link{[[,topo-method}},
\code{\link{[[,windrose-method}},
\code{\link{[[,xbt-method}},
\code{\link{[[<-,adv-method}}

Other things related to landsat data: 
\code{\link{[[<-,landsat-method}},
\code{\link{landsat-class}},
\code{\link{landsatAdd}()},
\code{\link{landsatTrim}()},
\code{\link{landsat}},
\code{\link{plot,landsat-method}},
\code{\link{read.landsat}()},
\code{\link{summary,landsat-method}}
}
\author{
Dan Kelley
}
\concept{functions that extract parts of oce objects}
\concept{things related to landsat data}
