% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\docType{methods}
\name{[[,section-method}
\alias{[[,section-method}
\title{Extract Something From a section Object}
\usage{
\S4method{[[}{section}(x, i, j, ...)
}
\arguments{
\item{x}{A section object, i.e. one inheriting from \code{\link{section-class}}.}

\item{i}{The item to extract. This may be the name of a variable stored in the station
data, or the word \code{"station"}, which will return a \code{\link{list}} of 
\code{\link{ctd-class}} objects holding the station data, if \code{j} is not specified,
or the j-th station in the section, if \code{j} is given.}

\item{j}{Optional additional information on the \code{i} item. If this is \code{"byStation"}
and \code{i} is either \code{"latitude"} or \code{"longitude"}, then only one value
is returned for each station.}

\item{...}{Optional additional information (ignored).}
}
\description{
Extract Something From a section Object
}
\examples{
data(section)
length(section[["latitude"]])
length(section[["latitude", "byStation"]])

}
\author{
Dan Kelley
}
\seealso{
Other functions that access oce data and metadata: \code{\link{[[,adp-method}},
  \code{\link{[[,adv-method}},
  \code{\link{[[,argo-method}}, \code{\link{[[,ctd-method}}

Other functions that handle section data: \code{\link{as.section}},
  \code{\link{plot,section-method}},
  \code{\link{read.section}},
  \code{\link{sectionAddStation}},
  \code{\link{sectionGrid}}, \code{\link{sectionSmooth}},
  \code{\link{sectionSort}},
  \code{\link{subset,section-method}},
  \code{\link{summary,section-method}}
}

