% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\docType{methods}
\name{[[<-,argo-method}
\alias{[[<-,argo-method}
\title{Change something within an argo object}
\usage{
\S4method{[[}{argo}(x, i, j) <- value
}
\arguments{
\item{x}{An argo object, i.e. one inheriting from \code{\link{argo-class}}.}

\item{i}{The item to insert}

\item{j}{Optional additional information on the \code{i} item.}

\item{value}{The value recoverd from \code{x}.}

\item{...}{Optional additional information (ignored).}
}
\description{
In addition to the usual insertion of elements by name, note
that e.g. \code{pitch} gets stored into \code{pitchSlow}.
}
\seealso{
Other functions that alter oce data and metadata: \code{\link{[[<-,adp-method}},
  \code{\link{[[<-,adv-method}},
  \code{\link{[[<-,ctd-method}}

Other functions that handle argo data: \code{\link{[[,argo-method}},
  \code{\link{argoGrid}}, \code{\link{as.argo}},
  \code{\link{plot,argo-method}}, \code{\link{read.argo}},
  \code{\link{subset,argo-method}},
  \code{\link{summary,argo-method}}
}

