% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{[[<-,ctd-method}
\alias{[[<-,ctd-method}
\title{Change Something within a ctd object}
\usage{
\S4method{[[}{ctd}(x, i, j) <- value
}
\arguments{
\item{x}{A ctd object}

\item{i}{The item to insert}

\item{j}{Optional additional information on the \code{i} item.}

\item{value}{The value to be inserted into \code{x}.}

\item{...}{Optional additional information (ignored).}
}
\description{
Change Something within a ctd object
}
\seealso{
Other functions that alter oce data and metadata: \code{\link{[[<-,adp-method}},
  \code{\link{[[<-,adv-method}},
  \code{\link{[[<-,argo-method}}
}

