% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\docType{methods}
\name{subset,adp-method}
\alias{subset,adp-method}
\title{Subset an adp object}
\usage{
\S4method{subset}{adp}(x, subset, ...)
}
\arguments{
\item{x}{An \code{\link{adp-class}} object.}

\item{subset}{A condition to be applied to the \code{data} portion of
\code{x}.  See \sQuote{Details}.}

\item{...}{Ignored.}
}
\value{
A new \code{\link{adp-class}} object.
}
\description{
Subset an adp (acoustic Doppler profile) object, in a manner that is function
is somewhat analogous to \code{\link{subset.data.frame}}.  Subsetting can be by
\code{time} or \code{distance}, but these may not be combined; use a sequence
of calls to subset by both.
}
\examples{
library(oce)
data(adp)
# First part of time series
plot(subset(adp, time < mean(range(adp[['time']]))))

}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{adp} data: \code{\link{[[,adp-method}},
  \code{\link{[[<-,adp-method}}, \code{\link{adp-class}},
  \code{\link{adp}}, \code{\link{as.adp}},
  \code{\link{xyzToEnuAdp}}
}

