% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\docType{methods}
\name{subset,odf-method}
\alias{subset,odf-method}
\title{Subset an ODF object}
\usage{
\S4method{subset}{odf}(x, subset, ...)
}
\arguments{
\item{x}{a \code{odf} object.}

\item{subset}{a condition to be applied to the \code{data} portion of
\code{x}.  See \sQuote{Details}.}

\item{\dots}{ignored.}
}
\value{
A new \code{odf} object.
}
\description{
This function is somewhat analogous to \code{\link{subset.data.frame}}.
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{odf} data: \code{\link{ODF2oce}},
  \code{\link{[[,odf-method}},
  \code{\link{[[<-,odf-method}}, \code{\link{odf-class}},
  \code{\link{plot,odf-method}},
  \code{\link{summary,odf-method}}
}

