% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk.R
\docType{methods}
\name{subset,rsk-method}
\alias{subset,rsk-method}
\title{Subset a Rsk Object}
\usage{
\S4method{subset}{rsk}(x, subset, ...)
}
\arguments{
\item{x}{a \code{rsk} object, i.e. inheriting from \code{\link{rsk-class}}.}

\item{subset}{a condition to be applied to the \code{data} portion of \code{x}.
See \sQuote{Details}.}

\item{\dots}{ignored.}
}
\value{
A new \code{rsk} object.
}
\description{
Subset a rsk object.  This function is somewhat analogous to
\code{\link{subset.data.frame}}, but subsetting is only permitted by time.
}
\examples{
library(oce)
data(rsk)
plot(rsk)
plot(subset(rsk, time < mean(range(rsk[['time']]))))

}
\seealso{
Other things related to \code{rsk} data: \code{\link{[[,rsk-method}},
  \code{\link{[[<-,rsk-method}}, \code{\link{as.rsk}},
  \code{\link{plot,rsk-method}}, \code{\link{read.rsk}},
  \code{\link{rsk-class}}, \code{\link{rskPatm}},
  \code{\link{rskToc}}, \code{\link{rsk}},
  \code{\link{summary,rsk-method}}
}
\author{
Dan Kelley
}
