% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\docType{methods}
\name{subset,section-method}
\alias{subset,section-method}
\title{Subset a Section Object}
\usage{
\S4method{subset}{section}(x, subset, ...)
}
\arguments{
\item{x}{A \code{section} object, i.e. one inheriting from \code{\link{section-class}}.}

\item{subset}{an optional indication of either (a) stations to be kept,
or (b) data to be kept within the stations.  Mode (a)
is enabled if the subset string contains \code{distance}
(in which case stations are kept only when the distance in km
from the first stations matches the specification), either
\code{longitude} or \code{latitude} (in which case stations
are selected by location), or \code{stationId} (in which case stations
are selected by the station-ID stored within the CTDs within
the section). If none of these words is present
in \code{subset}, then mode (b) is used, and the subset is applied
within each station, by passing it as the argument of the same
name to \code{\link{subset,ctd-method}}.}

\item{...}{optional arguments, of which only the first is examined. The only
possibility is that this argument be named \code{indices}, and that it be
a numeric or logical vector specifying the indices of stations to be kept,
starting at 1 for the first station.}
}
\value{
A \code{\link{section-class}} object.
}
\description{
Return a subset of a section object.
}
\details{
This function that may be used to subset data within the
stations of a section, or to choose a subset of the stations
themselves. The first case is handled with the \code{subset} argument,
while the second is handled if \code{...} contains a vector named
\code{indices}.
}
\examples{
library(oce)
data(section)
# Gulf Stream
GS <- subset(section, 109<=stationId&stationId<=129)
# Sections with more than 5 levels
long <- subset(section,
   indices=unlist(lapply(section[["station"]], function(s) 10<length(s[["pressure"]]))))

}
\seealso{
Other functions that subset \code{oce} objects: \code{\link{subset,adp-method}},
  \code{\link{subset,adv-method}},
  \code{\link{subset,amsr-method}},
  \code{\link{subset,argo-method}},
  \code{\link{subset,cm-method}},
  \code{\link{subset,coastline-method}},
  \code{\link{subset,ctd-method}},
  \code{\link{subset,echosounder-method}},
  \code{\link{subset,lobo-method}},
  \code{\link{subset,met-method}},
  \code{\link{subset,oce-method}},
  \code{\link{subset,odf-method}},
  \code{\link{subset,rsk-method}},
  \code{\link{subset,sealevel-method}},
  \code{\link{subset,topo-method}}

Other things related to \code{section} data: \code{\link{[[,section-method}},
  \code{\link{[[<-,section-method}},
  \code{\link{as.section}},
  \code{\link{handleFlags,section-method}},
  \code{\link{plot,section-method}},
  \code{\link{read.section}}, \code{\link{section-class}},
  \code{\link{sectionAddStation}},
  \code{\link{sectionGrid}}, \code{\link{sectionSmooth}},
  \code{\link{sectionSort}}, \code{\link{section}},
  \code{\link{summary,section-method}}
}
\author{
Dan Kelley
}
