% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.R
\name{summary,adv-method}
\alias{summary,adv-method}
\title{Summarize an adv Object}
\usage{
\S4method{summary}{adv}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"adv"}, usually, a result of a call to
\code{\link[=read.adv]{read.adv()}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarize data in an \code{adv} object.
}
\examples{
library(oce)
data(adv)
summary(adv)

}
\seealso{
Other things related to adv data: 
\code{\link{[[,adv-method}},
\code{\link{[[<-,adv-method}},
\code{\link{adv}},
\code{\link{adv-class}},
\code{\link{advSontekAdrFileTrim}()},
\code{\link{applyMagneticDeclination,adv-method}},
\code{\link{beamName}()},
\code{\link{beamToXyz}()},
\code{\link{enuToOther}()},
\code{\link{enuToOtherAdv}()},
\code{\link{plot,adv-method}},
\code{\link{read.adv}()},
\code{\link{read.adv.nortek}()},
\code{\link{read.adv.sontek.adr}()},
\code{\link{read.adv.sontek.serial}()},
\code{\link{read.adv.sontek.text}()},
\code{\link{rotateAboutZ}()},
\code{\link{subset,adv-method}},
\code{\link{toEnu}()},
\code{\link{toEnuAdv}()},
\code{\link{velocityStatistics}()},
\code{\link{xyzToEnu}()},
\code{\link{xyzToEnuAdv}()}
}
\author{
Dan Kelley
}
\concept{things related to adv data}
