% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.R
\docType{methods}
\name{summary,cm-method}
\alias{summary,cm-method}
\title{Summarize a CM Object}
\usage{
\S4method{summary}{cm}(object, ...)
}
\arguments{
\item{object}{A \code{cm} object, i.e. one inheriting from \code{\link{cm-class}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in a \code{cm} object, presenting such information
as the station name, sampling location, data ranges, etc.
}
\author{
Dan Kelley
}
\seealso{
The documentation for \code{\link{cm-class}} explains the structure
of \code{cm} objects, and also outlines the other functions dealing with them.

Other things related to \code{cm} data: \code{\link{[[,cm-method}},
  \code{\link{[[<-,cm-method}}, \code{\link{cm-class}},
  \code{\link{cm}}, \code{\link{plot,cm-method}},
  \code{\link{read.cm}}, \code{\link{subset,cm-method}}
}

