% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\docType{methods}
\name{summary,coastline-method}
\alias{summary,coastline-method}
\title{Summarize a Coastline Object

Summarizes coastline length, bounding box, etc.}
\usage{
\S4method{summary}{coastline}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"coastline"}, usually, a result of a
call to \code{\link{read.coastline}} or \code{\link{read.oce}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{coastline} data: \code{\link{[[,coastline-method}},
  \code{\link{[[<-,coastline-method}},
  \code{\link{as.coastline}},
  \code{\link{coastline-class}},
  \code{\link{coastlineBest}}, \code{\link{coastlineCut}},
  \code{\link{coastlineWorld}},
  \code{\link{plot,coastline-method}},
  \code{\link{read.coastline.openstreetmap}},
  \code{\link{read.coastline.shapefile}},
  \code{\link{subset,coastline-method}}
}

