% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echosounder.R
\name{summary,echosounder-method}
\alias{summary,echosounder-method}
\title{Summarize an Echosounder Object}
\usage{
\S4method{summary}{echosounder}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"echosounder"}, usually, a result of
a call to \code{\link[=read.echosounder]{read.echosounder()}}, \code{\link[=read.oce]{read.oce()}}, or
\code{\link[=as.echosounder]{as.echosounder()}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in an \linkS4class{echosounder} object.
}
\seealso{
Other things related to echosounder data: 
\code{\link{[[,echosounder-method}},
\code{\link{[[<-,echosounder-method}},
\code{\link{as.echosounder}()},
\code{\link{echosounder-class}},
\code{\link{echosounder}},
\code{\link{findBottom}()},
\code{\link{plot,echosounder-method}},
\code{\link{read.echosounder}()},
\code{\link{subset,echosounder-method}}
}
\author{
Dan Kelley
}
\concept{things related to echosounder data}
