% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladp.R
\docType{methods}
\name{summary,ladp-method}
\alias{summary,ladp-method}
\title{Summarize an ladp object}
\usage{
\S4method{summary}{ladp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"ladp"}, usually, a result of a call to
\code{\link{as.ladp}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A matrix containing statistics of the elements of the \code{data} slot.
}
\description{
Pertinent summary information is presented, including the station name,
sampling location, data ranges, etc.
}
\author{
Dan Kelley
}
\seealso{
Other functions that handle ladp data: \code{\link{[[,ladp-method}},
  \code{\link{as.ladp}}, \code{\link{ladp-class}},
  \code{\link{plot,ladp-method}}
}

