% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\docType{methods}
\name{summary,lisst-method}
\alias{summary,lisst-method}
\title{Summarize a LISST Object}
\usage{
\S4method{summary}{lisst}(object, ...)
}
\arguments{
\item{object}{a \linkS4class{lisst} object.}

\item{...}{Ignored.}
}
\description{
Summarizes some of the data in a \code{lisst} object, presenting such information
as the station name, sampling location, data ranges, etc.
}
\examples{
library(oce)
data(lisst)
summary(lisst)

}
\seealso{
Other things related to lisst data: \code{\link{[[,lisst-method}},
  \code{\link{[[<-,lisst-method}}, \code{\link{as.lisst}},
  \code{\link{lisst-class}},
  \code{\link{plot,lisst-method}}, \code{\link{read.lisst}}
}
\author{
Dan Kelley
}
\concept{things related to lisst data}
