% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobo.R
\docType{methods}
\name{summary,lobo-method}
\alias{summary,lobo,missing-method}
\alias{summary,lobo-method}
\alias{summary.lobo}
\title{Summarize a lobo data object}
\usage{
\S4method{summary}{lobo}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"lobo"}, usually, a result of a call
to \code{\link{read.lobo}} or \code{\link{read.oce}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A matrix containing statistics of the elements of the \code{data}
slot.
}
\description{
Summarizes some of the data in a lobo object.
}
\details{
Pertinent summary information is presented, including the sampling interval,
data ranges, etc.
}
\examples{

library(oce)
data(lobo)
summary(lobo)
}
\author{
Dan Kelley
}
\references{
\url{http://lobo.satlantic.com} \url{http://www.mbari.org/lobo/}
}
\seealso{
The documentation for \code{\link{lobo-class}} explains the
structure of LOBO objects, and also outlines the other functions dealing
with them.

Other functions that handle \code{lobo} data: \code{\link{[[,lobo-method}},
  \code{\link{[[<-,lobo-method}}, \code{\link{as.lobo}},
  \code{\link{plot,lobo-method}},
  \code{\link{subset,lobo-method}}
}
\keyword{misc}

