% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\docType{methods}
\name{summary,met-method}
\alias{summary,met-method}
\title{Summarize a met Object}
\usage{
\S4method{summary}{met}(object, ...)
}
\arguments{
\item{object}{a \linkS4class{met} object.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Pertinent summary information is presented, including the sampling location,
data ranges, etc.
}
\seealso{
Other things related to met data: \code{\link{[[,met-method}},
  \code{\link{[[<-,met-method}}, \code{\link{as.met}},
  \code{\link{download.met}}, \code{\link{met-class}},
  \code{\link{met}}, \code{\link{plot,met-method}},
  \code{\link{read.met}}, \code{\link{subset,met-method}}
}
\author{
Dan Kelley
}
\concept{things related to met data}
