% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{summary,oce-method}
\alias{summary,oce-method}
\title{Summarize an oce object}
\usage{
\S4method{summary}{oce}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Extra arguments (ignored)}
}
\description{
Provide a summary of some pertinent aspects of the object, including
important elements of its \code{metadata} slot and \code{data} slot,
and a listing of the contents of its \code{processingLog} slot.
}
\details{
The output is tailored to the object. This is not the way to learn
everything about the object; programmers, for example, will often
use \code{\link{str}} to get more details of the object structure.
}

