% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\docType{methods}
\name{summary,odf-method}
\alias{summary,odf-method}
\title{Summarize an ODF Object}
\usage{
\S4method{summary}{odf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"odf"}, usually, a result of a call
to \code{\link{read.odf}} or \code{\link{read.oce}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A matrix containing statistics of the elements of the \code{data}
slot.
}
\description{
Pertinent summary information is presented, including the station name,
sampling location, data ranges, etc.
}
\seealso{
Other things related to \code{odf} data: \code{\link{ODF2oce}},
  \code{\link{ODFNames2oceNames}},
  \code{\link{[[,odf-method}},
  \code{\link{[[<-,odf-method}}, \code{\link{odf-class}},
  \code{\link{plot,odf-method}},
  \code{\link{read.ctd.odf}}, \code{\link{read.odf}},
  \code{\link{subset,odf-method}}
}
\author{
Dan Kelley
}
