% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.R
\name{summary,satellite-method}
\alias{summary,satellite-method}
\title{Summarize a satellite Object}
\usage{
\S4method{summary}{satellite}(object, ...)
}
\arguments{
\item{object}{a \linkS4class{satellite} object.}

\item{...}{Ignored.}
}
\description{
Summarize a satellite Object
}
\author{
Dan Kelley
}
