% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel.R
\name{summary,sealevel-method}
\alias{summary,sealevel-method}
\title{Summarize a sealevel Object}
\usage{
\S4method{summary}{sealevel}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{sealevel} object.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A matrix containing statistics of the elements of the \code{data}
slot.
}
\description{
Summarizes some of the data in a sealevel object.
}
\examples{
library(oce)
data(sealevel)
summary(sealevel)

}
\seealso{
Other things related to sealevel data: 
\code{\link{[[,sealevel-method}},
\code{\link{[[<-,sealevel-method}},
\code{\link{as.sealevel}()},
\code{\link{plot,sealevel-method}},
\code{\link{read.sealevel}()},
\code{\link{read.sealevel.gc2026}()},
\code{\link{sealevel}},
\code{\link{sealevel-class}},
\code{\link{sealevelTuktoyaktuk}},
\code{\link{subset,sealevel-method}}
}
\author{
Dan Kelley
}
\concept{things related to sealevel data}
