% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\name{summary,windrose-method}
\alias{summary,windrose-method}
\title{Summarize a windrose Object}
\usage{
\S4method{summary}{windrose}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{windrose} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in a \linkS4class{windrose} object.
}
\seealso{
Other things related to windrose data: 
\code{\link{[[,windrose-method}},
\code{\link{[[<-,windrose-method}},
\code{\link{as.windrose}()},
\code{\link{plot,windrose-method}},
\code{\link{windrose-class}}
}
\author{
Dan Kelley
}
\concept{things related to windrose data}
