% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbt.R
\docType{methods}
\name{summary,xbt-method}
\alias{summary,xbt-method}
\title{Summarize an xbt Object}
\usage{
\S4method{summary}{xbt}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{xbt} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in a \code{xbt} object.
}
\seealso{
The documentation for the \linkS4class{xbt} class explains the structure
of \code{xbt} objects, and also outlines the other functions dealing with them.

Other things related to xbt data: \code{\link{[[<-,xbt-method}},
  \code{\link{as.xbt}}, \code{\link{plot,xbt-method}},
  \code{\link{read.xbt.noaa1}}, \code{\link{read.xbt}},
  \code{\link{subset,xbt-method}}, \code{\link{xbt-class}},
  \code{\link{xbt.edf}}
}
\author{
Dan Kelley
}
\concept{things related to xbt data}
