\name{summary.gps}

\alias{summary.gps}
\alias{summary,gps,missing-method}
\alias{summary,gps-method}

\title{Summarize a gps object}

\description{Summarize a \code{gps} object}

\usage{\S4method{summary}{gps}(object, \dots)}

\arguments{
  \item{object}{an object of class \code{"gps"}}
  \item{\dots}{further arguments passed to or from other methods.}
}

\details{Pertinent summary information is presented.}

\value{A matrix containing statistics of the elements of the \code{data} slot.}

\seealso{The documentation for \code{\link{gps-class}} explains the structure
    of GPS objects, and also outlines the other functions dealing with them.}

\author{Dan Kelley}

\keyword{misc}
