% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.R
\name{sunDeclinationRightAscension}
\alias{sunDeclinationRightAscension}
\title{Sun Declination and Right Ascension}
\usage{
sunDeclinationRightAscension(time, apparent = FALSE)
}
\arguments{
\item{time}{a POSIXct time. This ought to be in UTC timezone; if not,
the behaviour of this function is unlikely to be correct.}

\item{apparent}{logical value indicating whether to return
the 'apparent' angles.}
}
\value{
A list containing \code{declination} and \code{rightAscension}, in degrees.
}
\description{
The formulae are from Meeus (1991), chapter 24 (which uses chapter 21).
}
\examples{
## Example 24.a in Meeus (1991) (page 158 PDF, 153 print)
time <- as.POSIXct("1992-10-13 00:00:00", tz="UTC")
a <- sunDeclinationRightAscension(time, apparent=TRUE)
expect_equal(a$declination, -7.78507,
             tol=            0.00004, scale=1)
expect_equal(a$rightAscension, -161.61919,
             tol=                 0.00003, scale=1)
b <- sunDeclinationRightAscension(time)
## check against previous results, to protect aginst code-drift errors
expect_equal(b$declination, -7.785464443,
             tol=            0.000000001, scale=1)
expect_equal(b$rightAscension, -161.6183305,
             tol=                 0.0000001, scale=1)

}
\references{
\enumerate{
\item Meeus, Jean H. Astronomical Algorithms. Second edition. Willmann-Bell, Incorporated, 1991.
}
}
\seealso{
Other things related to astronomy: 
\code{\link{eclipticalToEquatorial}()},
\code{\link{equatorialToLocalHorizontal}()},
\code{\link{julianCenturyAnomaly}()},
\code{\link{julianDay}()},
\code{\link{moonAngle}()},
\code{\link{siderealTime}()},
\code{\link{sunAngle}()}
}
\author{
Dan Kelley, based on formulae in Meeus (1991).
}
\concept{things related to astronomy}
