% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{tail.oce}
\alias{tail.oce}
\title{Extract the End of an Oce Object}
\usage{
\method{tail}{oce}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{oce} object.}

\item{n}{Number of elements to extract.}

\item{...}{ignored}
}
\description{
NOTE: this is a preliminary function, subject to change. In particular,
it is written as an S3 generic, but problably should be converted
to a S4 generic. So far, this function only handles objects of
\code{\link{adp-class}},
\code{\link{adv-class}},
and
\code{\link{ctd-class}}.
For all other classes, it simply calls \code{\link{tail}}
with \code{n} as provided, for each item in the \code{data}
slot.
}
\seealso{
\code{\link{head.oce}}, which yields the start of an \code{oce} object.
}
\author{
Dan Kelley
}
