% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{test_met_csv2.csv}
\alias{test_met_csv2.csv}
\title{Sample met dataset (CSV format as of October 2019)}
\description{
This is a subset of a file downloaded with \code{\link[=download.met]{download.met()}}
from the Environment Canada "Climate Data"
website \verb{https://climate.weather.gc.ca/index_e.html}
on October 13, 2019.
}
\examples{
\dontrun{
read.met(system.file("extdata", "test_met_csv2.csv", package="oce"))
}

}
\seealso{
Other raw datasets: 
\code{\link{CTD_BCD2014666_008_1_DN.ODF.gz}},
\code{\link{adp_rdi.000}},
\code{\link{ctd.cnv}},
\code{\link{d200321-001.ctd}},
\code{\link{d201211_0011.cnv}},
\code{\link{test_met_csv1.csv}},
\code{\link{test_met_xml2.xml}},
\code{\link{xbt.edf}}

Other things related to met data: 
\code{\link{[[,met-method}},
\code{\link{[[<-,met-method}},
\code{\link{as.met}()},
\code{\link{download.met}()},
\code{\link{met-class}},
\code{\link{met}},
\code{\link{plot,met-method}},
\code{\link{read.met}()},
\code{\link{subset,met-method}},
\code{\link{summary,met-method}},
\code{\link{test_met_csv1.csv}},
\code{\link{test_met_xml2.xml}}
}
\concept{raw datasets}
\concept{things related to met data}
