\name{tideconst}

\docType{data}

\alias{tideconst}

\title{Tidal constituent information}

\description{Tidal constituent information, for use in fitting tidal
  models to sealevel data.}

\usage{data(tidesetup)}

\format{See section 1.3 of Foreman (1977).}

\source{The data come from the \code{tide3.dat} file of the
  \code{T_TIDE} package, and derive from Appendix 7.1 of Foreman
  (1977).  The data are scanned using \file{tests/tide.R} in this
  package, which also tests the veracity of the data, to a limited
  extent.
}

\references{
  Foreman, M. G. G., 1977.
  Manual for tidal heights analysis and prediction.
  Pacific Marine Science Report 77-10,
  Institute of Ocean Sciences, Patricia Bay, Sidney, BC, 58pp.
  
  Pawlowicz, Rich, Bob Beardsley, and Steve Lentz, 2002.
  Classical tidal harmonic analysis including error estimates in MATLAB using \code{T\_TIDE}.
  Computers and Geosciences, 28, 929-937.
}

\author{Dan Kelley}

\keyword{misc}
