% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidem.R
\name{tidemVuf}
\alias{tidemVuf}
\title{Ephemeris Calculations for Tidem}
\usage{
tidemVuf(t, j, lat = NULL)
}
\arguments{
\item{t}{time in \code{POSIXct} format.  (It is \strong{very} important to
use \code{tz="GMT"} in constructing \code{t}.)}

\item{j}{indices of tidal constituents to use}

\item{lat}{latitude (if missing, something is done for that)}
}
\value{
A \code{\link[base]{data.frame}} containing
\item{list("astro")}{(fill in later)} \item{list("ader")}{(fill in later)}
}
\description{
Do ephemeris calculations for tidem. This is based directly
on \code{t_vuf}, from the \code{T_TIDE} package.
}
\examples{

tidemVuf(as.POSIXct("2008-01-22 18:50:24"), 43, 45.0)

}
\author{
Dan Kelley, based directly on \code{t_vuf} from the \code{T_TIDE}
package.
}
\seealso{
\code{\link{tidem}}, which uses this.

Other things related to \code{tidem} data: \code{\link{[[,tidem-method}},
  \code{\link{[[<-,tidem-method}},
  \code{\link{plot,tidem-method}},
  \code{\link{predict.tidem}},
  \code{\link{summary,tidem-method}},
  \code{\link{tidedata}}, \code{\link{tidem-class}},
  \code{\link{tidemAstron}}, \code{\link{tidem}}
}

