% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{timeToArgoJuld}
\alias{timeToArgoJuld}
\title{Convert time to Argo Julian Day (juld)}
\usage{
timeToArgoJuld(t)
}
\arguments{
\item{t}{A POSIXct time or a string that can be converted to a POSIXct time}
}
\description{
This converts a POSIXct time into an Argo julian day, with the convention
that juld=0 at 1950-01-01.
}
\examples{
timeToArgoJuld("2020-07-01")

}
\author{
Jaimie Harbin and Dan Kelley
}
