\name{topo-class}

\docType{class}

\alias{topo-class}
\alias{[[,topo-method}

\title{Class to store topographic data}

\description{Class to store topographic data, with standard slots
    \code{metadata}, \code{data} and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{topo[["latitude"]]}, where
    the string could also be \code{"longitude"} or \code{"z"}, or
    abbreviations matching uniquely to any of these.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

