% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\docType{data}
\name{topoNS}
\alias{topoNS}
\title{High-resolution Topographic Dataset for Nova Scotia}
\source{
This is created with \code{\link[=read.topo]{read.topo()}}, using a file downloaded with
\link{download.topo}\verb{(-67,-59.5,43.3,47.2,resolution=1)}.
}
\usage{
data(topoNS)
}
\description{
One-minute (under 2km) dataset in a region including Nova Scotia, Prince
Edward Island and part of New Brunswick, Canada.
}
\examples{
\donttest{
library(oce)
data(coastlineWorldFine, package="ocedata")
data(topoNS)
# 1. Contour plot (with no legend)
plot(topoNS, location="none")
# 2. Image plot (note that whitespace can occur if plot does not match aspect ratio)
H <- quantile(abs(topoNS[["z"]]), 0.98)
cm <- colormap(H*c(-1, 1),
               col=function(n) oceColorsGebco(region="both",n=n))
imagep(topoNS, colormap=cm, xlab="", ylab="")
lines(coastlineWorldFine[["longitude"]], coastlineWorldFine[["latitude"]])
}

}
\seealso{
Other datasets provided with oce: 
\code{\link{adp}},
\code{\link{adv}},
\code{\link{amsr}},
\code{\link{argo}},
\code{\link{cm}},
\code{\link{coastlineWorld}},
\code{\link{ctdRaw}},
\code{\link{ctd}},
\code{\link{echosounder}},
\code{\link{landsat}},
\code{\link{lisst}},
\code{\link{lobo}},
\code{\link{met}},
\code{\link{ocecolors}},
\code{\link{rsk}},
\code{\link{sealevelTuktoyaktuk}},
\code{\link{sealevel}},
\code{\link{section}},
\code{\link{topoWorld}},
\code{\link{wind}},
\code{\link{xbt}}

Other things related to topo data: 
\code{\link{[[,topo-method}},
\code{\link{[[<-,topo-method}},
\code{\link{as.topo}()},
\code{\link{download.topo}()},
\code{\link{plot,topo-method}},
\code{\link{read.topo}()},
\code{\link{subset,topo-method}},
\code{\link{summary,topo-method}},
\code{\link{topo-class}},
\code{\link{topoInterpolate}()},
\code{\link{topoWorld}}
}
\concept{datasets provided with oce}
\concept{things related to topo data}
