% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trap}
\alias{trap}
\title{trapezoidal integration}
\usage{
trap(x, y, type)
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}

\item{type}{number indicating type, 0, 1 or 2}
}
\description{
This is a utility function called by \code{\link{integrateTrapezoid}},
not intended for direct use.
}
