% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trimString}
\alias{trimString}
\title{Remove leading and trailing whitespace from strings (deprecated)}
\usage{
trimString(s)
}
\arguments{
\item{s}{vector of character strings}
}
\value{
a new vector formed by trimming leading and trailing whitespace
from the elements of \code{s}.
}
\description{
This function will be removed from an upcoming version of oce, because
the base function \code{\link[=trimws]{trimws()}} does the same and more, without having
problems with encoding (see \verb{https://github.com/dankelley/oce/issues/1993}).
}
\concept{deprecated functions}
