% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unduplicateName}
\alias{unduplicateName}
\title{Rename a duplicated item (used in reading CTD files)}
\usage{
unduplicateName(name, existingNames)
}
\arguments{
\item{name}{String with a candidate name.}

\item{existingNames}{Vector of strings with names already processed.}
}
\value{
names String with an unduplicated name.
}
\description{
Determine a new name for an item that is already in a list of names. This is
done by e.g. appending a \code{2} to the second occurrence of a name, etc.
The purpose is to create distinct variable names for
\code{\link{read.ctd.sbe}}.
}
\examples{
unduplicateName("a", c("a", "b", "a")) # returns "a3"
}
\seealso{
\code{\link{unduplicateNames}} is similar, but considers
a vector of names.
}

