% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unitBrackets}
\alias{unitBrackets}
\title{Bracketing strings to put to the left of a unit}
\usage{
unitBrackets()
}
\value{
A list containing \code{L} and \code{R}, suitable for use with \code{\link[=expression]{expression()}}.
}
\description{
This returns strings built up of either parentheses or square brackets,
depending on the value of \code{\link{getOption}}\code{("oceUnitBracket")}, which is
\code{"["} by default.  Space may be added as appropriate, depending on the
value of \code{\link{getOption}}\code{("oceUnitSep")}, which is an empty string by
default.  This function is used by \code{\link[=resizableLabel]{resizableLabel()}}) in the
construction of labels for axis names.
}
\examples{
day <- seq(0, 1, length.out=100)
temperature <- 10 + sin(day * 2 * pi / 1)
ylab <- substitute("Temperature "*L*degree*"C"*R, unitBrackets())
plot(day, temperature, ylab=ylab, type="l")
}
\author{
Dan Kelley
}
