% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{usrLonLat}
\alias{usrLonLat}
\title{Calculate lon-lat coordinates of plot-box trace}
\usage{
usrLonLat(n = 25, debug = getOption("oceDebug"))
}
\arguments{
\item{n}{number of points to check along each side of the plot box}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
a list containing \code{lonmin}, \code{lonmax},
\code{latmin}, \code{latmax}, and \code{ok}; the last
of which indicates whether at least one point on the plot box
is invertable.
}
\description{
Trace along the plot box, converting from xy coordinates to lonlat
coordinates. The results are used by \code{\link{mapGrid}}
and \code{\link{mapAxis}} to ignore out-of-frame grid
lines and axis labels.
}
\seealso{
Other functions related to maps: \code{\link{lonlat2map}},
  \code{\link{lonlat2utm}}, \code{\link{map2lonlat}},
  \code{\link{mapArrows}}, \code{\link{mapAxis}},
  \code{\link{mapContour}},
  \code{\link{mapDirectionField}}, \code{\link{mapGrid}},
  \code{\link{mapImage}}, \code{\link{mapLines}},
  \code{\link{mapLocator}},
  \code{\link{mapLongitudeLatitudeXY}},
  \code{\link{mapPlot}}, \code{\link{mapPoints}},
  \code{\link{mapPolygon}}, \code{\link{mapScalebar}},
  \code{\link{mapText}}, \code{\link{mapTissot}},
  \code{\link{oceCRS}}, \code{\link{shiftLongitude}},
  \code{\link{utm2lonlat}}
}
\author{
Dan Kelley
}
