\name{windrose-class}

\docType{class}

\alias{windrose-class}
\alias{[[,windrose-method}

\title{Class to store windrose data}

\description{Class to store windrose data, with standard slots \code{metadata},
    \code{data} and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{windrose[["theta"]]},
    where the string could also be \code{"count"}, or \code{"fives"}, or
    abbreviations matching uniquely to any of these.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

