% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{woceNames2oceNames}
\alias{woceNames2oceNames}
\title{Translate WOCE data names to \code{oce} data names}
\usage{
woceNames2oceNames(names)
}
\arguments{
\item{names}{vector of strings holding WOCE-style names.}
}
\value{
vector of strings holding \code{oce}-style names.
}
\description{
Translate WOCE-style names to \code{oce} names, using \code{\link{gsub}}
to match patterns. For example, the pattern \code{"CTDOXY.*"} is taken
to mean \code{oxygen}.
}
\author{
Dan Kelley
}

