% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_odf}
\alias{write_odf}
\title{write_odf}
\usage{
write_odf(odf_object, output_file)
}
\arguments{
\item{odf_object}{Object to be written to file}

\item{output_file}{location and name of the file the ODF object is to be written to}
}
\description{
Description:
takes an ODF object and an output location and writes an ODF compliant output file
}
\details{
date: July 4, 2014
}
\author{
Patrick Upson
}

