% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{xbt2.edf}
\alias{xbt2.edf}
\title{Sample xbt File in tab-separated .edf Format}
\description{
Sample xbt File in tab-separated .edf Format
}
\examples{
xbt2 <- read.xbt(system.file("extdata", "xbt2.edf", package = "oce"),
    type = "sippican2"
)
summary(xbt2)

}
\seealso{
Other raw datasets: 
\code{\link{CTD_BCD2014666_008_1_DN.ODF.gz}},
\code{\link{D4902337_219.nc}},
\code{\link{adp_rdi.000}},
\code{\link{ctd.cnv.gz}},
\code{\link{ctd_aml_type1.csv.gz}},
\code{\link{ctd_aml_type3.csv.gz}},
\code{\link{d200321-001.ctd.gz}},
\code{\link{d201211_0011.cnv.gz}},
\code{\link{xbt.edf}}

Other things related to xbt data: 
\code{\link{[[,xbt-method}},
\code{\link{[[<-,xbt-method}},
\code{\link{as.xbt}()},
\code{\link{plot,xbt-method}},
\code{\link{read.xbt}()},
\code{\link{read.xbt.noaa1}()},
\code{\link{read.xbt.noaa2}()},
\code{\link{subset,xbt-method}},
\code{\link{summary,xbt-method}},
\code{\link{xbt}},
\code{\link{xbt-class}},
\code{\link{xbt.edf}}
}
\concept{raw datasets}
\concept{things related to xbt data}
