\name{a03}
\docType{data}
\alias{a03}
\title{WOCE section a03}
\description{
This is line A03 (ExpoCode 90CT40\_1, with nominal sampling date
1993-09-11).  The chief scientist was Tereschenkov of SOI, working
aboard the Russian ship MULTANOVSKY.
}
\usage{data(a03)}
\source{\url{http://cchdo.ucsd.edu/data_access?ExpoCode=90CT40_1} (use the \emph{BOT} link in the \emph{Exchange} section of this page.)}
\references{\url{http://cchdo.ucsd.edu}}
\seealso{This was read with \code{\link{read.section}(a03_hy1.csv)}, working a data file downloaded from the CCHDO site mentioned above.}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
