% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagep.R
\name{abbreviateTimeLabels}
\alias{abbreviateTimeLabels}
\title{Abbreviate a vector of times by removing commonalities}
\usage{
abbreviateTimeLabels(t, ...)
}
\arguments{
\item{t}{vector of times.}

\item{\dots}{optional arguments passed to the \code{\link{format}}, e.g.
\code{format}.}
}
\value{
None.
}
\description{
Abbreviate a vector of times by removing commonalities (e.g. year)
}
\author{
Dan Kelley, with help from Clark Richards
}
\seealso{
This is used by various functions that draw time labels on axes,
e.g.  \code{\link{plot,adp-method}}.
}

