% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.nortek.ad2cp.R
\name{ad2cpCodeToName}
\alias{ad2cpCodeToName}
\title{Map AD2CP ID Code to oce Name}
\usage{
ad2cpCodeToName(code = NULL)
}
\arguments{
\item{code}{a \link{raw} (or corresponding integer) vector indicating the IDs of
interest, or NULL to get a summary of possible values.}
}
\value{
An indication of the mapping.  If \code{code} is NULL, this is a data
frame.  Otherwise, it is a character vector with the relevant mappings,
with the raw form of the code linked with the name, as in the example.
}
\description{
As explained in Nortek (2022, section 6.1, page 80), AD2CP files use a
hexadecimal (in R, "raw") code to indicate the nature of each data chunk, and
\code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}} uses the present function as it analyses AD2CP files.
}
\details{
The mapping from code (hex or decimal) to oce name is as follows.\tabular{rrr}{
   code (raw) \tab code (integer) \tab oce name \cr
   ---------- \tab -------------- \tab ----------------- \cr
   0x15 \tab 21 \tab burst \cr
   0x16 \tab 22 \tab average \cr
   0x17 \tab 23 \tab bottomTrack \cr
   0x18 \tab 24 \tab interleavedBurst \cr
   0x1a \tab 26 \tab burstAltimeterRaw \cr
   0x1b \tab 27 \tab DVLBottomTrack \cr
   0x1c \tab 28 \tab echosounder \cr
   0x1d \tab 29 \tab DVLWaterTrack \cr
   0x1e \tab 30 \tab altimeter \cr
   0x1f \tab 31 \tab averageAltimeter \cr
   0xa0 \tab 160 \tab text \cr
}
}
\examples{
stopifnot(ad2cpCodeToName(0x15) == "0x15=burst")

}
\references{
Nortek AS. \dQuote{Signature Integration 55|250|500|1000kHz.} Nortek AS,
March 31, 2022.
}
\author{
Dan Kelley
}
