% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addColumn}
\alias{addColumn}
\title{Add a Column to the Data Slot of an Oce object [DEPRECATED]}
\usage{
addColumn(x, data, name)
}
\arguments{
\item{x}{A \code{ctd} object, e.g. as read by \code{\link{read.ctd}}.}

\item{data}{the data.  The length of this item must match that of the
existing data entries in the \code{data} slot).}

\item{name}{the name of the column.}
}
\value{
An object of \code{\link[base]{class}} \code{oce}, with a new
column.
}
\description{
\strong{Warning:} this function will be removed soon;
see \link{oce-deprecated}. The more powerful function
\code{\link{oceSetData}} should be used instead.
}
\details{
If there is already a column with the given name, its contents are replaced
by the new value.
}
\author{
Dan Kelley
}
\seealso{
Please use \code{\link{oceSetData}} instead of the present function.
}

