% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addColumn}
\alias{addColumn}
\title{Add a Column to the data Slot of an oce object (defunct)}
\usage{
addColumn(x, data, name)
}
\arguments{
\item{x}{a \linkS4class{ctd} object.}

\item{data}{the data.  The length of this item must match that of the
existing data entries in the \code{data} slot).}

\item{name}{the name of the column.}
}
\value{
A \linkS4class{oce} object.
}
\description{
\strong{WARNING:} This function will be removed soon; see \link{oce-defunct}.
}
\details{
Use \code{\link[=oceSetData]{oceSetData()}} instead of the present function.
}
\seealso{
Please use \code{\link[=oceSetData]{oceSetData()}} instead of the present function.

Other functions that will be removed soon: \code{\link{ctdAddColumn}},
  \code{\link{ctdUpdateHeader}},
  \code{\link{findInOrdered}}, \code{\link{mapMeridians}},
  \code{\link{mapZones}}
}
\author{
Dan Kelley
}
\concept{functions that will be removed soon}
