% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\docType{data}
\name{adp}
\alias{adp}
\title{ADP (acoustic-doppler profiler) dataset}
\source{
This file came from the SLEIWEX-2008 experiment.
}
\usage{
data(adp)
}
\description{
This is degraded subsample of measurements that were made with an
upward-pointing ADP manufactured by Teledyne-RDI, as part of the St Lawrence
Internal Wave Experiment (SLEIWEX).
}
\examples{
\dontrun{
library(oce)
data(adp)

# Velocity components.  (Note: we should probably trim some bins at top.)
plot(adp)

# Note that tides have moved the mooring.
plot(adp, which=15:18)
}


}
\seealso{
Other datasets provided with \code{oce}: \code{\link{adv}},
  \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{lisst}},
  \code{\link{lobo}}, \code{\link{section}},
  \code{\link{topoWorld}}
}

