% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\docType{data}
\name{adp}
\alias{adp}
\title{ADP (acoustic-doppler profiler) dataset}
\source{
This file came from the SLEIWEX-2008 experiment.
}
\usage{
data(adp)
}
\description{
This is degraded subsample of measurements that were made with an
upward-pointing ADP manufactured by Teledyne-RDI, as part of the St Lawrence
Internal Wave Experiment (SLEIWEX).
}
\examples{
\dontrun{
library(oce)
data(adp)

# Velocity components.  (Note: we should probably trim some bins at top.)
plot(adp)

# Note that tides have moved the mooring.
plot(adp, which=15:18)
}


}
\seealso{
Other datasets provided with \code{oce}: \code{\link{adv}},
  \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{echosounder}},
  \code{\link{lisst}}, \code{\link{lobo}},
  \code{\link{met}}, \code{\link{sealevel}},
  \code{\link{section}}, \code{\link{topoWorld}}

Other things related to \code{adp} data: \code{\link{[[,adp-method}},
  \code{\link{[[<-,adp-method}}, \code{\link{adp-class}},
  \code{\link{as.adp}}, \code{\link{plot,adp-method}},
  \code{\link{read.adp.nortek}},
  \code{\link{read.adp.rdi}},
  \code{\link{read.adp.sontek.serial}},
  \code{\link{read.adp.sontek}}, \code{\link{read.adp}},
  \code{\link{read.aquadoppHR}},
  \code{\link{read.aquadoppProfiler}},
  \code{\link{read.aquadopp}},
  \code{\link{subset,adp-method}},
  \code{\link{summary,adp-method}},
  \code{\link{xyzToEnuAdp}}
}

