% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amsr.R
\docType{class}
\name{amsr-class}
\alias{amsr-class}
\title{Class to hold AMSR data}
\description{
The bands are stored in \code{\link{raw}} form, to save storage. The accessor
function, \code{\link{plot.amsr}}, and \code{\link{summary.amsr}} all
apply linear scaling to create values appropriate to the units of
the individual band.  The band names are \code{timeDay},
\code{SSTDay}, \code{LFwindDay}, \code{MFwindDay}, \code{vaporDay} (note
the U.S. spelling), \code{cloudDay}, and \code{rainDay}, along with
similarly-named items that end in \code{Night}.
}
\author{
Dan Kelley
}
\references{
\url{http://www.remss.com/missions/amsre} describes the data.
}

