% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\docType{data}
\name{argo}
\alias{argo}
\title{ARGO drifter dataset}
\source{
This is the profile stored in the file \code{6900388_prof.nc}
downloaded from the \code{usgodae.org} website in March 2012.
}
\description{
This is an ARGO drifter data object, for drifter 6900388, downloaded as
\code{6900388_prof.nc} from \code{usgodae.org} in March 2012.
}
\examples{
\dontrun{
library(oce)
data(argo)
summary(argo)
data(coastlineWorld)
plot(argo, which="trajectory", coastline=coastlineWorld)
}


}
\seealso{
Other datasets provided with oce: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{section}}
}

