% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{argoDataNames}
\alias{argoDataNames}
\title{Convert Data Names From the Argo Convention to the Oce Convention}
\usage{
argoDataNames(names)
}
\arguments{
\item{names}{vector of character strings containing names in the Argo convention.}
}
\description{
For example, \code{"PSAL"} becomes \code{"salinity"}
}

