% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\name{argoJuldToTime}
\alias{argoJuldToTime}
\title{Convert Argo Julian Day (juld) to time}
\usage{
argoJuldToTime(jday)
}
\arguments{
\item{jday}{A numerical value indicating the julian day in the Argo convention,
with day=0 at 1950-01-01.}
}
\description{
Convert Argo Julian Day (juld) to time
}
\examples{
argoJuldToTime(25749)

}
\author{
Jaimie Harbin and Dan Kelley
}
