% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobo.R
\name{as.lobo}
\alias{as.lobo}
\title{Coerce Data into a Lobo Object}
\usage{
as.lobo(time, u, v, salinity, temperature, pressure, nitrate, fluorescence,
  filename = "")
}
\arguments{
\item{time}{vector of times of observation}

\item{u}{vector of x velocity component observations}

\item{v}{vector of y velocity component observations}

\item{salinity}{vector of salinity observations}

\item{temperature}{vector of temperature observations}

\item{pressure}{vector of pressure observations}

\item{nitrate}{vector of nitrate observations}

\item{fluorescence}{vector of fluoresence observations}

\item{filename}{source filename}
}
\value{
An object of \code{\link{lobo-class}}.
}
\description{
Coerce a dataset into a lobo dataset.
}
\seealso{
Other things related to code{lobo} data: \code{\link{[[,lobo-method}},
  \code{\link{[[<-,lobo-method}}, \code{\link{lobo-class}},
  \code{\link{lobo}}, \code{\link{read.lobo}},
  \code{\link{subset,lobo-method}},
  \code{\link{summary,lobo-method}}
}
\author{
Dan Kelley
}
\concept{things related to code{lobo} data}
