% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{as.topo}
\alias{as.topo}
\title{Coerce Data into Topo Object}
\usage{
as.topo(longitude, latitude, z, filename = "")
}
\arguments{
\item{longitude}{Either a vector of longitudes (in degrees east, and bounded by
-180 and 180), or a \code{bathy} object created by \code{getNOAA.bathy()} from
the \code{marmap} package; in the second case, all other arguments are ignored.}

\item{latitude}{A vector of latitudes.}

\item{z}{A matrix of heights (positive over land).}

\item{filename}{Name of data (used when called by \code{\link[=read.topo]{read.topo()}}.}
}
\value{
A \code{\link[=topo-class]{topo-class()}} object.
}
\description{
Coerce Data into Topo Object
}
\seealso{
Other things related to topo data: \code{\link{[[,topo-method}},
  \code{\link{[[<-,topo-method}},
  \code{\link{download.topo}},
  \code{\link{plot,topo-method}}, \code{\link{read.topo}},
  \code{\link{subset,topo-method}},
  \code{\link{summary,topo-method}},
  \code{\link{topo-class}}, \code{\link{topoInterpolate}},
  \code{\link{topoWorld}}
}
\author{
Dan Kelley
}
\concept{things related to topo data}
