% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/units.R
\name{as.unit}
\alias{as.unit}
\title{Convert a String to a Unit}
\usage{
as.unit(u, default = list(unit = expression(), scale = ""))
}
\arguments{
\item{u}{A character string.}

\item{default}{A default, if the string is not recognized.}
}
\value{
A list with elements \code{unit}, an \code{\link{expression}}, and \code{scale}, a string.
}
\description{
Convert a String to a Unit
}
\details{
String matching is used to try to guess a unit from a character string.
}
\examples{
as.unit("IPTS-69")
}

