% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\name{beamName}
\alias{beamName}
\title{Get names of Acoustic-Doppler Beams}
\usage{
beamName(x, which)
}
\arguments{
\item{x}{an \code{adp} or \code{adv} object, i.e. one inheriting from
\code{\link{adp-class}} or \code{\link{adv-class}}.}

\item{which}{an integer indicating beam number.}
}
\value{
A character string containing a reasonable name for the beam, of the
form \code{"beam 1"}, etc., for beam coordinates, \code{"east"}, etc. for
enu coordinates, \code{"u"}, etc. for \code{"xyz"}, or \code{"u'"}, etc.,
for \code{"other"} coordinates.  The coordinate system is determined
with \code{x[["coordinate"]]}.
}
\description{
Get names of Acoustic-Doppler Beams
}
\author{
Dan Kelley
}
\seealso{
This is used by \code{\link{read.oce}}.
}

