% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{byteToBinary}
\alias{byteToBinary}
\title{Format bytes as binary [deprecated]}
\usage{
byteToBinary(x, endian = c("little", "big"))
}
\arguments{
\item{x}{an integer to be interpreted as a byte.}

\item{endian}{character string indicating the endian-ness ("big" or
"little").}
}
\value{
A character string representing the bit strings for the elements of
\code{x}, in order of significance for the \code{endian="big"} case.
(The \code{"little"} case should not be used.)
See \dQuote{Examples} for how this relates to the output from
\link{rawToBits}.
}
\description{
\strong{WARNING:} This function will be removed soon; see \link{oce-deprecated}.
Use \code{\link{oceSetData}} instead of the present function.
}
\examples{
library(oce)
## In the big-endian case, the results match those of byteToBinary(),
## but reversed.
a <- as.raw(0x0a)
byteToBinary(a, "big")
rev(rawToBits(a))
}
\author{
Dan Kelley
}
