% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.R
\docType{class}
\name{cm-class}
\alias{cm-class}
\title{Class to store current meter data}
\description{
Class to store current meter data from an interocean S4 device.  A file
containing CM profile data may be read with \code{\link{read.cm}}. The results
may be plotted with \code{\link{plot,cm-method}} or summarized with
\code{\link{summary,cm-method}}.  Data may be retrieved with
\code{\link{[[,cm-method}} or replaced with \ \code{\link{[[<-,cm-method}}.
}
\author{
Dan Kelley
}
\seealso{
Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}},
  \code{\link{coastline-class}}, \code{\link{ctd-class}},
  \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{met-class}}, \code{\link{odf-class}},
  \code{\link{rsk-class}}, \code{\link{sealevel-class}},
  \code{\link{sealevel}}, \code{\link{section-class}},
  \code{\link{topo-class}}, \code{\link{windrose-class}}

Other things related to \code{cm} data: \code{\link{[[,cm-method}},
  \code{\link{[[<-,cm-method}}, \code{\link{cm}},
  \code{\link{plot,cm-method}}, \code{\link{read.cm}},
  \code{\link{subset,cm-method}},
  \code{\link{summary,cm-method}}
}

