% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.sbe.R
\name{cnvName2oceName}
\alias{cnvName2oceName}
\title{Decode an SBE variable name from a header line}
\usage{
cnvName2oceName(h)
}
\arguments{
\item{h}{The header line.}
}
\value{
a list containing \code{name} (the oce name), \code{nameOriginal} (the SBE name) and \code{unit}.
}
\description{
This function is used by \code{\link{read.ctd.sbe}} to infer data names
and units from the coding used by Teledyne/Seabird (SBE) \code{.cnv}
files.  Lacking access to documentation on the SBE format,
the present function is based on inspection of a suite of CNV files
available to the \code{oce} developers.
}
\details{
A few sample header lines to be decoded are as follows.
\preformatted{
# name 3 = t090C: Temperature [ITS-90, deg C]
# name 4 = t190C: Temperature, 2 [ITS-90, deg C]
}
Examination of several CNV files suggests that it is best to
try to infer the name from the characters between the "\code{=}"
and "\code{:}" characters, because the material after the colon
seems to vary more between sample files.

The table given below indicates the translation patterns used.
The \code{.cnv} convention for multiple sensors is to include digits in the
name, and these are indicated with \code{N} in the leftmost column of
the table below. This maps to \code{M} in the second column in the following
way: \code{M} is omitted if \code{N} is zero, or \code{M} equals \code{N}

\tabular{llll}{
  \strong{Key}       \tab \strong{Result}                     \tab \strong{Unit, scale} \tab \strong{Notes} \cr
  \code{altM}        \tab \code{altimeter}                    \tab m                    \tab                \cr
  \code{cNms/cm}     \tab \code{conductivityM}                \tab ms/cm                \tab                \cr
  \code{CStarAtN}    \tab \code{beamAttenuation}              \tab 1/m                  \tab                \cr
  \code{CStarTrN}    \tab \code{beamTransmission}             \tab percent              \tab                \cr
  \code{depS}        \tab \code{depth}                        \tab m                    \tab                \cr
  \code{depSM}       \tab \code{depth}                        \tab m                    \tab                \cr
  \code{dz/dtM}      \tab \code{descentRate}                  \tab m/s                  \tab                \cr
  \code{flag}        \tab \code{flag}                         \tab -                    \tab                \cr
  \code{flC}         \tab \code{fluorescenceChelsea}          \tab ug/l                 \tab                \cr
  \code{flC1}        \tab \code{fluorescenceChelseaN}         \tab ug/l                 \tab                \cr
  \code{flEC-AFLM}   \tab \code{fluorescenceN}                \tab mg/m^3               \tab                \cr
  \code{flsP}        \tab \code{fluorescence}                 \tab -                    \tab                \cr
  \code{latitude}    \tab \code{latitude}                     \tab degN                 \tab                \cr
  \code{longitude}   \tab \code{longitude}                    \tab degE                 \tab                \cr
  \code{nbin}        \tab \code{nbin}                         \tab -                    \tab                \cr
  \code{oxsatML/L}   \tab \code{oxygenSaturationWeiss}        \tab ml/l                 \tab                \cr
  \code{oxsatMg/L}   \tab \code{oxygenSaturationWeiss}        \tab mg/l                 \tab                \cr
  \code{oxsatMm/Kg}  \tab \code{oxygenSaturationWeiss}        \tab umol/kg              \tab                \cr
  \code{oxsolML/L}   \tab \code{oxygenSaturationGarciaGordon} \tab ml/l                 \tab                \cr
  \code{oxsolMg/L}   \tab \code{oxygenSaturationGarciaGordon} \tab mg/l                 \tab                \cr
  \code{oxsolMm/Kg}  \tab \code{oxygenSaturationGarciaGordon} \tab umol/kg              \tab                \cr
  \code{potempN68C}  \tab \code{thetaM}                       \tab degC, IPTS-68        \tab                \cr
  \code{potempN90C}  \tab \code{thetaM}                       \tab degC, ITS-90         \tab                \cr
  \code{pr}          \tab \code{pressure}                     \tab dbar                 \tab                \cr
  \code{prDM}        \tab \code{pressure}                     \tab dbar                 \tab 1              \cr
  \code{ptempC}      \tab \code{pressureTemperature}          \tab degC, ITS-90         \tab 2              \cr
  \code{pumps}       \tab \code{pumpStatus}                   \tab                      \tab                \cr
  \code{salNN}       \tab \code{salinityM}                    \tab unitless, PSS-78     \tab 3              \cr
  \code{sbeoxNML/L}  \tab \code{oxygenConcentrationVolumeM}   \tab ml/l                 \tab                \cr
  \code{sbeoxNMm/Kg} \tab \code{oxygenConcentrationMoleM}     \tab ml/l                 \tab                \cr
  \code{sbeoxNPs}    \tab \code{oxygenSaturationM}            \tab percent              \tab                \cr
  \code{sbeoxNV}     \tab \code{oxygenVoltageM}               \tab ml/l                 \tab                \cr
  \code{scan}        \tab \code{scan}                         \tab -                    \tab                \cr
  \code{sigma-theta} \tab \code{sigmaTheta}                   \tab kg/m^3               \tab 4              \cr
  \code{spar}        \tab \code{spar}                         \tab -                    \tab                \cr
  \code{svCM}        \tab \code{soundSpeed}                   \tab m/s                  \tab                \cr
  \code{tN68}        \tab \code{temperatureM}                 \tab degC, IPTS-68        \tab                \cr 
  \code{tN68C}       \tab \code{temperatureM}                 \tab degC, IPTS-68        \tab                \cr 
  \code{tN90c}       \tab \code{temperatureM}                 \tab degC, ITS-90         \tab                \cr
  \code{upolyN}      \tab \code{upolyM}                       \tab -                    \tab                \cr 
  \code{vN}          \tab \code{voltageM}                     \tab V                    \tab                \cr 
  \code{wetCDOM}     \tab \code{fluorescence}                 \tab mg/m^3               \tab                \cr 
}
Notes:
\itemize{
\item{1. what should be done if a file has both \code{pr} and \code{prDM}?}
\item{2: assume ITS-90 temperature scale, since sample \code{.cnv} file headers do not specify it.}
\item{3: some files have PSU for this. Should we handle that? And are there other S scales to consider?}
\item{4: 'theta' may appear in different ways with different encoding configurations, set up
within R or in the operating system.}
}
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{ctd-class}}, \code{\link{ctdAddColumn}},
  \code{\link{ctdDecimate}}, \code{\link{ctdFindProfiles}},
  \code{\link{ctdRaw}}, \code{\link{ctdTrim}},
  \code{\link{ctdUpdateHeader}}, \code{\link{ctd}},
  \code{\link{gps-class}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}}, \code{\link{plotProfile}},
  \code{\link{plotScan}}, \code{\link{plotTS}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.odv}}, \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}},
  \code{\link{woceNames2oceNames}}, \code{\link{write.ctd}}
}

