\name{coastline-class}

\docType{class}

\alias{coastline-class}
\alias{[[,coastline-method}

\title{Class to store coastline data}

\description{Class to store coastline data, with standard slots \code{metadata}
    (containing \code{fillable} and \code{filename}),  \code{data} (containing
    \code{latitude} and \code{longitude}) and \code{processingLog}.}

\section{Methods}{Data may be accessed as e.g. \code{coastline[["latitude"]]}
    or \code{coastline[["longitude"]]}.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

