\name{coastline-class}

\docType{class}

\alias{coastline-class}
\alias{[[}
\alias{[[,coastline-method,index,ANY,ANY-method}

\title{Class to store coastline data}

\description{Class to store coastline data, with standard slots \code{metadata}
    (containing \code{fillable} and \code{filename}),  \code{data} (containing
    \code{latitude} and \code{longitude}) and \code{processingLog}.}

\section{Methods}{
    \describe{
        \item{[[}{\code{signature(x="Coastline")}:
            provide access to \code{latitude} and \code{longitude}.}
    }
}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

