% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\docType{class}
\name{coastline-class}
\alias{coastline-class}
\title{Class to store coastline data}
\description{
Class to store coastline data, which may be read with
\code{\link{read.coastline}} or constructed with \code{\link{as.coastline}},
plotted with \code{\link{plot,coastline-method}} or summarized with
\code{\link{summary,coastline-method}}. Data within \code{coastline}
objects may be retrieved with \code{\link{[[,coastline-method}}
or replaced with \code{\link{[[<-,coastline-method}}.
}
\author{
Dan Kelley
}
\seealso{
Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{ctd-class}}, \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{rsk-class}}, \code{\link{section-class}},
  \code{\link{topo-class}}, \code{\link{windrose-class}}
}

