% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\docType{class}
\name{coastline-class}
\alias{coastline-class}
\title{Class to Store Coastline Data}
\description{
Class to store coastline data, which may be read with
\code{\link{read.coastline}} or constructed with \code{\link{as.coastline}},
plotted with \code{\link{plot,coastline-method}} or summarized with
\code{\link{summary,coastline-method}}. Data within \code{coastline}
objects may be retrieved with \code{\link{[[,coastline-method}}
or replaced with \code{\link{[[<-,coastline-method}}.
}
\author{
Dan Kelley
}
\seealso{
Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{ctd-class}}, \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{met-class}}, \code{\link{oce-class}},
  \code{\link{odf-class}}, \code{\link{rsk-class}},
  \code{\link{sealevel-class}},
  \code{\link{section-class}}, \code{\link{topo-class}},
  \code{\link{windrose-class}}

Other things related to \code{coastline} data: \code{\link{[[,coastline-method}},
  \code{\link{[[<-,coastline-method}},
  \code{\link{as.coastline}}, \code{\link{coastlineBest}},
  \code{\link{coastlineCut}}, \code{\link{coastlineWorld}},
  \code{\link{plot,coastline-method}},
  \code{\link{read.coastline.openstreetmap}},
  \code{\link{read.coastline.shapefile}},
  \code{\link{subset,coastline-method}},
  \code{\link{summary,coastline-method}}
}

