\name{coastline.maritimes}
\docType{data}
\alias{coastline.maritimes}
\title{Sample coastline data set, showing Maritime provinces of Canada}
\description{A sample coastline dataset, showing Nova Scotia, 
Prince Edward Island, and part of New Brunswick.
This is the oce author's stomping ground, and so he feels compelled
to apologize for the error at the Strait of Canso.}
\usage{data(coastline.maritimes)}
\source{The data were extracted from the the 1:2,000,000 ``World Data Bank II''
dataset at the NOAA site.}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/index.jsp?llat=43.0&ulat=47.2&llon=-66.8&rlon=-59.5}}
\seealso{Coastlines may be read with
\code{\link{read.coastline}}, or created with
\code{\link{as.coastline}}; see these for details of the data format.  Other coastline datasets provided with
oce are \code{\link{coastline.halifax}} and \code{\link{coastline.world}}.}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
