\name{coastline.world}
\docType{data}
\alias{coastline.world}
\title{World coastline at 1:5000000 resolution}
\description{World coastline at 1:5000000 resolution}
\usage{data(coastline.world)}
\source{http://rimmer.ngdc.noaa.gov/mgg/coast/getcoast.html}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/}}
\seealso{Coastlines may be read with
\code{\link{read.coastline}}, or created with
\code{\link{as.coastline}}; see these for details of the data format.  Other coastline datasets provided with
oce are \code{\link{coastline.halifax}} and \code{\link{coastline.maritimes}}.}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
