\name{coastline.world}
\docType{data}
\alias{coastline.world}
\title{World coastline at coarse resolution}
\description{World coastline at 1:5000000 resolution.}
\usage{data(coastline.world)}
\details{The longitudes are in the range from -180 to 180 degrees, i.e.
western longitudes have negative values.}
\source{\url{http://rimmer.ngdc.noaa.gov/mgg/coast/getcoast.html}}
\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/}}
\seealso{Coastlines may be read with
\code{\link{read.coastline}}, or created with
\code{\link{as.coastline}}; see these for details of the data format.  Other coastline datasets provided with
oce are \code{\link{coastline.halifax}} and \code{\link{coastline.maritimes}}.}
\author{Dan Kelley}
\keyword{misc}
