% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coastline.R
\name{coastlineCut}
\alias{coastlineCut}
\title{Cut a coastline file at specified longitude}
\usage{
coastlineCut(coastline, lon_0 = 0)
}
\arguments{
\item{coastline}{original coastline object}

\item{lon_0}{longitude as would be given in a \code{+lon_0=} item in a proj.4 string}
}
\value{
a new coastline object
}
\description{
Cut a coastline file at specified longitude
}
\details{
This can be helpful in preventing \code{\link{mapPlot}} from producing ugly
horizontal lines in world maps. These lines occur when a coastline segment
is intersected by longitude lon_0+180.  Since the coastline files in the oce
and ocedata packages are already "cut" at longitudes of -180 and 180, the present
function is not needed for default maps, which have \code{+lon_0=0}. However,
may help with other values of \code{lon_0}.

This function is provisional, and both its behaviour and name may change
through the summer and possibly autumn of 2015.
}
\examples{
\dontrun{
library(oce)
data(coastlineWorld)
mapPlot(coastlineCut(coastlineWorld, lon_0=100), proj="+proj=robin +lon_0=100", fill='gray')
}
}

