\name{coastlineHalifax}

\docType{data}

\alias{coastlineHalifax}

\title{Coastline data set, showing Halifax Harbour}

\description{A sample coastline dataset, showing Halifax Harbour.
The little island at about 44.64N is visible outside the author's window.}
\usage{data(coastlineHalifax)}

\references{\url{http://www.ngdc.noaa.gov/mgg_coastline/}}

\seealso{The documentation for \code{\link{coastline-class}} explains the
    structure of coastline objects, and also outlines the other functions
    dealing with them.}

\author{Dan Kelley}

\keyword{datasets}
