\name{coastlineSLE}

\docType{data}

\alias{coastlineSLE}

\title{Coastline of St Lawrence Estuary}

\description{A sample coastline dataset, showing the St Lawrence Estuary.}

\usage{data(coastlineSLE)}

\source{The data were extracted from the the 1:250,000 ``World Vector Shoreline''
dataset at the NOAA site.}

\references{\url{http://www.ngdc.noaa.gov/mgg_coastline}}

\seealso{The documentation for \code{\link{coastline-class}} explains the
    structure of coastline objects, and also outlines the other functions
    dealing with them.}

\author{Dan Kelley}

\keyword{datasets}
