% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colormapGMT}
\alias{colormapGMT}
\title{Create a GMT-type colormap}
\usage{
colormapGMT(name, debug = getOption("oceDebug"))
}
\arguments{
\item{name}{character value specifying the GMT scheme, or a source for such
a scheme. Four pre-defined schemes are available, accessed by setting \code{name} to
\code{"gmt_gebco"}, \code{"gmt_globe"}, \code{"gmt_ocean"}, or \code{"gmt_relief"}. If \code{name} is
not one of these values, then it is taken to be the name of a local file
in GMT format or, if no such file is found, a URL holding such a file.}

\item{debug}{integer that, if positive, indicates to print some debugging output}
}
\value{
a list, in the same format as the return value for \code{\link[=colormap]{colormap()}}.
}
\description{
\code{colormapGMT} creates colormaps in the Generic Mapping Tools (GMT)
scheme (see Reference 1).  A few such schemes are built-in, and may be referred to
by name (\code{"gmt_gebco"}, \code{"gmt_globe"}, \code{"gmt_ocean"}, or \code{"gmt_relief"})
while others are handled by reading local files that are in GMT
format, or URLs providing such files (see Reference 2).
}
\details{
GMT files start with optional comment lines that begin with
the \verb{#} character, followed by a sequence of lines containing 8 numbers,
followed by a line with the character \code{F} followed by three integers
giving red, green and blue values in the range 0 to 255, followed
by a similar line starting with the character \code{B},
and finally by a similar line starting
with the character \code{N}. The last line
is interpreted as the missing-value color.  The lines containing 8 numbers each
describe a color band. The first number is the minimum value in the band,
and this is followed by integers in the range 0:255 that specify the
red, green and blue components of the color.  The fourth to eight numbers
similarly describe the upper end of that color band.
}
\references{
\enumerate{
\item \url{https://docs.generic-mapping-tools.org/dev/cookbook/cpts.html}
\item \url{https://beamreach.org/maps/gmt/share/cpt}
}
}
\seealso{
Other things related to colors: 
\code{\link{colormap}()},
\code{\link{oceColors9B}()},
\code{\link{oceColorsCDOM}()},
\code{\link{oceColorsChlorophyll}()},
\code{\link{oceColorsClosure}()},
\code{\link{oceColorsDensity}()},
\code{\link{oceColorsFreesurface}()},
\code{\link{oceColorsGebco}()},
\code{\link{oceColorsJet}()},
\code{\link{oceColorsOxygen}()},
\code{\link{oceColorsPAR}()},
\code{\link{oceColorsPalette}()},
\code{\link{oceColorsPhase}()},
\code{\link{oceColorsSalinity}()},
\code{\link{oceColorsTemperature}()},
\code{\link{oceColorsTurbidity}()},
\code{\link{oceColorsTurbo}()},
\code{\link{oceColorsTwo}()},
\code{\link{oceColorsVelocity}()},
\code{\link{oceColorsViridis}()},
\code{\link{oceColorsVorticity}()},
\code{\link{ocecolors}}
}
\author{
Dan Kelley
}
\concept{things related to colors}
