% vim:textwidth=80:expandtab:shiftwidth=2:softtabstop=2
\name{colors}

\docType{data}

\alias{colors}

\title{Data that define some color palettes}

\description{The \code{colors} dataset is a list containing vectors of colors,
  e.g. \code{colors$viridis} holds colors for the color palette known as
  Viridis, which in 2015 became the default color palette in the matplotlib 2.0
  Python library [1].}

\usage{data(colors)}

\source{The data come from the code in matplotlib [1].}

\seealso{\code{\link{oceColorsViridis}} uses this dataset.}
    
\references{
  
  1. Matplotlib is developed on github; see
  \url{https://github.com/matplotlib/matplotlib}
  
}

\author{Authored by matplotlib contributes, packaged in oce by Dan Kelley}

\keyword{datasets}

\concept{hplot}

