% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors}
\alias{colors}
\title{Data that define some colour palettes}
\source{
The data come from the code in matplotlib [1].
}
\description{
The \code{colors} dataset is a list containing vectors of colour-scheme
names, e.g. \code{colors$viridis} holds colours for the colour palette known
as Viridis, which in 2015 became the default colour palette in the
matplotlib 2.0 Python library [1].
}
\references{
1. Matplotlib is developed on github; see
\url{https://github.com/matplotlib/matplotlib}
}
\seealso{
\code{\link{oceColorsViridis}} uses this dataset.

Other datasets provided with \code{oce}: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{echosounder}},
  \code{\link{landsat}}, \code{\link{lisst}},
  \code{\link{lobo}}, \code{\link{met}}, \code{\link{rsk}},
  \code{\link{sealevelTuktoyaktuk}},
  \code{\link{sealevel}}, \code{\link{section}},
  \code{\link{topoWorld}}, \code{\link{wind}}

Other things related to colors: \code{\link{colormap}},
  \code{\link{oce.colorsGebco}}, \code{\link{oceColors9A}},
  \code{\link{oceColors9B}}, \code{\link{oceColorsCDOM}},
  \code{\link{oceColorsChlorophyll}},
  \code{\link{oceColorsDensity}},
  \code{\link{oceColorsFreesurface}},
  \code{\link{oceColorsJet}},
  \code{\link{oceColorsOxygen}},
  \code{\link{oceColorsPAR}},
  \code{\link{oceColorsPalette}},
  \code{\link{oceColorsPhase}},
  \code{\link{oceColorsSalinity}},
  \code{\link{oceColorsTemperature}},
  \code{\link{oceColorsTurbidity}},
  \code{\link{oceColorsTwo}},
  \code{\link{oceColorsVelocity}},
  \code{\link{oceColorsViridis}},
  \code{\link{oceColorsVorticity}}
}
\author{
Authored by matplotlib contributes, packaged in oce by Dan Kelley
}
