% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{composite,ANY-method}
\alias{composite,ANY-method}
\title{Combine several oce objects into a composite}
\usage{
\S4method{composite}{ANY}(object, ...)
}
\arguments{
<<<<<<< HEAD
\item{object}{An object inheriting from \link{oce-class}.}

\item{...}{Other oce objects of the same inheritance as \code{object}.}
}
\description{
Combine several oce objects into a composite

Items within the \code{data} slots of the
objects that are supplied as arguments are averaged in a way that
makes sense for the object class, i.e. taking into account
the particular bad-data codes of that particular class.
}
\seealso{
Other functions that create composite objects: \code{\link{composite,amsr-method}},
  \code{\link{composite}}
=======
\item{object}{Any object.}
}
\description{
Combine several oce objects into a composite
>>>>>>> 8fbff8a36b48d85c0e9d20ebed94172b929524d5
}

