% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\docType{methods}
\name{concatenate,adp-method}
\alias{concatenate,adp-method}
\title{Specialize "concatenate" for adp objects.}
\usage{
\S4method{concatenate}{adp}(object, ...)
}
\description{
Specialize "concatenate" for adp objects.
}
