% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{concatenate,list-method}
\alias{concatenate,list-method}
\title{Concatenate a list of oce objects}
\usage{
\S4method{concatenate}{list}(object)
}
\arguments{
\item{object}{A list holding objects of \link{oce-class}.}
}
\value{
An object of class corresponding to that in \code{object}.
}
\description{
Concatenate a list of oce objects
}
\seealso{
Other functions that concatenate oce objects.: \code{\link{concatenate}}
}
\concept{functions that concatenate oce objects.}
