% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{concatenate,list-method}
\alias{concatenate,list-method}
\title{Concatenate a list of oce objects}
\usage{
\S4method{concatenate}{list}(object)
}
\arguments{
\item{object}{a \link{list} of \linkS4class{oce} objects.}
}
\value{
An object of class corresponding to that in \code{object}.
}
\description{
Concatenate a list of oce objects
}
\concept{functions that concatenate \link{oce} objects.}
