% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{concatenate,oce-method}
\alias{concatenate,oce-method}
\title{Concatenate oce objects}
\usage{
\S4method{concatenate}{oce}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{oce-class}}, or a list containing such
objects (in which case the remaining arguments are ignored).}

\item{...}{Optional additional objects of \code{\link{oce-class}}.}
}
\description{
Concatenate oce objects

This function concatenates oce objects.
}
\examples{
## 1. Split, then recombine, a ctd object.
data(ctd)
a <- subset(ctd, scan <= median(ctd[["scan"]]))
b <- subset(ctd, scan > median(ctd[["scan"]]))
ab <- concatenate(a,b)

## 2. Split, then recombine an adp object.
data(adp)
midtime <- median(adp[["time"]])
a <- subset(adp, time <= midtime)
b <- subset(adp, time > midtime)
ab <- concatenate(a, b)

\dontrun{
## 3. Download and combine EnvCan monthly met datasets.
d8 <- read.met(download.met(id=6358, year=2003, month=8, destdir="."))
d9 <- read.met(download.met(id=6358, year=2003, month=9, destdir="."))
dd <- concatenate(d8, d9)
}

}
\seealso{
Other functions that concatenate \code{oce} objects.: \code{\link{concatenate,adp-method}}
}
\author{
Dan Kelley
}
