% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{concatenate,oce-method}
\alias{concatenate,oce-method}
\title{Create composite objects}
\usage{
\S4method{concatenate}{oce}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{oce}}, or a list containing such
objects (in which case the remaining arguments are ignored).}

\item{...}{Optional additional object of class \code{\link{oce}}.}
}
\description{
Create an oce object by stringing together some other objects.
}
\section{Historical note}{

This was added on April 10, 2018, and has only been tested on \code{\link{met-class}}
and \code{\link{adp-class}} to date. It is written in a general way, and therefore
may do something (without generating errors) on other classes, but that does not
imply that the results will be meaningful because some classes require special
customization.  It is hoped that users will let the developers
know (via github issues) of other classes that need support.
}

\examples{
## 1. Combine two adp objects (created by splitting one)
data(adp)
midtime <- median(adp[["time"]])
a <- subset(adp, time <= midtime)
b <- subset(adp, time > midtime)
ab <- concatenate(a, b)
plot(ab)

\dontrun{
## 2. Combine two monthly "met" datasets from Environment Canada.
d8 <- read.met(download.met(id=6358, year=2003, month=8, destdir="."))
d9 <- read.met(download.met(id=6358, year=2003, month=9, destdir="."))
dd <- concatenate(d8, d9)
}

}
