% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate oce Objects (Generic)}
\usage{
concatenate(object, ..., debug = getOption("oceDebug"))
}
\arguments{
\item{object}{an \linkS4class{oce} object.}

\item{...}{optional additional \linkS4class{oce} objects.}

\item{debug}{integer indicating a debugging level. If this is 0,
the work is done silently.  If it is a larger integer, some information
may be printed during the processing.}
}
\value{
An object of class corresponding to that of \code{object}.
}
\description{
Concatenate oce Objects (Generic)
}
\seealso{
Other functions that concatenate oce objects: 
\code{\link{concatenate,adp-method}},
\code{\link{concatenate,list-method}},
\code{\link{concatenate,oce-method}}
}
\concept{functions that concatenate oce objects}
