% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate oce objects}
\usage{
concatenate(object, ...)
}
\arguments{
\item{object}{An object of \linkS4class{oce} class.}

\item{...}{Optional additional objects of \linkS4class{oce} class.}
}
\value{
An object of class corresponding to that of \code{object}.
}
\description{
Concatenate oce objects
}
\concept{functions that concatenate oce objects.}
