% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate oce objects}
\usage{
concatenate(...)
}
\arguments{
\item{...}{one or more oce objects or a list containing such
objects, potentially followed by other items that are ignored.}
}
\value{
An oce object that is formed by concatenating the input objects.
}
\description{
The elements in the \code{data} slot are combined sequentially, as
appropriate (e.g. it makes no sense to combine the \code{distance}
item within an \code{\link{adp-class}} object). Some elements of
the \code{metadata} slot may also be combined as appropriate (e.g.
the \code{flags} are concatenated, if extant). Some alignment
is also required for some object classes (e.g. \code{\link{adp-class}}
objects have \code{numberOfSamples} in the \code{metadata} slot,
and \code{\link{plot,adp-method}} requires that this values
matches up with the first dimension of \code{v} and similar
elements in the \code{data} slot). See \dQuote{Examples} for
illustrations with the object types that have been tested
to date, and \dQuote{History} for notes on development.
}
\section{History}{

As of Apr 9, 2018, this function handles only objects of
\code{\link{met-class}} and \code{\link{adp-class}}.
Other classes may be handled at least partially, but a fair
bit of special-case coding may be required, because the internal
workings of the various classes differ. Users are asked
to report problems via github issues.
}

\examples{

## 1. Combine two adp objects (created by splitting one)
data(adp)
midtime <- median(adp[["time"]])
a <- subset(adp, time <= midtime)
b <- subset(adp, time > midtime)
ab <- concatenate(a, b)
plot(ab)

\dontrun{
## 2. Combine two monthly "met" datasets from Environment Canada.
d8 <- read.met(download.met(id=6358, year=2003, month=8, destdir="."))
d9 <- read.met(download.met(id=6358, year=2003, month=9, destdir="."))
dd <- concatenate(d8, d9)
plot(dd)
}
}
\author{
Dan Kelley
}
