% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate oce objects}
\usage{
concatenate(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{oce-class}}.}

\item{...}{Optional additional objects of \code{\link{oce-class}}.}
}
\value{
An object of class corresponding to that of \code{object}.
}
\description{
Concatenate oce objects
}
\seealso{
Other functions that concatenate \code{oce} objects.: \code{\link{concatenate,adp-method}},
  \code{\link{concatenate,list-method}},
  \code{\link{concatenate,oce-method}}
}
\concept{functions that concatenate \code{oce} objects.}
