% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{class}
\name{ctd-class}
\alias{ctd-class}
\title{Class to Store CTD (or general hydrographic) Data}
\description{
Class to store hydrographic data such as measured with a CTD (conductivity,
temperature, depth) instrument, or with other systems that produce
similar data.  Data repositories may store conductivity, temperature
and depth, as in the instrument name, but it is also common to store
salinity, temperature and pressure instead (or in addition). For this
reason, \code{ctd} objects are required to hold \code{salinity},
\code{temperature} and \code{pressure} in their \code{data} slot,
with other data being optional. Formulae are available for converting
between variants of these data triplets, e.g. \code{\link{swSCTp}}
can calculate \code{salinity} given \code{conductivity}, \code{temperature}
and \code{pressure}, and these are used by the main functions that
create \code{ctd} objects. For example, if \code{\link{read.ctd.sbe}}
is used to read a Seabird file that contains only conductivity, temperature
and pressure, then that function will automatically append a data
item to hold salinity.  \code{\link{as.ctd}} acts similarly. The result
this is that all \code{ctd} objects hold \code{salinity},
\code{temperature} and \code{pressure}, which are henceforth called
the three basic quantities.
}
\details{
Different units and scales are permitted for the three basic quantities, and
most \code{oce} functions check those units and scales before
doing calculations (e.g. of seawater density), because those calculations
demand certain units and scales. The way this is handled is that the
accessor function \code{\link{[[,ctd-method}} returns values in standardized
form. For example, a \code{ctd} object might hold temperature defined on the
IPTS-68 scale, but e.g. \code{ctd[["temperature"]]} returns a value on the ITS-90
scale. (The conversion is done with \code{\link{T90fromT68}}.)  Similarly,
pressure may be stored in either dbars or PSI, but e.g. \code{ctd[["pressure"]]}
returns a value in dbars, after multiplying by 0.689476 if the value is
stored in PSI. Luckily, there is (as of early 2016) only one salinity scale in
common use in data files, namely PSS-78.
}
\section{Slots}{

\describe{
\item{\code{data}}{As with all \code{oce} objects, the \code{data} slot is a
\code{\link{list}} containing the main data for the object.}

\item{\code{metadata}}{As with all \code{oce} objects, the \code{metadata} slot is
a \code{\link{list}} containing information about the \code{data} or about the
object itself.  An example of the former might be the location at which a
measurement was made, and of the latter might be the name of the experiment in
which the data were collected.}

\item{\code{processingLog}}{As with all \code{oce} objects, the \code{processingLog}
slot is a \code{\link{list}} with entries describing the creation and evolution
of the object. The contents are updated by various \code{oce} functions to
keep a record of processing steps.  Object summaries and
\code{\link{processingLogShow}} both display the log.}
}}

\section{Retrieving slot contents (general)}{


The full contents of the \code{data} and \code{metadata} slots of a \code{oce}
object named \code{obj} may be retrieved in the standard R way, with
\code{slot(obj, "data")} and \code{slot(obj, "metadata")}, or with
\code{obj[["data"]]} and \code{obj[["metadata"]]}, where the latter uses an
overloading of the \code{[[} operator for \code{oce} objects.

The \code{[[} mechanism can also be used to look for named items within
the slots. It first checks the \code{metadata} slot, returning the value
if it is there, and then tries in the \code{data} slot if not. If this
order of precedence is unsuitable, \code{\link{oceGetData}} and
\code{\link{oceGetMetadata}} can be used.

In addition to looking up stored quantities, the \code{[[} operator
has other useful capabilities. If an object holds data-quality flags, then
\code{obj[["flags"]]} will retrieve them, as a list with names matching
the variable names.  Similarly, \code{obj[["units"]]} return a list of units.
Individual items may be found in the usual R way within these lists, but there
is a simpler way: use e.g. \code{obj[["temperatureFlag"]]} or
\code{obj[["temperatureUnit"]]}. In such cases, \code{[[} is returning
information that is stored somewhere in the object, but it can also return
wholly \strong{derived} quantities, depending on the object type. For
example, \code{obj[["N2"]]} will return the square of the buoyancy
frequency, if \code{obj} is an object from which that can be calculated
(such as a \code{ctd} object). Information about accessing such derived
quantities may be found in the \dQuote{Special access methods}
section of the documentation for those classes that support it.
}

\section{Modifying slot contents (general)}{


Although there is a \code{[[<-} notation for modifying the contents
of \code{oce} objects, it is wiser to use \code{\link{oceSetData}}
\code{\link{oceSetMetadata}}, respectively, to be sure where the change will be
made, and to record the change in the \code{processingLog}.
}

\section{Special access methods for ctd objects}{


The TEOS-10 notation for these quantities also works, with \code{ctd[["SP"]]},
\code{ctd[["t"]]} and \code{ctd[["p"]]} returning identical values to those
returned for the longer names.

After the names listed above have been checked, the remaining names in the
\code{data} slot are checked using \code{\link{pmatch}}, so that e.g.
\code{ctd[["sal"]]} will recover practical \code{salinity}, \code{ctd[["sc"]]}
will recover \code{scan} (if it exists), etc.

Data may be extracted with \code{\link{[[,ctd-method}}.

Note that \code{\link{[[,ctd-method}}
returns temperature in the ITS-90 scale and pressure in dbar, regardless of the
scale and unit of the data within the object. Type \code{?"[[,ctd-method"}
or \code{?"[[<-,ctd-method"} to learn more.

Depth is accessed with e.g. \code{ctd[["depth"]]}, while its negative, the
vertical coordinate, is accessed with e.g.  \code{ctd[["z"]]}; note that these
are calculated using \code{\link{swDepth}} and \code{\link{swZ}}, and that any
values that may have been read in a data file are ignored.

Potential temperature defined according to UNESCO-1980 is calculated with
\code{ctd[["theta"]]} or \code{ctd[["potential temperature"]]}.  Salinity is
retrieved with \code{ctd[["S"]]} or \code{ctd[["salinity"]]}.

Conservative Temperature defined according to TEOS-2010 is calculated with
\code{ctd[["CT"]]} or \code{ctd[["conservative temperature"]]}.  Absolute
salinity is calculated with \code{ctd[["SA"]]} or \code{ctd[["absolute
salinity"]]}. Note that the \code{CT}, \code{SA} and \code{Sstar} calculations
require latitude and longitude, and so errors result if these items
are sought for a \code{ctd} object that lacks latitude or longitude.
(Until 2017 May 14, defaults of 300E and 30N were used if position
was not stored in the object.)

The square of buoyancy frequency is retrieved with \code{ctd[["N2"]]} or
\code{\link{swN2}}, density ratio with \code{ctd[["Rrho"]]} and spiciness with
\code{ctd[["spice"]]}.
}

\section{Reading/creating CTD objects}{

A file containing CTD profile data may be read with
\code{\link{read.ctd}}, and a CTD object can also be created with
\code{\link{as.ctd}}.  See \code{\link{read.ctd}} for references on data
formats used in CTD files. Data can also be assembled into
\code{ctd} objects with \code{\link{as.ctd}}.

Statistical summaries are provided by \code{\link{summary,ctd-method}}, while
\code{\link{show}} displays an overview.

CTD objects may be plotted with \code{\link{plot,ctd-method}}, which does much of its
work by calling \code{\link{plotProfile}} or \code{\link{plotTS}}, both of
which can also be called by the user, to get fine control over the plots.

A CTD profile can be isolated from a larger record with \code{\link{ctdTrim}},
a task made easier when \code{\link{plotScan}} is used to examine the results.
Towyow data can be split up into sets of profiles (ascending or descending)
with \code{\link{ctdFindProfiles}}.  CTD data may be smoothed and/or cast onto
specified pressure levels with \code{\link{ctdDecimate}}.

As with all oce objects, low-level manipulation may be done with
\code{\link{oceSetData}} and \code{\link{oceSetMetadata}}. Additionally,
many of the contents of CTD objects may be altered with the \code{\link{[[,ctd-method}} scheme
discussed above, and skilled users may also manipulate the contents directly.
}

\examples{

# 1. Create a ctd object with fake data.
a <- as.ctd(salinity=35+1:3/10, temperature=10-1:3/10, pressure=1:3)
summary(a)

# 2. Fix a typo in a station latitude (fake! it's actually okay)
data(ctd)
ctd <- oceSetMetadata(ctd, "latitude", ctd[["latitude"]]-0.001,
                     "fix latitude typo in log book")

}
\seealso{
Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{cnvName2oceName}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdRaw}},
  \code{\link{ctdTrim}}, \code{\link{ctd}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{initialize,ctd-method}},
  \code{\link{oceNames2whpNames}},
  \code{\link{oceUnits2whpUnits}},
  \code{\link{plot,ctd-method}}, \code{\link{plotProfile}},
  \code{\link{plotScan}}, \code{\link{plotTS}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}},
  \code{\link{woceNames2oceNames}},
  \code{\link{woceUnit2oceUnit}}, \code{\link{write.ctd}}

Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{coastline-class}},
  \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{met-class}}, \code{\link{oce-class}},
  \code{\link{odf-class}}, \code{\link{rsk-class}},
  \code{\link{sealevel-class}},
  \code{\link{section-class}}, \code{\link{topo-class}},
  \code{\link{windrose-class}}
}
\author{
Dan Kelley
}
