\name{ctd}

\docType{data}

\alias{ctd}

\title{Seawater CTD profile}

\description{This is sample CTD profile provided for testing, created by
  the code given in \dQuote{Examples}.}

\usage{data(ctd)}

\examples{ctd <- read.oce(system.file("data", "ctd.cnv", package="oce"))}

\source{Peter Galbraith acquired these data during the \acronym{SLEIWEX}
field program, part of a Dalhousie/IML cooperative study of
internal-wave mixing in the St Lawrence estuary.}

\seealso{See \code{\link{read.ctd}} for the data format, and for a list
  of the many functions that work with objects of class \code{"ctd"}, of
  which this is a sample.}

\keyword{misc}
