\name{ctd}
\docType{data}
\alias{ctd}
\title{Sample seawater (CTD) profile}
\description{This is sample CTD profile provided for testing.
It has been pruned by selecting only the portion of the data
acquired when the instrument was being lowered through the water
column.}
\usage{data(ctd)}
\format{
  This file is from a \emph{Sea-Bird SBE 19} CTD instrument.
}
\source{Peter Galbraith acquired these data during the 2002 SLEIWEX
  field program, part of a Dalhousie/IML cooperative study of
  internal-wave mixing in the St Lawrence estuary.} 
\references{
  The Seabird CTD instrument is described at
  \url{http://www.seabird.com/products/spec_sheets/19plusdata.htm}.  
  See link{read.ctd} for the data format.
}
\author{Dan Kelley \email{Dan.Kelley@Dal.Ca}}
\keyword{misc}
