\name{ctd}

\docType{data}

\alias{ctd}

\title{Seawater CTD profile}

\description{This is sample CTD profile provided for testing, created by
  the code given in \dQuote{Examples}.}

\usage{data(ctd)}

\examples{
library(oce)
data(ctd)
plot(ctd)
}

\source{This is a trimmed version of the \code{\link{ctdRaw}} dataset.}

\seealso{See \code{\link{ctdRaw}}, from which this is derived.  See
    \code{\link{read.ctd}} for the data format, and for a list of the many
    functions that work with objects of class \code{"ctd"}, of which this is a
    sample.}

\keyword{datasets}
