% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctdUpdateHeader}
\alias{ctdUpdateHeader}
\title{Update a CTD header}
\usage{
ctdUpdateHeader(x, debug = FALSE)
}
\arguments{
\item{x}{A \code{ctd} object, e.g. as read by \code{\link{read.ctd}}.}

\item{debug}{Set to \code{TRUE} for debugging.}
}
\value{
A new \code{\link{ctd-class}} object.
}
\description{
Update the header of a \code{ctd} object, e.g. adjusting \code{nvalues} and the
\code{span} of each column. This is done automatically by \code{ctdTrim}, for
example.
}
\examples{
library(oce)
data(ctd) 
ctd[["pressure"]] <- ctd[["pressure"]] + 3
ctdNew <- ctdUpdateHeader(ctd)

}
\author{
Dan Kelley
}
\references{
The Seabird CTD instrument is described at
  \url{http://www.seabird.com/products/spec_sheets/19plusdata.htm}.
}
\seealso{
Other functions that handle CTD data: \code{\link{[[,ctd-method}},
  \code{\link{as.ctd}}, \code{\link{ctd-class}},
  \code{\link{ctdAddColumn}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdTrim}},
  \code{\link{handleFlags,argo-method}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.odv}}, \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}}, \code{\link{write.ctd}}
}

