% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctdUpdateHeader}
\alias{ctdUpdateHeader}
\title{Update a CTD Header (defunct)}
\usage{
ctdUpdateHeader(x, debug = FALSE)
}
\arguments{
\item{x}{a \linkS4class{ctd} object.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
A \linkS4class{ctd} object.
}
\description{
\emph{W}ARNING:** This function will be removed soon; see \link{oce-defunct}.
}
\details{
Update the header of a \code{ctd} object, e.g. adjusting \code{nvalues} and the
\code{span} of each column. This is done automatically by \code{ctdTrim}, for
example.
}
\references{
The Seabird CTD instrument is described at
\code{http://www.seabird.com/products/spec_sheets/19plusdata.htm}.
}
\seealso{
Other functions that will be removed soon: \code{\link{addColumn}},
  \code{\link{ctdAddColumn}}, \code{\link{findInOrdered}},
  \code{\link{mapMeridians}}, \code{\link{mapZones}},
  \code{\link{oce.as.POSIXlt}}
}
\author{
Dan Kelley
}
\concept{functions that will be removed soon}
