% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\name{download.amsr}
\alias{download.amsr}
\title{Download an AMSR2 file from the server, caching result in destdir}
\usage{
download.amsr(year, month, day, destdir = ".",
  server = "ftp.ssmi.com/amsr2/bmaps_v07.2")
}
\arguments{
\item{year, month, day}{Numerical values of the year, month, and day.
If the first two of these are missing and the third is negative, 
\code{download.amsr} will seek data relative to the present date, e.g.
specifying \code{day=-3} will seek an image from three days ago (which
may be the most recent image that is available, depending on the timezone
and the time the function is called).}

\item{destdir}{Character value of the directory in which to cache resultant files}

\item{sever}{Character value of the server}
}
\description{
If the file is already present in \code{destdir}, then it is not
downloaded again. The default \code{destdir} is the present directory,
but it probably makes more sense to use something like \code{"~/data/amsr"}
or something of that nature, so that other scripts can also use the
cached data.
}
\details{
Messages are printed to indicate the progress of the download, or to
indicate that a cached value is being used.
}
\section{Limitation}{

This function relies on the system utility \code{ftp}, and also on local directories
being separated by forward slashes in the file system. That means it will probably
only work on unix-like systems.
}

