% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{download.met}
\alias{download.met}
\title{Download and Cache a met File}
\usage{
download.met(id, year, month, destdir = "~/data/met", destfile,
  debug = getOption("oceDebug"))
}
\arguments{
\item{id}{A number giving the "Station ID" of the station of interest. If not
provided, \code{id} defaults to 6358, for Halifax International Airport. See
\dQuote{Details}.}

\item{year}{A number giving the year of interest. This defaults to the present
year, if not given.}

\item{month}{A number giving the month of interest. This defaults to the present
month, if not given.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, the file name is constructed from the other
parameters of the function call, so subsequent calls with the same
parameters will yield the same result, thus providing the key
to the caching scheme.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
Data are downloaded from \url{http://climate.weather.gc.ca} and cached locally.
}
\details{
The data are downloaded with \code{\link[utils]{download.file}}
pointed to the Environment Canada website \url{http://climate.weather.gc.ca},
using queries that had to be devised by reverse-engineering, since the agency
provides no documentation about the queries. Changes to query format are not
by any means unheard-of; queries that worked in 2016 failed in early 2017, for
example.

The Station ID that is provided in the \code{id} argument
becomes part of the query used to download the data.
Confusingly, this is \emph{neither} the "Climate ID"
\emph{nor} the "WMO ID". Instead, it seems to be a creation of Environment
Canada.  Even worse, the Environment Canada documents state that the ID for 
a particular location may be changed at any time. Users are therefore advised to
look up the codes at
\code{ftp://ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees//Station Inventory EN.csv}
before using this function. This file can be searched by city name, and in other
ways.  (A future version of \code{download.met} might use this file, if enough
users request this feature.)
}
\section{Warning}{

All of the \code{download.*} functions are subject to change through the
year 2017. The plan is to create individual functions
as needs arise in everyday work, thereby discovering patterns of 
arguments that might make sense more generally. Therefore, it is likely
that argument order (and even name) may change up until approximately
December 2016. Users are advised to refer to arguments by name,
i.e. not to rely on their order, and they are asked for patience,
because argument names and defaults may vary through the early development.
}
\examples{
\dontrun{
library(oce)
## Download data for Halifax International Airport, in September
## of 2003. (This dataset is used for data(met) provided with oce.)
metFileName <- download.met(6358, 2003, 9)
}

}
\author{
Dan Kelley
}
\seealso{
The work is done with \code{\link[utils]{download.file}}.

Other functions that download files: \code{\link{download.amsr}},
  \code{\link{download.coastline}},
  \code{\link{download.topo}}

Other things related to \code{met} data: \code{\link{[[,met-method}},
  \code{\link{[[<-,met-method}}, \code{\link{as.met}},
  \code{\link{met-class}}, \code{\link{met}},
  \code{\link{plot,met-method}}, \code{\link{read.met}},
  \code{\link{subset,met-method}},
  \code{\link{summary,met-method}}
}

