% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{download.topo}
\alias{download.topo}
\title{Download and Cache a topo File}
\usage{
download.topo(
  west,
  east,
  south,
  north,
  resolution = 4,
  destdir = ".",
  destfile,
  format,
  server = "https://gis.ngdc.noaa.gov",
  debug = getOption("oceDebug")
)
}
\arguments{
\item{west, east}{Longitudes of the western and eastern sides of the box.}

\item{south, north}{Latitudes of the southern and northern sides of the box.}

\item{resolution}{numeric value of grid spacing, in geographical minutes.
The default value of 4 minutes corresponds to 4 nautical miles, or 7.4km.}

\item{destdir}{Optional string indicating the directory in which to store
downloaded files. If not supplied, \code{"."} is used, i.e. the data file
is stored in the present working directory.}

\item{destfile}{Optional string indicating the name of the file.
If not supplied, the file name is constructed from the other
parameters of the function call, so subsequent calls with the same
parameters will yield the same result, thus providing the key
to the caching scheme.}

\item{format}{Deprecated, and ignored, as of June 2020.}

\item{server}{character value specifying the base from which a
download URL will be constructed.  It is unlikely that any value
other than the default will work, unless it is a similarly-constructed
mirrored site.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
Topographic data are downloaded from a data server that holds the ETOPO1
dataset (Amante, C. and B.W. Eakins, 2009), and saved as a netCDF file whose
name specifies the data request, if a file of that name is not already
present on the local file system.  The return value is the name of the data
file, and it's typical use is as the filename for a call to \code{\link[=read.topo]{read.topo()}}.
Given the rules on file naming, subsequent calls to \code{download.topo()}
with identical parameters will simply return the name of the cached file,
assuming the user has not deleted it in the meantime.
}
\details{
For years, the server used by \code{download.topo()} was capable of
returning netCDF files, but this was found to fail in late May of 2020.
Luckily, the \code{\link[marmap:getNOAA.bathy]{marmap::getNOAA.bathy()}} function in the \CRANpkg{marmap}
package was updated to handle the changes in the server, and
\code{download.topo()} was able to pattern its URL construction on that function.
The server also now has an API, which might suggest it will be less subject
to change in the future.
}
\examples{
\dontrun{
library(oce)
topoFile <- download.topo(west=-66, east=-60, south=43, north=47,
                          resolution=1, destdir="~/data/topo")
topo <- read.topo(topoFile)
imagep(topo, zlim=c(-400, 400), drawTriangles=TRUE)
if (requireNamespace("ocedata", quietly=TRUE)) {
    data(coastlineWorldFine, package="ocedata")
    lines(coastlineWorldFine[["longitude"]], coastlineWorldFine[["latitude"]])
}
}

}
\references{
\itemize{
\item Amante, C. and B.W. Eakins, 2009. ETOPO1 1 Arc-Minute Global Relief
Model: Procedures, Data Sources and Analysis. NOAA Technical Memorandum
NESDIS NGDC-24. National Geophysical Data Center, NOAA. \doi{doi:10.7289/V5C8276M}
}
}
\seealso{
Other functions that download files: 
\code{\link{download.amsr}()},
\code{\link{download.coastline}()},
\code{\link{download.met}()}

Other things related to topo data: 
\code{\link{[[,topo-method}},
\code{\link{[[<-,topo-method}},
\code{\link{as.topo}()},
\code{\link{plot,topo-method}},
\code{\link{read.topo}()},
\code{\link{subset,topo-method}},
\code{\link{summary,topo-method}},
\code{\link{topo-class}},
\code{\link{topoInterpolate}()},
\code{\link{topoNS}},
\code{\link{topoWorld}}
}
\author{
Dan Kelley
}
\concept{functions that download files}
\concept{things related to topo data}
