% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echosounder.R
\docType{data}
\name{echosounder}
\alias{echosounder}
\title{echosounder dataset}
\source{
This file came from the SLEIWEX-2008 experiment, and was decimated
using \code{\link{decimate}} with \code{by=c()}.
}
\description{
This is degraded subsample of measurements that were made with a Biosonics
scientific echousounder, as part of the St Lawrence Internal Wave Experiment
(SLEIWEX).
}
\author{
Dan Kelley
}
\seealso{
Other datasets provided with \code{oce}: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{lisst}},
  \code{\link{lobo}}, \code{\link{met}},
  \code{\link{section}}, \code{\link{topoWorld}}
}

