% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echosounder.R
\docType{data}
\name{echosounder}
\alias{echosounder}
\title{Echosounder Dataset}
\source{
This file came from the SLEIWEX-2008 experiment, and was decimated
using \code{\link[=decimate]{decimate()}} with \code{by=c()}.
}
\description{
This is degraded subsample of measurements that were made with a Biosonics
scientific echosounder, as part of the St Lawrence Internal Wave Experiment
(SLEIWEX).
}
\seealso{
Other datasets provided with oce: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{landsat}},
  \code{\link{lisst}}, \code{\link{lobo}},
  \code{\link{met}}, \code{\link{ocecolors}},
  \code{\link{rsk}}, \code{\link{sealevelTuktoyaktuk}},
  \code{\link{sealevel}}, \code{\link{section}},
  \code{\link{topoWorld}}, \code{\link{wind}},
  \code{\link{xbt}}

Other things related to echosounder data: \code{\link{[[,echosounder-method}},
  \code{\link{[[<-,echosounder-method}},
  \code{\link{as.echosounder}},
  \code{\link{echosounder-class}},
  \code{\link{findBottom}},
  \code{\link{plot,echosounder-method}},
  \code{\link{read.echosounder}},
  \code{\link{subset,echosounder-method}},
  \code{\link{summary,echosounder-method}}
}
\author{
Dan Kelley
}
\concept{datasets provided with oce}
\concept{things related to echosounder data}
