% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findInOrdered}
\alias{findInOrdered}
\title{Find indices of times in an ordered vector [defunct]}
\usage{
findInOrdered(x, f)
}
\arguments{
\item{x}{Ignored, since this function is defunct.}

\item{f}{Ignored, since this function is defunct.}
}
\description{
\strong{WARNING:} This function will be removed soon; see \link{oce-defunct}.
}
\seealso{
Other functions that will be removed soon: \code{\link{addColumn}},
  \code{\link{ctdAddColumn}},
  \code{\link{ctdUpdateHeader}},
  \code{\link{mapMeridians}}, \code{\link{mapZones}},
  \code{\link{oce.as.POSIXlt}}
}
\author{
Dan Kelley
}
