% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{firstFinite}
\alias{firstFinite}
\title{Get first finite value in a vector or array, or NULL if none}
\usage{
firstFinite(v)
}
\arguments{
\item{v}{A numerical vector or array.}
}
\description{
Get first finite value in a vector or array, or NULL if none
}

