% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gappyIndex}
\alias{gappyIndex}
\title{Create an indexing array}
\usage{
gappyIndex(starts, from, to)
}
\arguments{
\item{starts}{Integer vector of one or more values.}

\item{from, to}{Integer value specifying the region to be created.}
}
\description{
This is used internally to construct indexing arrays, mainly
for adv and adp functions, in which \code{\link[=readBin]{readBin()}} is used to
read chunks from a buffer of \link{raw} values.  The example
shows how to create the sequence \code{c(3,4, 103,104)}.
}
\examples{
library(oce)
gappyIndex(c(1, 101), 2, 3)
}
