% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gappyIndex}
\alias{gappyIndex}
\title{Create a Possibly Gappy Indexing Vector}
\usage{
gappyIndex(starts, from, to)
}
\arguments{
\item{starts}{integer vector of one or more values.}

\item{from, to}{integer values controlling the generated sequence, as if
\link{seq(from,to)} had been used. See \dQuote{Details}.}
}
\description{
This is used internally to construct indexing arrays, mainly for adv and adp
functions, in which \code{\link[=readBin]{readBin()}} is used to access isolated regions within a
\link{raw} vector. The work is done in C++, for speed.
}
\details{
For example, suppose data elements in a buffer named \code{buf} start at bytes
1000 and 2000, and that the goal is to read the first 4 bytes of each of
these sequences as an unsigned, 4-byte integer.  This could be accomplished
as follows.  Note that the last argument to \code{gappyIndex()} is 3 in this
example, because the \code{0:3} has 4 elements.

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(oce)
buf <- readBin("filename", "raw", n=5000, size=1)
i <- gappyIndex(c(1000, 2000), 0, 3) # Note the length(0:3) is 4.
# i is 1000,1001,1002,1003,2000,2001,2002,2003
readBin(buf[i], "integer", size=4, n=1, endian="little")
}\if{html}{\out{</div>}}
}
\author{
Dan Kelley
}
