% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{geodXyInverse}
\alias{geodXyInverse}
\title{inverse geodesic calculation}
\usage{
geodXyInverse(x, y, lon.ref = 0, lat.ref = 0)
}
\arguments{
\item{x}{value of x in metres, as given by \code{\link{geodXy}}}

\item{y}{value of y in metres, as given by \code{\link{geodXy}}}

\item{lon.ref}{reference longitude, as supplied to \code{\link{geodXy}}}

\item{lat.ref}{reference latitude, as supplied to \code{\link{geodXy}}}
}
\value{
a data frame containing \code{longitude} and \code{latitude}
}
\description{
inverse geodesic calculation
}
\details{
The calculation is done by finding a minimum value of a cost
function that is the vector difference between (\code{x},\code{y})
and the corresponding values returned by \code{\link{geodXy}}. This
minimum is calculated in C for speed, using the \code{nmmin} function
that is the underpinning for the Nelder-Meade version of the R function
\code{\link{optim}}.
}

