% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\docType{class}
\name{gps-class}
\alias{gps-class}
\title{Class to Store GPS Data}
\description{
Class to store gps data. These objects may be read with
\code{\link{read.gps}} or assembled with \code{\link{as.gps}}.
}
\seealso{
Other things related to \code{gps} data: \code{\link{[[,gps-method}},
  \code{\link{[[<-,gps-method}}, \code{\link{as.gps}},
  \code{\link{plot,gps-method}}, \code{\link{read.gps}},
  \code{\link{summary,gps-method}}
}
\author{
Dan Kelley
}
