% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\docType{class}
\name{gps-class}
\alias{gps-class}
\title{Class to store gps data}
\description{
Class to store gps data. These objects may be read with
\code{\link{read.gps}} or assembled with \code{\link{as.gps}}.
}
\author{
Dan Kelley
}
\seealso{
Other functions that handle \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{ctdAddColumn}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdTrim}},
  \code{\link{ctdUpdateHeader}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}}, \code{\link{plotScan}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.odv}}, \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}}, \code{\link{write.ctd}}
}

