% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{handleFlags,ctd-method}
\alias{handleFlags,ctd-method}
\title{Handle flags in \code{ctd} objects}
\usage{
\S4method{handleFlags}{ctd}(object, flags = list(), actions = list())
}
\arguments{
\item{object}{An object of \code{\link{ctd-class}}.}

\item{flags}{An optional \code{\link{list}} containing (a)
items with names of entries in the \code{data} slot of \code{object},
or (b) a single unnamed item. In the first case, the attention is
focussed on the named items, while in the second case the 
all the data in the \code{object}'s \code{data} slot are examined.
Each element in the list must be set to an integer or vector of integers,
specifying conditions to be met before actions are to be taken.}

\item{actions}{An optional \code{\link{list}} that contains items with
names that match those in the \code{flags} argument.  If \code{actions}
is not supplied, the default will be to set all values identified by
\code{flags} to \code{NA}; this can also be specified by
specifying \code{actions=list("NA")}. It is also possible to specify 
functions that calculate replacement values. These are provided
with \code{object} as the single argument, and must return a
replacement for the data item in question.}
}
\description{
Data-quality flags are stored in the \code{metadata}
slot of \code{\link{oce-class}} objects in a 
\code{\link{list}} named \code{flags}.
The present function (a generic that has specialized versions
for various data classes) provides a way to
manipulate the core data based on
the data-quality flags. For example, a common operation is to replace suspicious
or erroneous data with \code{NA}.

If \code{metadata$flags} in the object supplied as the first argument
is empty, then that object is returned, unaltered.
Otherwise, \code{handleFlags} analyses the data-quality flags within
the object, in relation to the \code{flags} argument, and interprets
the \code{action} argument to select an action to be applied to mached
data.

Reasonable defaults are used if \code{flags} and \code{actions}
are not supplied (see \sQuote{Details}),
but different schemes are used in different
data archives, so it is risky to rely on these defaults.
It is usually necessary to tailor \code{flags} and \code{actions} 
to the data and the analysis goals.
}
\details{
If \code{flags} and \code{actions} are not provided, the
default is to use WHP (World Hydrographic Program) flags [1], in which the
value 2 indicates good data, and other values indicate either unchecked,
suspicious, or bad data. Any data not flagged as good are set
to \code{NA} in the returned value. Since WHP flag codes run
from 1 to 9, this default is equivalent to
setting \code{flags=list(all=c(1, 3:9))} along with
\code{action=list("NA")}.
}
\section{Implementation status}{
 \code{handleFlags} is a new function as of March 2016,
and it will probably change through the Spring of 2016.
Almost nothing works yet, and users should not be doing
any more than looking at the documentation and telling the developers
whether the planned functioning seems reasonable.
The hope is to get trial code working for \code{ctd}
and \code{argo} data types by the end of April. After that, there
will likely  be a month or more of testing with real-world
work, with possible changes to the user interface. Then other
types will be added, as needed.
}
\examples{
library(oce)
data(section)
stn <- section[["station", 100]]
# 1. Default: anything not flagged as 2 is set to NA, to focus
# solely on 'good', in the World Hydrographic Program scheme.
STN <- handleFlags(stn)
data.frame(old=stn[['salinity']], flag=stn[['salinityFlag']], new=STN[['salinity']])

# 2. A less restrictive case: include also 'questionable' data,
# and only apply this action to salinity.
STN <- handleFlags(stn, flags=list(salinity=c(1, 4:9)))

# 3. Use smoothed TS relationship to nudge questionable data.
# This is perhaps a silly idea, but at least it illustrates
# how to write a nontrivial function for an action.
f<-function(x) {
  S <- x[["salinity"]]
  T <- x[["temperature"]]
  df <- 0.5 * length(S) # smooths a bit
  sp <- smooth.spline(T, S, df=df)
  0.5 * (S + predict(sp, T)$y)
}
par(mfrow=c(1,2))
STN <- handleFlags(stn, flags=list(salinity=c(1,3:9)), action=list(salinity=f))
plotProfile(stn, "salinity", mar=c(3, 3, 3, 1))
p <- stn[['pressure']]
par(mar=c(3, 3, 3, 1))
plot(STN[['salinity']] - stn[['salinity']], p, ylim=rev(range(p)))

}
\references{
1. \url{https://www.nodc.noaa.gov/woce/woce_v3/wocedata_1/whp/exchange/exchange_format_desc.htm}
}
\seealso{
Other functions that handle data-quality flags: \code{\link{handleFlags,argo-method}},
  \code{\link{handleFlags,section-method}},
  \code{\link{handleFlags}}

Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{cnvName2oceName}}, \code{\link{ctd-class}},
  \code{\link{ctdAddColumn}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdRaw}},
  \code{\link{ctdTrim}}, \code{\link{ctdUpdateHeader}},
  \code{\link{ctd}}, \code{\link{gps-class}},
  \code{\link{plot,ctd-method}}, \code{\link{plotScan}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.odv}}, \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}}, \code{\link{write.ctd}}
}

