% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{handleFlags,vector-method}
\alias{handleFlags,vector-method}
\title{Signal erroneous application to non-oce objects}
\usage{
\S4method{handleFlags}{vector}(object, flags = list(),
  actions = list(), debug = getOption("oceDebug"))
}
\arguments{
\item{object}{A vector, which cannot be the case for \code{oce} objects.}

\item{flags}{Ignored.}

\item{actions}{Ignored.}

\item{debug}{Ignored.}
}
\description{
Signal erroneous application to non-oce objects
}
