% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{handleFlags}
\alias{handleFlags}
\title{Set any flagged data to NA}
\usage{
handleFlags(x, action = "NA")
}
\arguments{
\item{x}{an oce object}

\item{action}{the action to be undertaken.}
}
\value{
an oce object
}
\description{
If this is applied to a non-oce object, then that object is returned
unchanged. If it is applied to an object that lacks \code{metadata$flags},
then \code{x} is again returned unchanged, but a warning is issued.
}
\details{
If \code{action} is \code{"NA"}, then any flagged data are set to \code{NA}.
Any other \code{action} yields an error.
}

