% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{handleFlags}
\alias{handleFlags}
\title{Handle flags in oce objects}
\usage{
handleFlags(object, flags, actions)
}
\arguments{
\item{object}{An object of \code{\link{oce}}.}

\item{flags}{An optional \code{\link{list}} containing items with names of
entries in the \code{data} slot of \code{object}, or a single
item named \code{all} (which indicates that the operations are to take
place on each element in the \code{data} slot of \code{object}).
For example, the WOCE CTD scheme has flags ranging from 1 to 9, with 2
indicating good data. Therefore, all the data that are not flagged
as "good" are identified with \code{flags=list(all=(1:9)[-2])}.
If only certain data types are of interest, they may be named in the list,
e.g. \code{flags=list(temperature=(1:9)[-2])} focusses on the questional
temperature values. \code{handleFlags} ignores flag values that are not
used in the dataset, so the above could use \code{(1:100)[-2]}
to be robust against the addition of codes in excess of 9.}

\item{actions}{An optional \code{\link{list}} that contains items with
names that match those in the \code{flags} argument.  If \code{actions}
is not supplied, the default will be to set all values identified by
\code{flags} to \code{NA}. This is also accomplished by e.g.
\code{actions=list(temperature="NA")}. It is also possible to specify a
function that will calculate a replacement value. This function is provided
with \code{object} as its single argument, and it must return a
replacement for the data item in question. Thus, for example, it might
handle bad salinity data by filling in values from a known TS relationship,
using good temperatures in \code{object}.}
}
\description{
\code{handleFlags} is a generic function that has specialized variants
that are used for different types of \code{oce} object.  In each case, 
the purpose is to help analysts
deal with data-quality flags, by performing specified actions on data elements that 
have been flagged (e.g. by data-provision agencies) in specified ways.  For example,
the action could be to set the data to the R missing-value code, \code{NA},
and the identification could be a particular value of a data-quality flag.

The flag identification is set with the \code{flags} argument, while the action
is set with the \code{actions} argument.  Both of these can be omitted, in which
case a reasonable default is used for the particular data type (i.e. for the 
class of the \code{object} argument).  However, data-quality flags differ
between archiving agencies, and it is common to tailor 
\code{flags} and \code{actions} to match the peculiarities of the data
under analysis. These flags also permit the modification of data, e.g.
to fill gaps.

It is important to note that \code{oce} stores flags as 
non-negative integers. The flags are set up when the data are read.
If the data hold numerical (integer) flags, then these are copied into
the \code{oce} object. However, if the data source specifies flags with letters
or other non-integer values, then the function used to read the data will 
map the flags to integers, as described in the documentation for that
function.
}
\section{A general note}{

Each specialized variant of this function has its own defaults
for \code{flags} and \code{actions}; see the specialized 
documentation.
}

\section{Implementation status}{
 \code{handleFlags} is a new function as of March 2016,
and it will probably change through the Spring of 2016.
Almost nothing works yet, and users should not be doing
any more than looking at the documentation and telling the developers
whether the planned functioning seems reasonable.
The hope is to get trial code working for \code{ctd}
and \code{argo} data types by the end of April. After that, there
will likely  be a month or more of testing with real-world
work, with possible changes to the user interface. Then other
types will be added, as needed.
}
\seealso{
Other functions that handle data-quality flags: \code{\link{handleFlags,ctd-method}}
}

