% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{head.oce}
\alias{head.oce}
\title{Extract The Start of an Oce Object}
\usage{
\method{head}{oce}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{oce} object.}

\item{n}{Number of elements to extract.}

\item{...}{ignored}
}
\description{
NOTE: this is a preliminary function, subject to change. So far
it has been coded and tested only adp and ctd types; whetheter
it will do anything with other types is a question. Please contact
the author if you need this to work on a specific type.
}
\seealso{
\code{\link{tail.oce}}, which yields the end of an \code{oce} object.
}
\author{
Dan Kelley
}
