% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{head.oce}
\alias{head.oce}
\title{Extract The Start of an Oce Object}
\usage{
\method{head}{oce}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{oce} object.}

\item{n}{Number of elements to extract, as for \code{\link{head}}.}

\item{...}{ignored}
}
\description{
Extract The Start of an Oce Object

The plan is that most classes should be covered by July 2018.
Please contact the author if there is a class you need handled
before that date.
}
\seealso{
\code{\link{tail.oce}}, which yields the end of an \code{oce} object.
}
\author{
Dan Kelley
}
