% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{head.oce}
\alias{head.oce}
\title{Extract The Start of an Oce Object}
\usage{
\method{head}{oce}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{oce} object.}

\item{n}{Number of elements to extract, as for \code{\link{head}}.}

\item{...}{ignored}
}
\description{
Extract The Start of an Oce Object

This function handles the following object classes directly:
\code{\link{adp-class}},
\code{\link{adv-class}},
\code{\link{ctd-class}},
\code{\link{section-class}}.
and
\code{\link{topo-class}}. It does not handle
\code{\link{amst-class}} or
\code{\link{landsat-class}}
at all, and simply returns \code{x} in those cases.
For all other classes, it simply calls \code{\link{head}}
with \code{n} as provided, for each item in the \code{data}
slot, issuing a warning if that item is not a vector.
}
\details{
Please contact the author if there is a class you need handled.
}
\seealso{
\code{\link{tail.oce}}, which yields the end of an \code{oce} object.
}
\author{
Dan Kelley
}
