% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{integerToAscii}
\alias{integerToAscii}
\title{Decode integer to corresponding ASCII code}
\usage{
integerToAscii(i)
}
\arguments{
\item{i}{an integer, or integer vector.}
}
\value{
A character, or character vector.
}
\description{
Decode integer to corresponding ASCII code
}
\examples{

library(oce)
A <- integerToAscii(65)
cat("A=", A, "\\n")
}
\author{
Dan Kelley
}

