% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladp.R
\docType{class}
\name{ladp-class}
\alias{ladp-class}
\title{Class to store lowered-adp data}
\description{
Class to store data measured with a lowered ADP (also
known as ADCP) device.
}
\section{Accessing data}{


    Consider an object named \code{ladp}.

    Metadata (contained in the S4 slot named \code{metadata}) may be retrieved
    or set by name, \code{ladp[["longitude"]] <- ladp[["longitude"]] + 1}
    corrects a one-degree error.  Use \code{names(ladp@metadata)} to find the
    names of the metadata.

    Column data may be accessed by name, e.g. \code{ladp[["u"]]},
    \code{ladp[["v"]]}, \code{ladp[["pressure"]]}, etc.  There may also be
    columns for \code{"temperature"} and \code{"salinity"}, and possibly other
    things. Use \code{names(ladp@data)} to find the names of the data.
}

\seealso{
Other things related to \code{ladp} data: \code{\link{[[,ladp-method}},
  \code{\link{as.ladp}}, \code{\link{plot,ladp-method}},
  \code{\link{summary,ladp-method}}
}
\author{
Dan Kelley
}
