% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landsat.R
\docType{data}
\name{landsat}
\alias{landsat}
\title{Sample landsat dataset.}
\description{
This is a landsat-8 dataset
This is a subset of the Landsat-8 image designated LC80080292014065LGN00, an
image from March 2014 that covers Nova Scotia and portions of the Bay of
Fundy and the Scotian Shelf. The image is decimated to reduce the memory
requirements of this package, yielding a spatial resolution of about 2km.
}
\details{
The original data were downloaded from the USGS earthexplorer website, although
other sites can also be used to uncover it by name.  The code below shows how
the dataset was created.  The decimation by 100 reduces file size from about 1GB
to under 100Kb.
}
\keyword{data}
\keyword{satellite}

