% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat.R
\docType{data}
\name{landsat}
\alias{landsat}
\title{Sample landsat Dataset}
\description{
This is a subset of the Landsat-8 image designated LC80080292014065LGN00, an
image from March 2014 that covers Nova Scotia and portions of the Bay of
Fundy and the Scotian Shelf. The image is decimated to reduce the memory
requirements of this package, yielding a spatial resolution of about 2km.
}
\details{
The original data were downloaded from the USGS earthexplorer website, although
other sites can also be used to uncover it by name.  The original
data were decimation by a factor of 100 to reduce the file size from about 1GB
to under 100Kb.
}
\seealso{
Other datasets provided with \code{oce}: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{colors}},
  \code{\link{ctdRaw}}, \code{\link{ctd}},
  \code{\link{echosounder}}, \code{\link{lisst}},
  \code{\link{lobo}}, \code{\link{met}}, \code{\link{rsk}},
  \code{\link{sealevelTuktoyaktuk}},
  \code{\link{sealevel}}, \code{\link{section}},
  \code{\link{topoWorld}}, \code{\link{wind}}

Other things related to \code{landsat} data: \code{\link{[[,landsat-method}},
  \code{\link{landsat-class}}, \code{\link{landsatAdd}},
  \code{\link{landsatTrim}},
  \code{\link{plot,landsat-method}},
  \code{\link{read.landsat}},
  \code{\link{summary,landsat-method}}
}
