% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landsat.R
\name{landsatAdd}
\alias{landsatAdd}
\title{Add a Band to a landsat Object}
\usage{
landsatAdd(x, data, name, debug = getOption("oceDebug"))
}
\arguments{
\item{x}{A \code{landsat} object, e.g. as read by \code{\link{read.landsat}}.}

\item{data}{A matrix of data, with dimensions matching that of entries already in \code{x}.}

\item{name}{The name to be used for the data, i.e. the data can later be
accessed with \code{d[[name]]} where \code{d} is the name of the return value
from the present function.}

\item{debug}{A flag that turns on debugging.  Set to 1 to get a moderate amount of debugging
information, or a higher value for more debugging.}
}
\value{
An object of \code{\link{landsat-class}}, with a new data band.
}
\description{
Add a band to an object of \code{\link{landsat-class}}. Note that
it will be stored in numeric form, not raw form, and therefore
it will require much more storage than data read with
\code{\link{read.landsat}}.
}
\author{
Dan Kelley
}
\seealso{
The documentation for \code{\link{landsat-class}} explains the
structure of landsat objects, and also outlines the other functions dealing
with them.

Other functions.dealing.with.satellite.data: \code{\link{[[,amsr-method}};
  \code{\link{[[,landsat-method}};
  \code{\link{amsr-class}}; \code{\link{g1sst-class}};
  \code{\link{landsat-class}},
  \code{\link{satellite-class}},
  \code{\link{satellite-class}}; \code{\link{landsatTrim}};
  \code{\link{plot,amsr-method}}, \code{\link{plot.amsr}};
  \code{\link{plot,landsat-method}},
  \code{\link{plot.landsat}};
  \code{\link{plot,satellite-method}},
  \code{\link{plot.satellite}}; \code{\link{read.amsr}};
  \code{\link{read.g1sst}}; \code{\link{read.landsat}};
  \code{\link{summary,amsr-method}},
  \code{\link{summary.amsr}};
  \code{\link{summary,landsat-method}},
  \code{\link{summary.landsat}};
  \code{\link{summary,satellite-method}},
  \code{\link{summary.g1sst}},
  \code{\link{summary.satellite}}
}
\concept{
satellite
}

