\name{lat.format}

\alias{lat.format}

\title{Format a latitude}

\description{Format a latitude, using "S" for negative latitude.}

\usage{lat.label <- lat.format(lat)}

\arguments{
  \item{lat}{latitude in \eqn{^\circ}{deg}N north of the equator.}
}

\value{A character string.}

\seealso{\code{\link{lon.format}} and \code{\link{latlon.format}}.}

\author{Dan Kelley}

\keyword{misc}
