% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\docType{data}
\name{lisst}
\alias{lisst}
\title{LISST Dataset}
\source{
This was constructed artificially using \code{\link[=as.lisst]{as.lisst()}},
to approximately match values that might be measured in the field.
}
\usage{
data(lisst)
}
\description{
LISST (Laser in-situ scattering and transmissometry) dataset, constructed
artificially.
}
\seealso{
Other datasets provided with oce: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{echosounder}},
  \code{\link{landsat}}, \code{\link{lobo}},
  \code{\link{met}}, \code{\link{ocecolors}},
  \code{\link{rsk}}, \code{\link{sealevelTuktoyaktuk}},
  \code{\link{sealevel}}, \code{\link{section}},
  \code{\link{topoWorld}}, \code{\link{wind}},
  \code{\link{xbt}}
}
\author{
Dan Kelley
}
\concept{datasets provided with oce}
