\name{lisst}

\docType{data}

\alias{lisst}

\title{LISST dataset}

\description{LISST (Laser in-situ scattering and transmissometry) dataset,
    constructed artificially.}

\examples{
\dontrun{
library(oce)
data(lisst)
plot(lisst)
}
}

\usage{data(lisst)}

\source{This was constructed artifically, to roughly match the sort of values
    that might be measured in the field.}

\seealso{The documentation for \code{\link{lisst-class}} explains the structure
    of \code{lisst} objects, and also outlines the other functions dealing with
    them.}

\author{Dan Kelley}

\keyword{datasets}
