% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobo.R
\docType{class}
\name{lobo-class}
\alias{lobo-class}
\title{Class to Store LOBO Data

Class to store LOBO data.
A \code{lobo} object may be read with \code{\link{read.lobo}} or
constructed with \code{\link{as.lobo}}.  Plots can be made with
\code{\link{plot,lobo-method}}, while \code{\link{summary,lobo-method}} produces
statistical summaries. Data within a \code{lobo} object may be retrieved with
\code{\link{[[,lobo-method}} and altered with \code{\link{[[,lobo-method}}.}
\author{
Dan Kelley
}
\seealso{
Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{coastline-class}}, \code{\link{ctd-class}},
  \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{met-class}},
  \code{\link{odf-class}}, \code{\link{rsk-class}},
  \code{\link{sealevel-class}},
  \code{\link{section-class}}, \code{\link{topo-class}},
  \code{\link{windrose-class}}

Other things related to \code{lobo} data: \code{\link{[[,lobo-method}},
  \code{\link{[[<-,lobo-method}}, \code{\link{as.lobo}},
  \code{\link{lobo}}, \code{\link{plot,lobo-method}},
  \code{\link{subset,lobo-method}},
  \code{\link{summary,lobo-method}}
}

