\name{lobo}
\docType{data}
\alias{lobo}
\title{Sample lobo dataset}
\description{This is sample lobo dataset obtained in the Northwest Arm
of Halifax by Satlantic.}
\usage{data(ctd)}
\format{
}
\source{Satlantic.} 
\references{
  \url{http://www.satlantic.com/default.asp?mn=1.15.27.139}.
  \url{http://www.mbari.org/lobo/}.
}
\author{Dan Kelley}
\keyword{misc}
