\name{logger}
\alias{logger}

\docType{data}

\title{logger data, of temperature-data recorder variety}

\description{A sample object of Oce class \code{logger-class}.}

\usage{data(logger)}

\source{The \acronym{SLEIWEX} experiment.}

\references{\url{http://myweb.dal.ca/kelley/SLEIWEX/index.php}}

\seealso{The documentation for \code{logger-class} in the Oce package explains the
    structure of \code{logger} objects, and also outlines the other functions
    dealing with them.}

\examples{
\dontrun{
library(oce)
data(logger)
plot(logger)
}
}

\author{Dan Kelley}

\keyword{datasets}
