\name{lon.format}
\alias{lon.format}
\title{Format a longitude}
\description{
Format a longitude, using "W" for west longitude.
}
\usage{
 lon.label <- lon.format(lon)
}
\arguments{
  \item{lon}{longitude in \eqn{^\circ}{deg}N east of Greenwich.}
}
\details{
}
\value{
  A character string.
}

\seealso{
\code{\link{lat.format}} and \code{\link{latlon.format}}.
}
\examples{
}
\author{Dan Kelley}
\keyword{misc}
