% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lon360}
\alias{lon360}
\title{Alter longitudes from -180:180 to 0:360 convention}
\usage{
lon360(x)
}
\arguments{
\item{x}{either a numeric vector or array, or an \linkS4class{oce} object.}
}
\description{
For numerical input, including vectors, matrices and arrays,
this simply calls \code{\link[=ifelse]{ifelse()}} to add 360 to any negative values. For
\linkS4class{section} objects, it changes \code{longitude} in the \code{metadata} slot
and also in the \code{metadata} slots of the stations within the section.
For other \linkS4class{oce} objects, this changes any \code{longitude} values
that are present in both the \code{metadata} and \code{data} slots.
}
\details{
This function can be helpful in dealing with \linkS4class{section} data
that cross the dateline (i.e. longitude -180).  However, it is
certainly not a useful method for dealing with the dateline problem
in maps, because altering longitudes does not help with filled
polygons; see \code{\link[=coastlineCut]{coastlineCut()}} for an attempt to solve the
dateline problem.
}
\examples{
lon360(c(179, -179))
}
