% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{makeSection}
\alias{makeSection}
\title{Make a section (DEFUNCT)}
\usage{
makeSection(item, ...)
}
\arguments{
\item{item}{Ignored, since this function is defunct}

\item{...}{Ignored, since this function is defunct}
}
\description{
This is a defunct function; use \code{\link{as.section}} instead, and
see \link{oce-defunct} for more on the oce procedure for retiring functions.
}

