% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapMeridians}
\alias{mapMeridians}
\title{Add Meridians on a Map (defunct)}
\usage{
mapMeridians(latitude, lty = "solid", lwd = 0.5 * par("lwd"),
  col = "darkgray", ...)
}
\arguments{
\item{latitude}{either a logical value indicating whether to draw
a meridian grid, or a vector of latitudes at which to draw meridians.}

\item{lty}{line type.}

\item{lwd}{line width.}

\item{col}{line color.}

\item{...}{optional arguments passed to \code{\link[=lines]{lines()}}.}
}
\description{
\strong{WARNING:} This function will be removed soon; see \link{oce-deprecated}.
Use \code{\link[=mapGrid]{mapGrid()}} instead.
}
\seealso{
Other functions that will be removed soon: \code{\link{addColumn}},
  \code{\link{ctdAddColumn}},
  \code{\link{ctdUpdateHeader}},
  \code{\link{findInOrdered}}, \code{\link{mapZones}}
}
\author{
Dan Kelley
}
\concept{functions that will be removed soon}
