% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapZones}
\alias{mapZones}
\title{Add Zones to a Map [deprecated]}
\usage{
mapZones(longitude, polarCircle = 0, lty = "solid", lwd = 0.5 *
  par("lwd"), col = "darkgray", ...)
}
\arguments{
\item{longitude}{either a logical indicating whether to draw a zonal grid,
or a vector of longitudes at which to draw zones.}

\item{polarCircle}{a number indicating the number of degrees of latitude
extending from the poles, within which zones are not drawn.}

\item{lty}{line type.}

\item{lwd}{line width.}

\item{col}{line colour.}

\item{...}{optional arguments passed to \code{\link{lines}}.}
}
\description{
\strong{WARNING:} This function will be removed soon; see \link{oce-deprecated}.
}
\details{
Use \code{\link{mapGrid}} instead of the present function.

Plot zones (lines of constant longitude) on a existing map.


This function should not be used, since it will be removed soon.
Please use mapGrid() instead.
}
\author{
Dan Kelley
}
\seealso{
Other functions that will be removed soon: \code{\link{addColumn}},
  \code{\link{ctdAddColumn}},
  \code{\link{ctdUpdateHeader}}, \code{\link{mapMeridians}}
}

