% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapZones}
\alias{mapZones}
\title{Add Zones to a Map (defunct)}
\usage{
mapZones(longitude, polarCircle = 0, lty = "solid", lwd = 0.5 *
  par("lwd"), col = "darkgray", ...)
}
\arguments{
\item{longitude}{either a logical indicating whether to draw a zonal grid,
or a vector of longitudes at which to draw zones.}

\item{polarCircle}{a number indicating the number of degrees of latitude
extending from the poles, within which zones are not drawn.}

\item{lty}{line type.}

\item{lwd}{line width.}

\item{col}{line color.}

\item{...}{optional arguments passed to \code{\link[=lines]{lines()}}.}
}
\description{
\strong{WARNING:} This function will be removed soon; see \link{oce-deprecated}.
Use \code{\link[=mapGrid]{mapGrid()}} instead.
}
\seealso{
Other functions that will be removed soon: \code{\link{addColumn}},
  \code{\link{ctdAddColumn}},
  \code{\link{ctdUpdateHeader}},
  \code{\link{findInOrdered}}, \code{\link{mapMeridians}}
}
\author{
Dan Kelley
}
\concept{functions that will be removed soon}
