\name{met}

\docType{data}

\alias{met}

\title{Sample meteorological object}

\description{This is sample \code{met} object, provided for testing, created by
    the code given in \dQuote{Examples}.}

\usage{data(met)}

\examples{
\dontrun{
library(oce)
data(met)
plot(met)
}
}

\source{Downloaded from the Environment Canada website
    \url{http://climate.weatheroffice.gc.ca/climateData/bulkdata_e.html?timeframe=1&Prov=XX&StationID=6358&Year=2011&Month=7&Day=15&format=csv}
    on November 17, 2011.}

\seealso{The documentation for \code{met-class} in the Oce package explains the
    structure of \code{met} objects, and also outlines the other functions
    dealing with them.}

\keyword{datasets}
