% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\docType{data}
\name{met}
\alias{met}
\title{Sample meteorological object}
\source{
Downloaded from the Environment Canada website on May 29, 2014, and
processed as follows.
\preformatted{
met <- read.met('http://climate.weather.gc.ca/climateData/bulkdata_e.html?format=csv&stationID=6358&Year=2003&Month=9&Day=17&timeframe=1&submit=Download+Data')
met[['time']] <- met[['time']] + 4 * 3600
}
(Note the conversion from local standard time to UTC.)
}
\description{
This is sample \code{met} object containing data for Halifax, Nova Scotia,
during September of 2003 (the period during which Hurricane Juan struck the
city).
}
\seealso{
Other datasets provided with \code{oce}: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{echosounder}},
  \code{\link{lisst}}, \code{\link{lobo}},
  \code{\link{section}}, \code{\link{topoWorld}}
}

