% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{new,ctd-method}
\alias{new,ctd-method}
\title{Initialize storage for a ctd object}
\usage{
new("ctd", pressure, salinity, temperature, conductivity, units, 
   pressureType, deploymentType)
}
\arguments{
\item{pressure}{optional numerical vector of pressures.}

\item{salinity}{optional numerical vector of salinities.}

\item{temperature}{optional numerical vector of temperatures.}

\item{conductivity}{optional numerical vector of conductivities.}

\item{units}{optional list indicating units for the quantities specified
in the previous arguments. If this
is not supplied, a default is set up, based on which of the
\code{pressure} to \code{conductivity} arguments were specified.
If all of those 4 arguments were specified, then \code{units} is set
up as if the call included the following:
\code{units=list(temperature=list(unit=expression(degree*C), scale="ITS-90"),
     salinity=list(unit=expression(), scale="PSS-78"),
     conductivity=list(unit=expression(), scale=""),
     pressure=list(unit=expression(dbar), scale=""),
     depth=list(unit=expression(m), scale=""))}. This list is trimmed
of any of the 4 items that were not specified in the previous
arguments.}

\item{pressureType}{optional character string indicating the type of pressure;
if not supplied, this defaults to \code{"sea"}, which indicates the excess of
pressure over the atmospheric value, in dbar.}

\item{deploymentType}{optional character string indicating the type of deployment, which may
be \code{"unknown"}, \code{"profile"}, \code{"towyo"}, or \code{"thermosalinograph"}.
If this is not set, the value defaults to \code{"unknown"}.}
}
\description{
This function creates \code{oce} objects of \code{\link{ctd-class}}. It is mainly
used by \code{oce} functions such as \code{\link{read.ctd}} and \code{\link{as.ctd}},
and it is not intended for novice users, so it may change at any time, without
following the usual rules for transitioning to deprecated and defunct status
(see \link{oce-deprecated}).
}
\details{
To save storage, this function has arguments only for quantities that are often present in data
files all cases. For example, not
all data files will have oxygen, so that's not present here.
Extra data may be added after the object is created, using
\code{\link{oceSetData}}.
Similarly, \code{\link{oceSetMetadata}} may be used to add metadata (station ID, etc),
while bearing in mind that other functions look for such information
in very particular places (e.g. the station ID is a string named \code{station}
within the \code{metadata} slot). See \code{\link{ctd-class}} for more information
on elements stored in \code{ctd} objects.
}
\seealso{
Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{cnvName2oceName}}, \code{\link{ctd-class}},
  \code{\link{ctdAddColumn}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdRaw}},
  \code{\link{ctdTrim}}, \code{\link{ctdUpdateHeader}},
  \code{\link{ctd}}, \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}}, \code{\link{plotProfile}},
  \code{\link{plotScan}}, \code{\link{plotTS}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}},
  \code{\link{woceNames2oceNames}}, \code{\link{write.ctd}}
}

