% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{NEWhandleFlags,ctd-method}
\alias{NEWhandleFlags,ctd-method}
\title{Handle flags in CTD objects}
\usage{
\S4method{NEWhandleFlags}{ctd}(object, flags = list(), actions = list())
}
\arguments{
\item{object}{An object of \code{\link{ctd-class}}.}

\item{flags}{An optional \code{\link{list}} that contains items with names matching the names of entries in the object's \code{data} slot. See \dQuote{Examples}.}

\item{actions}{An optional \code{\link{list}} that contains items with names that match those in the \code{flags} argument.  If this is not supplied, the action is to set flagged values to \code{NA}. See \dQuote{Examples}.}
}
\description{
\code{NEWhandleFlags} is a generic function that has specialized variants
that are used for different types of \code{oce} object.  Each is designed to help users
deal with data flags, by performing specified actions on data elements that are
flagged in particular ways.  A common case is to set flagged data to the R missing-value
code, \code{NA}.  The specialized forms of this function are set up to handle
flag codes that are typical for the target data type.
Specifying the \code{flags} and \code{actions} arguments permits finer-grained
control, as well as the handling of cases employing unusual conventions for the
values of flags.
}
\details{
If \code{flags} and \code{actions} are not provided, the
default is to use WHP (World Hydrographic Program) flags, in which the
value 2 indicates good data, and other values indicate either unchecked,
suspicious, or bad data [1]; any data not flagged as good are set
to \code{NA} in the returned value.
}
\section{Caution}{
 \code{NEWhandleFlags} is a new function as of March 2016, and it is very likely to change
through the Spring of 2016. Almost nothing works yet, and users should not be doing
any more than looking at the documentation and telling the developers whether the planned 
functioning seems reasonable.
The hope is to get trial code working for \code{ctd}
and \code{argo} data types by the end of April. After that, there will likely
be a month or more of testing with real-world
work, with possible changes to the user interface. After that, other types will be added,
needed.
}
\examples{
\dontrun{
# 1. Default: anything not flagged as 2 is set to NA, thus focussing on
# data flagged as 'good' in the World Hydrographic Program scheme.
ctd <- handleFlags(ctd)

# 2. A less restrictive case: only set to NA if flag=4, and only apply 
this action to the salinity, temperature, and pressure data.
ctd <- handleFlags(ctd, flags=list(salinity=4, temperature=4, pressure=4))

# 3. Add 0.002degC to temperatures flagged as bad (NOT WORKING YET)
addT <- function(T) T + 0.002
ctd <- handleFlags(ctd, flags=list(temperature=4), actions=list(temperature=addT))
}
}
\references{
1. \url{https://www.nodc.noaa.gov/woce/woce_v3/wocedata_1/whp/exchange/exchange_format_desc.htm}
}

