% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{NEWhandleFlags,ctd-method}
\alias{NEWhandleFlags,ctd-method}
\title{Handle flags in CTD objects.}
\usage{
\S4method{NEWhandleFlags}{ctd}(object, action = "NA", flags = list())
}
\arguments{
\item{object}{An object inheriting from `ctd`.}

\item{action}{A character string indicating the action to be done. The choices are \code{"NA"}, which will cause any flagged items to be set to \code{NA}, and \code{"omit"}, which will cause deletion of the flagged data.}

\item{flags}{FIXME: explain how to set up flags, and also the default. Please put generalizable stuff into
man-roxygen/handleFlagsTemplate.R in order to avoid a lot of editing work later, when
we handle flags for over a dozen types.}
}
\description{
Handle flags in CTD objects.

FIXME: write general description here (in man-roxygen/handleFlagsTemplate.R)
}
\details{
FIXME: Put ctd-specific details here.
}

