% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oce.as.POSIXlt}
\alias{oce.as.POSIXlt}
\title{Oce Variant of as.POSIXlt [defunct]}
\usage{
oce.as.POSIXlt(x, tz = "")
}
\arguments{
\item{x}{a date, as for \code{as.POSIXlt}, but also including forms in which
the month name appears.}

\item{tz}{the timezone, as for \code{as.POSIXlt}}
}
\value{
A POSIXlt object.
}
\description{
\strong{WARNING:} This function will be removed soon; see \link{oce-deprecated}.
}
\details{
It was realized in December of 2016 that this function was not used within
oce, and also that \code{\link[lubridate]{parse_date_time}} in the
\CRANpkg{lubridate} package was superior and therefore a better choice for
\dQuote{oce} users.
}
\seealso{
Other functions that will be removed soon: \code{\link{addColumn}},
  \code{\link{ctdAddColumn}},
  \code{\link{ctdUpdateHeader}},
  \code{\link{findInOrdered}}, \code{\link{mapMeridians}},
  \code{\link{mapZones}}
}
\author{
Dan Kelley
}
