% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oce.colorsGebco}
\alias{oce.colors.gebco}
\alias{oce.colorsGebco}
\alias{oceColorsGebco}
\title{Gebco colors}
\usage{
oce.colorsGebco(n = 9, region = c("water", "land", "both"),
  type = c("fill", "line"))
}
\arguments{
\item{n}{Number of colors to return}

\item{region}{String indicating application region, one of \code{"water"}, \code{"land"},
or \code{"both"}.}

\item{type}{String indicating the purpose, one of \code{"fill"} or \code{"line"}.}
}
\description{
Gebco colors
}
\seealso{
Other things related to colors: \code{\link{colormap}},
  \code{\link{colors}}, \code{\link{oce.colorsTwo}}
}

