% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oceColorsChlorophyll}
\alias{oceColorsChlorophyll}
\alias{oce.colorsChlorophyll}
\alias{aliases}
\title{Create colours suitable for chlorophyll fields}
\usage{
oceColorsChlorophyll(n)
}
\arguments{
\item{n}{number of colours to create.}
}
\value{
A vector of colour specificications.
}
\description{
Create a set of colours for displaying chlorophyll values,
based on the scheme devised by Kristen M. Thyng
in her \code{cmcolor} Python package, which is available at
\url{https://github.com/kthyng/cmocean}.
}
\examples{
library(oce)
imagep(volcano, col=oceColorsChlorophyll(128),
       zlab="volcano dataset, autoscaled with oceColorsChlorophyll(128)")

}
\seealso{
Other things related to colors: \code{\link{colormap}},
  \code{\link{colors}}, \code{\link{oceColors9B}},
  \code{\link{oceColorsCDOM}},
  \code{\link{oceColorsDensity}},
  \code{\link{oceColorsFreesurface}},
  \code{\link{oceColorsGebco}}, \code{\link{oceColorsJet}},
  \code{\link{oceColorsOxygen}},
  \code{\link{oceColorsPAR}},
  \code{\link{oceColorsPalette}},
  \code{\link{oceColorsPhase}},
  \code{\link{oceColorsSalinity}},
  \code{\link{oceColorsTemperature}},
  \code{\link{oceColorsTurbidity}},
  \code{\link{oceColorsTwo}},
  \code{\link{oceColorsVelocity}},
  \code{\link{oceColorsViridis}},
  \code{\link{oceColorsVorticity}}
}
\author{
Krysten M. Thyng (Python version), Dan Kelley (R transliteration)
}
