% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oceColorsJet}
\alias{oceColorsJet}
\alias{oce.colorsJet}
\title{Create colours similar to the Matlab Jet scheme}
\usage{
oceColorsJet(n)
}
\arguments{
\item{n}{number of colours}
}
\description{
Create colours similar to the Matlab Jet scheme
}
\seealso{
Other things related to colors: \code{\link{colormap}},
  \code{\link{colors}}, \code{\link{oce.colorsGebco}},
  \code{\link{oceColors9A}}, \code{\link{oceColors9B}},
  \code{\link{oceColorsCDOM}},
  \code{\link{oceColorsChlorophyll}},
  \code{\link{oceColorsDensity}},
  \code{\link{oceColorsFreesurface}},
  \code{\link{oceColorsOxygen}},
  \code{\link{oceColorsPAR}},
  \code{\link{oceColorsPalette}},
  \code{\link{oceColorsPhase}},
  \code{\link{oceColorsSalinity}},
  \code{\link{oceColorsTemperature}},
  \code{\link{oceColorsTurbidity}},
  \code{\link{oceColorsTwo}},
  \code{\link{oceColorsVelocity}},
  \code{\link{oceColorsViridis}},
  \code{\link{oceColorsVorticity}}
}
