% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oceColorsJet}
\alias{oceColorsJet}
\alias{oce.colorsJet}
\alias{oceColors9A}
\alias{oce.colors9A}
\title{Create colors similar to the Matlab Jet scheme}
\usage{
oceColorsJet(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
Create colors similar to the Matlab Jet scheme
}
\examples{
library(oce)
imagep(volcano, col=oceColorsJet(128),
       zlab="oceColorsJet")
}
\seealso{
Other things related to colors: 
\code{\link{colormap}()},
\code{\link{oceColors9B}()},
\code{\link{oceColorsCDOM}()},
\code{\link{oceColorsChlorophyll}()},
\code{\link{oceColorsClosure}()},
\code{\link{oceColorsDensity}()},
\code{\link{oceColorsFreesurface}()},
\code{\link{oceColorsGebco}()},
\code{\link{oceColorsOxygen}()},
\code{\link{oceColorsPAR}()},
\code{\link{oceColorsPalette}()},
\code{\link{oceColorsPhase}()},
\code{\link{oceColorsSalinity}()},
\code{\link{oceColorsTemperature}()},
\code{\link{oceColorsTurbidity}()},
\code{\link{oceColorsTurbo}()},
\code{\link{oceColorsTwo}()},
\code{\link{oceColorsVelocity}()},
\code{\link{oceColorsViridis}()},
\code{\link{oceColorsVorticity}()},
\code{\link{ocecolors}}
}
\concept{things related to colors}
