% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oceColorsSalinity}
\alias{oceColorsSalinity}
\alias{oce.colorsSalinity}
\alias{aliases}
\title{Create colours suitable for salinity fields}
\usage{
oceColorsSalinity(n)
}
\arguments{
\item{n}{number of colours to create.}
}
\value{
A vector of colour specificications.
}
\description{
Create a set of colours for displaying salinity values,
based on the scheme devised by Kristen M. Thyng
in her \code{cmcolor} Python package, which is available at
\url{https://github.com/kthyng/cmocean}.
}
\seealso{
Other things related to colors: \code{\link{colormap}},
  \code{\link{colors}}, \code{\link{oce.colorsGebco}},
  \code{\link{oceColorsCDOM}},
  \code{\link{oceColorsChlorophyll}},
  \code{\link{oceColorsDensity}},
  \code{\link{oceColorsFreesurface}},
  \code{\link{oceColorsJet}},
  \code{\link{oceColorsOxygen}},
  \code{\link{oceColorsPAR}},
  \code{\link{oceColorsPalette}},
  \code{\link{oceColorsPhase}},
  \code{\link{oceColorsTemperature}},
  \code{\link{oceColorsTurbidity}},
  \code{\link{oceColorsTwo}},
  \code{\link{oceColorsVelocity}},
  \code{\link{oceColorsViridis}},
  \code{\link{oceColorsVorticity}}
}
\author{
Krysten M. Thyng (Python version), Dan Kelley (R transliteration)
}
