% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{oceColorsViridis}
\alias{oceColorsViridis}
\alias{oce.colorsViridis}
\title{Create colours similar to the matlab Viridis scheme}
\usage{
oceColorsViridis(n)
}
\arguments{
\item{n}{number of colours to create.}
}
\seealso{
Other things related to colors: \code{\link{colormap}},
  \code{\link{colors}}, \code{\link{oce.colorsGebco}},
  \code{\link{oceColorsCDOM}},
  \code{\link{oceColorsChlorophyll}},
  \code{\link{oceColorsDensity}},
  \code{\link{oceColorsFreesurface}},
  \code{\link{oceColorsJet}},
  \code{\link{oceColorsOxygen}},
  \code{\link{oceColorsPAR}},
  \code{\link{oceColorsPalette}},
  \code{\link{oceColorsPhase}},
  \code{\link{oceColorsSalinity}},
  \code{\link{oceColorsTemperature}},
  \code{\link{oceColorsTurbidity}},
  \code{\link{oceColorsTwo}},
  \code{\link{oceColorsVelocity}},
  \code{\link{oceColorsVorticity}}
}
\author{
Dan Kelley
}
