% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{oceGetMetadata}
\alias{oceGetMetadata}
\title{Get Something From the metadata Slot in an oce Object}
\usage{
oceGetMetadata(object, name)
}
\arguments{
\item{object}{an \code{oce} object}

\item{name}{String indicating the name of the item to be found.}
}
\description{
In contrast to the various \code{[[} functions, this is
guaranteed to look only within the \code{metadata} slot. If
the named item is not found, \code{NULL} is returned.
}
\seealso{
Other things related to the metadata slot: 
\code{\link{oceDeleteMetadata}()},
\code{\link{oceRenameMetadata}()},
\code{\link{oceSetMetadata}()}
}
\author{
Dan Kelley
}
\concept{things related to the metadata slot}
