% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{oceProject}
\alias{oceProject}
\title{Wrapper to rgdal::project()}
\usage{
oceProject(xy, proj, inv = FALSE, use_ob_tran = FALSE, legacy = TRUE)
}
\arguments{
\item{xy}{As for the \code{\link[rgdal]{project}} function in the
\CRANpkg{rgdal} package.}

\item{proj}{"}

\item{inv}{"}

\item{use_ob_tran}{"}

\item{legacy}{"}
}
\value{
A two-column matrix, with first column holding either
\code{longitude} or \code{x}, and second column holding either
\code{latitude} or \code{y}.
}
\description{
This function is used to isolate other oce functions from
changes to the \code{\link[rgdal]{project}} function in the \CRANpkg{rgdal}
package, which calculations of both forward
and inverse map projections.
}
\details{
Some highlights of the evolving relationship with rgdal are:
\enumerate{
\item See https://github.com/dankelley/oce/issues/653#issuecomment-107040093
   for the reason why oce switched from using \code{\link[rgdal]{rawTransform}},
   to \code{\link[rgdal]{project}}, both functions provided by the
   \CRANpkg{rgdal} package.
\item 2016 Apr: rgdal::project started returning named quantities
\item 2019 Feb: allowNAs_if_not_legacy was added in rgdal 1.3-9 to prevent
   an error on i386/windows.
}
}
