% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{oceSetMetadata}
\alias{oceSetMetadata}
\title{Set something in the \code{metadata} slot of an \code{oce} object}
\usage{
oceSetMetadata(object, name, value, note = "")
}
\arguments{
\item{object}{an \code{oce} object}

\item{name}{String indicating the name of the item to be set.}

\item{value}{Value for the item.}

\item{note}{A note to be stored in the processing log.}
}
\description{
Set something in the \code{metadata} slot of an \code{oce} object
}
