% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{oceSmooth}
\alias{oceSmooth}
\alias{oce.smooth}
\title{Smooth an Oce Object}
\usage{
oceSmooth(x, ...)
}
\arguments{
\item{x}{an \code{oce} object.}

\item{\dots}{parameters to be supplied to \code{\link[=smooth]{smooth()}}, which does
the actual work.}
}
\value{
An object of \code{\link[base]{class}} \code{"oce"} that has been
smoothed appropriately.
}
\description{
Each data element is smoothed as a timeseries. For ADP data, this is done
along time, not distance.  Time vectors, if any, are not smoothed.  A good
use of \code{oce.smooth} is for despiking noisy data.
}
\examples{
library(oce)
data(ctd)
d <- oce.smooth(ctd)
plot(d)

}
\seealso{
The work is done with \code{\link[=smooth]{smooth()}}, and the \code{...}
arguments are handed to it directly by \code{oce.smooth}.
}
\author{
Dan Kelley
}
