% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\docType{class}
\name{odf-class}
\alias{odf-class}
\alias{.odf}
\title{Class to Store ODF data}
\description{
Class for data stored in a format used at Canadian Department of Fisheries
and Oceans laboratories. This is somewhat unusual amongst \code{oce}
classes, in that it does not map to a particular instrument, but rather to a
storage type; in that sense, it is similar to the \code{\link{bremen-class}}.
}
\section{Slots}{

\describe{
\item{\code{data}}{In this and all \code{oce} objects, the \code{data} slot is a
\code{\link{list}} containing the main data for the object.}

\item{\code{metadata}}{In this and all \code{oce} objects, the \code{metadata} slot is
a \code{\link{list}} containing information about the \code{data} or about the
object itself.  An example of the former might be the location at which a
measurement was made, and of the latter might be the name of the experiment in
which the data were collected.}

\item{\code{processingLog}}{In this and all \code{oce} objects, the \code{processingLog}
slot is a \code{\link{list}} with entries describing the creation and evolution
of the object. This slot is updated by various \code{oce} functions.  The
contents are listed in object summaries and with
\code{\link{processingLogShow}}.}
}}

\section{Retrieving slot contents}{


The contents of the \code{data} and \code{metadata} slots of a \code{oce}
object named \code{obj} may be retrieved with the standard R system, using
\code{slot(obj, "data")} and \code{slot(obj, "metadata")}.  In addition, the
\code{[[} operator is overloaded for \code{oce} objects, as a convenience.

If \code{"obj"} is an \code{oce} object, then \code{obj[["data"]]} retrieves
the \code{data} slot, and \code{obj[["metadata"]]} retrieves the \code{metadata}
slot.  Each of these is a list, so the items within them may be retrieved in
the usual way, e.g. either \code{ctd[["data"]]$salinity} or
\code{ctd[["data"]]["salinity"]} may be used to recover the salinities in a
\code{ctd} object. Similar actions apply to the \code{metadata} slot.

Individual items within \code{oce} objects can be retrieved in two ways.  The
first is to use \code{\link{oceGetData}} and \code{\link{oceGetMetadata}},
respectively.  The second is to use the \code{[[} operator.

Items within the slots may be retrieved by name alone, according to a somewhat
involved scheme. If \code{obj} is an \code{oce} object, then e.g.
\code{obj[["x"]]} will retrieve \code{x} from the \code{metadata} slot if it
exists there, or from the \code{data} slot, if it exists there. This order of
precedence seldom causes problems, because it is rare to have an
identically-named item in these two slots.
}

\section{Modifying slot contents}{


Although there is a \code{[[<-} notation for modifying the contents
of \code{oce} objects, it is wiser to use \code{\link{oceSetData}}
\code{\link{oceSetMetadata}}, respectively.
}

\references{
[1] Anthony W. Isenor and David Kellow, 2011. ODF Format Specification
Version 2.0. (This is a .doc file downloaded from a now-forgotten URL by Dan Kelley,
in June 2011.)

[2] The St Lawrence Global Observatory website has information on ODF format at
\url{https://slgo.ca/app-sgdo/en/docs_reference/format_odf.html}

[3] List of variable codes:
\url{https://slgo.ca/app-sgdo/en/docs_reference/code_parametre_odf.html}
(checked 2018-02-11); only a subset are handled.
}
\seealso{
Other things related to \code{odf} data: \code{\link{ODF2oce}},
  \code{\link{ODFNames2oceNames}},
  \code{\link{[[,odf-method}},
  \code{\link{[[<-,odf-method}},
  \code{\link{plot,odf-method}},
  \code{\link{read.ctd.odf}}, \code{\link{read.odf}},
  \code{\link{subset,odf-method}},
  \code{\link{summary,odf-method}}

Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{coastline-class}}, \code{\link{ctd-class}},
  \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{met-class}}, \code{\link{oce-class}},
  \code{\link{rsk-class}}, \code{\link{sealevel-class}},
  \code{\link{section-class}}, \code{\link{topo-class}},
  \code{\link{windrose-class}}
}
\author{
Dan Kelley
}
