% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odf.R
\name{ODFNames2oceNames}
\alias{ODFNames2oceNames}
\title{Translate from ODF names to oce names}
\usage{
ODFNames2oceNames(names)
}
\arguments{
\item{names}{Data names in ODF format.}
}
\value{
A vector of string strings.
}
\description{
Translate from ODF names to oce names
}
\details{
The following table gives the regular expressions that define recognized
ODF names, along with the translated names as used in oce objects. Note
that if an item is repeated, then the second one has a \code{2} appended
at the end, etc.
\tabular{lll}{
    \strong{Regexp} \tab \strong{Result}           \tab \strong{Notes}                                             \cr
    \code{BEAM_*.*} \tab \code{a}                  \tab Used in \code{adp} objects                                 \cr
    \code{CNTR_*.*} \tab \code{scan}               \tab Used in \code{ctd} objects                                 \cr
    \code{CRAT_*.*} \tab \code{conductivity}       \tab Used mainly in \code{ctd} objects                          \cr
    \code{DEPH_*.*} \tab \code{depth}              \tab Not used by any plotting code                              \cr
    \code{DOXY_*.*} \tab \code{oxygen_by_volume}   \tab Used mainly in \code{ctd} objects                          \cr
    \code{ERRV_*.*} \tab \code{error}              \tab Used in \code{adp} objects                                 \cr
    \code{EWCT_*.*} \tab \code{u}                  \tab Used in \code{adp} and \code{cm} objects                   \cr
    \code{FFFF_*.*} \tab \code{flag}               \tab Used in many objects                                       \cr
    \code{FLOR_*.*} \tab \code{fluorometer}        \tab Used mainly in \code{ctd} objects                          \cr
    \code{FWETLABS} \tab \code{fwetlabs}           \tab Used in ??                                                 \cr
    \code{NSCT_*.*} \tab \code{v}                  \tab Used in \code{adp} objects                                 \cr
    \code{OCUR_*.*} \tab \code{oxygen_by_mole}     \tab Used mainly in \code{ctd} objects                          \cr
    \code{OTMP_*.*} \tab \code{oxygen_temperature} \tab Used mainly in \code{ctd} objects                          \cr
    \code{POTM_*.*} \tab \code{theta}              \tab Used mainly in \code{ctd} objects                          \cr
    \code{PRES_*.*} \tab \code{pressure}           \tab Used mainly in \code{ctd} objects                          \cr
    \code{PSAL_*.*} \tab \code{salinity}           \tab Used mainly in \code{ctd} objects                          \cr
    \code{PSAR_*.*} \tab \code{par}                \tab Used mainly in \code{ctd} objects                          \cr
    \code{SIGP_*.*} \tab \code{sigmaTheta}         \tab Used mainly in \code{ctd} objecs                           \cr
    \code{SIGT_*.*} \tab \code{sigmat}             \tab Used mainly in \code{ctd} objects                          \cr
    \code{SYTM_*.*} \tab \code{time}               \tab Used in many objects                                       \cr
    \code{TE90_*.*} \tab \code{temperature}        \tab Used mainly in \code{ctd} objects                          \cr
    \code{TEMP_*.*} \tab \code{temperature}        \tab Used mainly in \code{ctd} objects (FIXME: check if IPTS68) \cr
    \code{UNKN_*.*} \tab \code{g}                  \tab Used in ??                                                 \cr
    \code{VCSP_*.*} \tab \code{w}                  \tab Used in \code{adp} objects                                 \cr
}
}

