% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\name{plot,amsr-method}
\alias{plot,amsr-method}
\alias{plot.amsr}
\alias{both}
\alias{xlim}
\alias{and}
\alias{ylim}
\alias{to}
\alias{compute}
\alias{zrange}
\title{Plot an amsr Object}
\usage{
\S4method{plot}{amsr}(
  x,
  y,
  asp = NULL,
  breaks,
  col,
  colormap,
  missingColor = list(land = "papayaWhip", none = "lightGray", bad = "gray", rain =
    "plum", ice = "mediumVioletRed"),
  debug = getOption("oceDebug"),
  ...
)
}
\arguments{
\item{x}{an \linkS4class{amsr} object.}

\item{y}{character value indicating the name of the band to plot; if not provided,
\code{SST} is used; see the documentation for the \linkS4class{amsr} class for a list of bands.}

\item{asp}{optional numerical value giving the aspect ratio for plot.  The
default value, \code{NULL}, means to use an aspect ratio of 1 for world views,
and a value computed from \code{ylim}, if the latter is specified in the
\code{...} argument.}

\item{breaks}{optional numeric vector of the z values for breaks in the color scheme.
See also \code{colormap}, which takes precedence over \code{breaks}.}

\item{col}{optional argument, either a vector of colors corresponding to the breaks, of length
1 less than the number of breaks, or a function specifying colors, e.g.
\code{\link[=oce.colorsJet]{oce.colorsJet()}} for a rainbow.
See also \code{colormap}, which takes precedence over \code{col}.}

\item{colormap}{a specification of the colormap to use, as created
with \code{\link[=colormap]{colormap()}}.  If \code{colormap} is NULL, which is the default, then
a colormap is created to cover the range of data values, using
\link{oceColorsTemperature} colour scheme.
If \code{colormap} is provided, it takes precedence over \code{breaks} and \code{col}.
provided, it takes precedence over \code{breaks} and \code{col}.
See \dQuote{Examples} for an example of using the "turbo" colour scheme.}

\item{missingColor}{List of colors for problem cases. The names of the
elements in this list must be as in the default, but the colors may
be changed to any desired values. These default values work reasonably
well for SST images, which are the default image, and which employ a
blue-white-red blend of colors, no mixture of which matches the
default values in \code{missingColor}.}

\item{debug}{A debugging flag, integer.}

\item{...}{extra arguments passed to \code{\link[=imagep]{imagep()}}, e.g. to control
the view with \code{xlim} (for longitude) and \code{ylim} (for latitude).}
}
\description{
Plot an amsr Object
}
\examples{
\dontrun{
library(oce)
data(coastlineWorld)

# Example 1
year <- 2020
month <- 8
day <- 6:8
d1 <- read.amsr(download.amsr(year, month, day[1], "~/data/amsr"))
d2 <- read.amsr(download.amsr(year, month, day[2], "~/data/amsr"))
d3 <- read.amsr(download.amsr(year, month, day[3], "~/data/amsr"))
d <- composite(d1, d2, d3)
plot(d, "SST", xlim=c(-80,-10), ylim=c(20,65))
lines(coastlineWorld[['longitude']], coastlineWorld[['latitude']])

# Example 2: 'turbo' colour scheme
cm <- colormap(zlim=range(d[["SST"]], na.rm=TRUE), col=oceColorsTurbo)
plot(d, "SST", colormap=cm, xlim=c(-80,-10), ylim=c(20,65))
lines(coastlineWorld[['longitude']], coastlineWorld[['latitude']])
}

}
\seealso{
Other functions that plot oce data: 
\code{\link{plot,adp-method}},
\code{\link{plot,adv-method}},
\code{\link{plot,argo-method}},
\code{\link{plot,bremen-method}},
\code{\link{plot,cm-method}},
\code{\link{plot,coastline-method}},
\code{\link{plot,ctd-method}},
\code{\link{plot,gps-method}},
\code{\link{plot,ladp-method}},
\code{\link{plot,landsat-method}},
\code{\link{plot,lisst-method}},
\code{\link{plot,lobo-method}},
\code{\link{plot,met-method}},
\code{\link{plot,odf-method}},
\code{\link{plot,rsk-method}},
\code{\link{plot,satellite-method}},
\code{\link{plot,sealevel-method}},
\code{\link{plot,section-method}},
\code{\link{plot,tidem-method}},
\code{\link{plot,topo-method}},
\code{\link{plot,windrose-method}},
\code{\link{plot,xbt-method}},
\code{\link{plotProfile}()},
\code{\link{plotScan}()},
\code{\link{plotTS}()},
\code{\link{tidem-class}}

Other things related to amsr data: 
\code{\link{[[,amsr-method}},
\code{\link{[[<-,amsr-method}},
\code{\link{amsr-class}},
\code{\link{amsr}},
\code{\link{composite,amsr-method}},
\code{\link{download.amsr}()},
\code{\link{read.amsr}()},
\code{\link{subset,amsr-method}},
\code{\link{summary,amsr-method}}
}
\author{
Dan Kelley
}
\concept{functions that plot oce data}
\concept{satellite}
\concept{things related to amsr data}
