% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amsr.R
\docType{methods}
\name{plot,amsr-method}
\alias{plot,amsr-method}
\alias{plot.amsr}
\title{Plot an amsr object}
\usage{
\S4method{plot}{amsr}(x, y, col = oce.colorsPalette, missingColor = "gray",
  debug = getOption("oceDebug"), ...)
}
\arguments{
\item{x}{An object inherting from \code{\link{amsr-class}}.}

\item{y}{String indicating the name of the band to plot; if not provided, \code{SSTDay} is used..}

\item{col}{Color palette.}

\item{missingColor}{String indicating the colour to be used for missing data (land, clouds, etc)}

\item{debug}{A debugging flag, integer.}

\item{...}{Ignored.}
}
\description{
Plot an amsr object
}
\author{
Dan Kelley
}
\concept{
satellite
}

