% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\docType{methods}
\name{plot,argo-method}
\alias{plot,argo-method}
\title{Plot argo data}
\usage{
\S4method{plot}{argo}(x, which = 1, level, coastline = c("best",
  "coastlineWorld", "coastlineWorldMedium", "coastlineWorldFine", "none"),
  cex = 1, pch = 1, type = "p", col, fill = FALSE, adorn = NULL,
  projection = NULL, mgp = getOption("oceMgp"), mar = c(mgp[1] + 1.5,
  mgp[1] + 1.5, 1.5, 1.5), tformat, debug = getOption("oceDebug"), ...)
}
\arguments{
\item{x}{object inheriting from \code{\link{argo-class}}.}

\item{which}{list of desired plot types, one of the following.
\itemize{
    \item \code{which=1} or \code{which="trajectory"} gives a 
    plot of the argo trajectory, with the coastline, if one is provided.

    \item \code{which=2} or \code{"salinity ts"} gives a time series of
    salinity at the indicated level(s)

    \item \code{which=3} or \code{"temperature ts"} gives a time series
    of temperature at the indicated level(s)

    \item \code{which=4} or \code{"TS"} gives a TS diagram at the
    indicated level(s)

    \item \code{which=5} or \code{"salinity profile"} gives a salinity
    profile of all the data (with S and p trimmed to the 1 and 99
    percentiles)

    \item \code{which=6} or \code{"temperature profile"} gives a
    temperature profile (with T and p trimmed to the 1 and 99
    percentiles)
}}

\item{level}{depth pseudo-level to plot, for \code{which=2} and higher.  May be an
integer, in which case it refers to an index of depth (1 being the top)
or it may be the string "all" which means to plot all data.}

\item{coastline}{character string giving the coastline to be used in an Argo-location
map, or \code{"best"} to pick the one with highest resolution, or
\code{"none"} to avoid drawing the coastline.}

\item{cex}{size of plotting symbols to be used if \code{type='p'}.}

\item{pch}{type of plotting symbols to be used if \code{type='p'}.}

\item{type}{plot type, either \code{"l"} or \code{"p"}.}

\item{col}{optional list of colours for plotting.}

\item{fill}{Either a logical, indicating whether to fill the land with
light-gray, or a colour name.  Owing to problems with some projections, the
default is not to fill.}

\item{adorn}{list of expressions to be executed for the panels in turn, e.g. to
adorn the plots.  If the number matches the number of panels, then the strings
are applied to the appropriate panels, as they are drawn from top-left to
bottom-right.   If only a single expression is provided, it is used for all
panels. (See \dQuote{Examples}.)}

\item{projection}{indication of the projection to be used
in trajetory maps. If this is \code{NULL}, no projection is used, although
the plot aspect ratio will be set to yield zero shape distortion at the 
mean float latitude.  If \code{projection="automatic"}, then one
of two projections is used: stereopolar (i.e. \code{"+proj=stere +lon_0=X"}
where \code{X} is the mean longitude), or Mercator (i.e. \code{"+proj=merc"})
otherwise.  Otherwise, \code{projection} must be a character string specifying
a projection in the notation used by \link[rgdal]{project} in the \CRANpkg{rgdal};
this will be familiar to many readers as the PROJ.4 notation;
see \code{\link{mapPlot}}.}

\item{mgp}{3-element numerical vector to use for \code{par(mgp)}, and also for
\code{par(mar)}, computed from this.  The default is tighter than the R
default, in order to use more space for the data and less for the axes.}

\item{mar}{value to be used with \code{\link{par}}("mar").}

\item{tformat}{optional argument passed to \code{\link{oce.plot.ts}}, for plot
types that call that function.  (See \code{\link{strptime}} for the format
used.)}

\item{debug}{debugging flag.}

\item{...}{optional arguments passed to plotting functions.}
}
\value{
None.
}
\description{
Plot a summary diagram for argo data.
}
\examples{
library(oce)
data(argo)
plot(argo, which="trajectory")


}
\author{
Dan Kelley
}
\references{
\url{http://www.argo.ucsd.edu/}
}
\seealso{
Other functions that handle \code{argo} data: \code{\link{[[,argo-method}},
  \code{\link{[[<-,argo-method}}, \code{\link{argoGrid}},
  \code{\link{as.argo}}, \code{\link{read.argo}},
  \code{\link{subset,argo-method}},
  \code{\link{summary,argo-method}}

Other functions that plot \code{oce} data: \code{\link{plot,adv-method}},
  \code{\link{plot,amsr-method}},
  \code{\link{plot,bremen-method}},
  \code{\link{plot,cm-method}},
  \code{\link{plot,coastline-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{plot,gps-method}},
  \code{\link{plot,ladp-method}},
  \code{\link{plot,lisst-method}},
  \code{\link{plot,lobo-method}},
  \code{\link{plot,rsk-method}},
  \code{\link{plot,satellite-method}},
  \code{\link{plot,section-method}},
  \code{\link{plot,topo-method}},
  \code{\link{plot,windrose-method}},
  \code{\link{plotScan}}
}

