% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{plot,oce-method}
\alias{plot,oce-method}
\title{Plot an oce Object}
\usage{
\S4method{plot}{oce}(x, y, ...)
}
\arguments{
\item{x}{An oce object.}

\item{...}{Passed to \code{\link{pairs}}, which does the plotting.}
}
\description{
This creates a \code{\link{pairs}} plot of the elements in the \code{data}
slot.
}
\examples{
library(oce)
o <- new("oce")
o <- oceSetData(o, 'x', rnorm(10))
o <- oceSetData(o, 'y', rnorm(10))
o <- oceSetData(o, 'z', rnorm(10))
plot(o)
}

