% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf.R
\docType{methods}
\name{plot,odf-method}
\alias{plot,odf-method}
\title{Plot an ODF Object

Plot data contained within an ODF object.

Uses \code{\link{oce.plot.ts}} to create panels of time-series plots for all
the columns contained in the \code{odf} object.}
\usage{
\S4method{plot}{odf}(x)
}
\arguments{
\item{x}{A \code{odf} object, e.g. one inheriting from \code{\link{odf-class}}.}
}
\author{
Dan Kelley
}
\seealso{
Other functions that plot \code{oce} data: \code{\link{plot,adp-method}},
  \code{\link{plot,adv-method}},
  \code{\link{plot,amsr-method}},
  \code{\link{plot,argo-method}},
  \code{\link{plot,bremen-method}},
  \code{\link{plot,cm-method}},
  \code{\link{plot,coastline-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{plot,gps-method}},
  \code{\link{plot,ladp-method}},
  \code{\link{plot,lisst-method}},
  \code{\link{plot,lobo-method}},
  \code{\link{plot,met-method}},
  \code{\link{plot,rsk-method}},
  \code{\link{plot,satellite-method}},
  \code{\link{plot,sealevel-method}},
  \code{\link{plot,section-method}},
  \code{\link{plot,tidem-method}},
  \code{\link{plot,topo-method}},
  \code{\link{plot,windrose-method}},
  \code{\link{plotProfile}}, \code{\link{plotScan}},
  \code{\link{plotTS}}, \code{\link{tidem-class}}

Other things related to \code{odf} data: \code{\link{ODF2oce}},
  \code{\link{[[,odf-method}},
  \code{\link{[[<-,odf-method}}, \code{\link{odf-class}},
  \code{\link{subset,odf-method}},
  \code{\link{summary,odf-method}}
}

