% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satellite.R
\docType{methods}
\name{plot,satellite-method}
\alias{plot,satellite-method}
\alias{plot.satellite}
\title{Plot a satellite object}
\usage{
\S4method{plot}{satellite}(x, y, asp, debug = getOption("oceDebug"), ...)
}
\arguments{
\item{x}{An object inherting from \code{\link{satellite-class}}.}

\item{y}{String indicating the quantity to be plotted.}

\item{asp}{Optional aspect ratio for plot.}

\item{debug}{A debugging flag, integer.}

\item{...}{extra arguments passed to \code{\link{imagep}}, e.g. set
\code{col} to control colours.}
}
\description{
For an example, see \code{\link{read.g1sst}}.
}
\author{
Dan Kelley
}
\concept{
satellite
}

