% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\docType{methods}
\name{plot,windrose-method}
\alias{plot,windrose-method}
\title{Plot a wind rose diagram}
\usage{
\S4method{plot}{windrose}(x, type = c("count", "mean", "median", "fivenum"),
  convention = c("meteorological", "oceanographic"),
  mgp = getOption("oceMgp"), mar = c(mgp[1], mgp[1], 1 + mgp[1], mgp[1]),
  col, ...)
}
\arguments{
\item{x}{A \code{windrose} object, e.g. inheriting from \code{\link{windrose-class}}.}

\item{type}{The thing to be plotted, either the number of counts in the angle
interval, the mean of the values in the interval, the median of the values, or
a \code{\link{fivenum}} representation of the values.}

\item{convention}{String indicating whether to use meteorological convention or
oceanographic convention for the arrows that emanate from the centre of the
rose.  In meteorological convenction, an arrow emanates towards the right on
the diagram if the wind is from the east; in oceanographic convention, such an
arrow indicates flow \emph{to} the east.}

\item{mgp}{Three-element numerical vector to use for \code{par(mgp)}, and also
for \code{par(mar)}, computed from this.  The default is tighter than the R
default, in order to use more space for the data and less for the axes.}

\item{mar}{Four-element numerical vector to be used with \code{\link{par}("mar")}.}

\item{col}{Optional list of colours to use.  If not set, the colours will be
\code{c("red", "pink", "blue", "lightgray")}.  For the first three types of
plot, the first colour in this list is used to fill in the rose, the third is
used for the petals of the rose, and the fourth is used for grid lines. For the
\code{"fivenum"} type, the first colour is used for the inter-quartile range,
the second is used outside this range, the third is used for the median, and
the fourth is, again, used for the grid lines.}

\item{...}{Optional arguments passed to plotting functions.}
}
\description{
Plot a wind rose diagram
}
\examples{
library(oce)
opar <- par(no.readonly = TRUE)
xcomp <- rnorm(360) + 1
ycomp <- rnorm(360)
wr <- as.windrose(xcomp, ycomp)
par(mfrow=c(1,2))
plot(wr)
plot(wr, "fivenum")
par(opar)

}
\author{
Dan Kelley
}
\seealso{
Other functions that handle \code{windrose} data: \code{\link{[[,windrose-method}},
  \code{\link{[[<-,windrose-method}},
  \code{\link{as.windrose}},
  \code{\link{summary,windrose-method}}

Other functions that plot \code{oce} data: \code{\link{plot,adv-method}},
  \code{\link{plot,amsr-method}},
  \code{\link{plot,argo-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{plot,ladp-method}},
  \code{\link{plot,satellite-method}},
  \code{\link{plot,section-method}}
}

