% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingLog.R
\name{processingLog<-}
\alias{processingLog<-}
\title{Add an item to a processing log (in place)}
\usage{
processingLog(x) <- value
}
\arguments{
\item{x}{An \code{oce} object.}

\item{value}{A character string with the description of the logged activity.}
}
\description{
Add an item to a processing log (in place)
}
\examples{
data(ctd)
processingLogShow(ctd)
processingLog(ctd) <- "test"
processingLogShow(ctd)
}
\seealso{
Other things related to processing logs: \code{\link{processingLogAppend}},
  \code{\link{processingLogItem}},
  \code{\link{processingLogShow}}
}
\concept{things related to processing logs}
