% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingLog.R
\name{processingLogAppend}
\alias{processingLogAppend}
\title{Append an item to a processing log}
\usage{
processingLogAppend(h, value = "")
}
\arguments{
\item{h}{either the \code{processingLog} slot of an object, or
an \code{oce} object from which the processingLog will be extracted}

\item{value}{A string indicating the text of the log entry.}
}
\value{
An \code{\link{list}} containing items named
\code{time} and \code{value}, i.e. the times of entries
and the text notations of those entries..
}
\description{
Append an item to a processing log
}

