% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingLog.R
\name{processingLogItem}
\alias{processingLogItem}
\title{Create an item that can be inserted into a processing log}
\usage{
processingLogItem(value = "")
}
\arguments{
\item{value}{A string that will be used for the item.}
}
\value{
A \code{\link[=list]{list()}} containing \code{time}, which is
the time in UTC (calculated with \code{\link[=presentTime]{presentTime()}})
at the moment the function is called and
\code{value}, a string that is set to the argument of the same name.
}
\description{
A function is used internally to initialize processing logs.
Users will probably prefer to use \code{\link[=processingLogAppend]{processingLogAppend()}}
instead.
}
\seealso{
Other things related to processing logs: \code{\link{processingLog<-}},
  \code{\link{processingLogAppend}},
  \code{\link{processingLogShow}}
}
\concept{things related to processing logs}
