% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingLog.R
\name{processingLogShow}
\alias{processingLogShow}
\title{Show the processing log of an oce object}
\usage{
processingLogShow(x)
}
\arguments{
\item{x}{an \linkS4class{oce} object.}
}
\description{
Show the processing log of an oce object
}
\seealso{
Other things related to processing logs: \code{\link{processingLog<-}},
  \code{\link{processingLogAppend}},
  \code{\link{processingLogItem}}
}
\concept{things related to processing logs}
