\name{pt}
\alias{pt}

\docType{data}

\title{pt (pressure-temperature) dataset}

\description{A sample object of class \code{\link{pt-class}}.}

\usage{data(pt)}

\source{The \acronym{SLEIWEX} experiment.}

\references{\url{http://myweb.dal.ca/kelley/SLEIWEX/index.php}}

\seealso{The documentation for \code{\link{pt-class}} explains the structure of
    PT objects, and also outlines the other functions dealing with them.}

\examples{
library(oce)
data(pt)
plot(pt)
}
\author{Dan Kelley}
\keyword{datasets}
