% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rangeExtended}
\alias{rangeExtended}
\title{Calculate Range, Extended a Little, as is Done for Axes}
\usage{
rangeExtended(x, extend = 0.04)
}
\arguments{
\item{x}{a numeric vector.}

\item{extend}{fraction to extend on either end}
}
\value{
A two-element vector with the extended range of \code{x}.
}
\description{
This is analogous to what is done as part of the R axis range calculation,
in the case where \code{xaxs="r"}.
}
\author{
Dan Kelley
}

