% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.nortek.ad2cp.R
\name{read.adp.ad2cp}
\alias{read.adp.ad2cp}
\title{Read a Nortek AD2CP File}
\usage{
read.adp.ad2cp(
  file,
  from = 1,
  to = 0,
  by = 1,
  dataType = NULL,
  tz = getOption("oceTz"),
  ignoreChecksums = FALSE,
  longitude = NA,
  latitude = NA,
  orientation,
  distance,
  plan,
  type,
  monitor = FALSE,
  despike = FALSE,
  processingLog,
  debug = getOption("oceDebug"),
  ...
)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to
load.}

\item{from}{an integer indicating the index number of the first record to
read. This must equal 1, for this version of \code{read.adp.ad2cp}.  (If not
provided, \code{from} defaults to 1.)}

\item{to}{an integer indicating the final record to read. If \code{to} is 0L,
which is the default, then the value is changed internally to 1e9, and
reading stops at the end of the file.}

\item{by}{ignored.}

\item{dataType}{an indication of the data type to be extracted.  If This is NULL
(the default) then \code{read.adp.ad2cp()} will produce a table of permitted values.
Otherwise, it must be either a numeric or character value.  In the numeric case,
it is converted to an integer that indicating the data type via ID. Either a
decimal or a raw value is permitted, and the convention is as set in Nortek
manuals; see the table at the start of the \dQuote{Details} section.  In the
character case, it must be a string as indicated in that same table, or
the string \code{"TOC"} to produce a table of contents listing data types and
times.}

\item{tz}{a character value indicating time zone. This is used in
interpreting times stored in the file.}

\item{ignoreChecksums}{a logical value indicating whether to ignore
checksums.  This is FALSE by default, meaning that any data chunk with an
improper checksum is ignored. It may be necessary to set this to TRUE to
parse some problematic files.}

\item{longitude, latitude}{numerical values indicating the observation
location.}

\item{orientation}{ignored by \code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}}, and provided only for
similarity to other \verb{read.adp.*} functions.}

\item{distance}{ignored by \code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}}, and provided only for similarity
to other \verb{read.adp.*} functions.}

\item{plan}{optional integer specifying which 'plan' to focus on (see
reference 1 for the meaning of 'plan').  If this is not given, it defaults to
the most common plan in the requested subset of the data.}

\item{type}{optional character value indicating the type of Nortek
instrument.  If this is not provided, an attempt is made to infer it from the
file header (if there is one), and \code{"Signature1000"} is used, otherwise. The
importance of knowing the type is for inferring the slantwise beam angle,
which is usd in the conversion from beam coordinates to xyz coordinates. If
\code{type} is provided, it must be one of \code{"Signature250"}, \code{"Signature500"}, or
\code{"Signature1000"}; the first of these has a 20 degree slant-beam angle, while
the others each have 20 degrees (see reference 2, section 2 on page 6). Note
that \code{\link[=oceSetMetadata]{oceSetMetadata()}} can be used to alter the slantwise beam angle of an
existing object, and this will alter any later conversion from beam to xyz
coordinates.}

\item{monitor}{a boolean value indicating whether to indicate the progress of
reading the file, by using \code{\link[=txtProgressBar]{txtProgressBar()}} or otherwise.  The value of
\code{monitor} is changed to \code{FALSE} automatically, for non-interactive sessions.}

\item{despike}{ignored by \code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}}, and provided only for similarity
to other \verb{read.adp.*} functions.}

\item{processingLog}{a character value that, if provided, is saved within the
\code{processingLog} slot of the returned value.}

\item{debug}{an integer value indicating the level of debugging.  Set to 1 to
get a moderate amount of debugging information, from the R code only, to 2 to
get some debugging information from the C++ code that is used to parse the
data chunks, or to 3 for intensive debugging at both levels.}

\item{\dots}{ignored.}
}
\value{
An \linkS4class{adp} object with \code{metadata$fileType} equal to \code{"AD2CP"}, a
table (if \code{which="?"}), a data frame (if \code{which="??"}), or a vector of
character (if \code{which="text"}).
}
\description{
This function reads Nortek AD2CP files, storing data elements in lists within
the \code{data} slot.  Those elements are named for the ID type in question.  For
example, data with ID code \code{0x16} are stored in \code{d@data$average}, if \code{d} is
the value returned by this function. See \code{\link[=ad2cpCodeToName]{ad2cpCodeToName()}} for the mapping
between ID hexadecimal code and storage name.  By default, \code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}}
reads all ID codes that are in the file, but the \code{which} argument permits a
refined focus, yielding faster processing times and smaller returned objects.
}
\details{
This function is still in active development, so be aware that the returned
data structures might change in future versions.

The coding is based mainly on descriptions in various versions of a Nortek
manual (see \dQuote{References}). However, there are some gaps and
contradictions in these manuals, owing partly to evolution of the data
format. These things posed a challenge in the writing of \code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}}.
Thankfully, personnel in Nortek technical support team were able to supply
the help that was needed. Lacking this, \code{\link[=read.adp.ad2cp]{read.adp.ad2cp()}} would be
significantly limited, both in what it can read and in the accuracy
of the results.

Comments in the code, along with some warnings and messages that may be
issued during processing, are used to highlight some areas that may need
attention in revisions to this function.
}
\examples{
# You can run this within the oce directory, if you clone from github.
file <- "tests/testthat/local_data/ad2cp/S102791A002_Barrow_v2.ad2cp"
if (file.exists(file)) {
    library(oce)
    d <- read.oce(file)
}

}
\references{
Nortek AS. \dQuote{Signature Integration 55|250|500|1000kHz.} Nortek AS,
2017.

Nortek AS. \dQuote{Signature Integration 55|250|500|1000kHz.} Nortek AS,
2018.

Nortek AS. \dQuote{Signature Integration 55|250|500|1000kHz.} Nortek AS,
March 31, 2022.
}
\seealso{
Other things related to adp data: 
\code{\link{[[,adp-method}},
\code{\link{[[<-,adp-method}},
\code{\link{ad2cpCodeToName}()},
\code{\link{ad2cpHeaderValue}()},
\code{\link{adp-class}},
\code{\link{adpAd2cpFileTrim}()},
\code{\link{adpConvertRawToNumeric}()},
\code{\link{adpEnsembleAverage}()},
\code{\link{adpFlagPastBoundary}()},
\code{\link{adpRdiFileTrim}()},
\code{\link{adp_rdi.000}},
\code{\link{adp}},
\code{\link{as.adp}()},
\code{\link{beamName}()},
\code{\link{beamToXyzAdpAD2CP}()},
\code{\link{beamToXyzAdp}()},
\code{\link{beamToXyzAdv}()},
\code{\link{beamToXyz}()},
\code{\link{beamUnspreadAdp}()},
\code{\link{binmapAdp}()},
\code{\link{enuToOtherAdp}()},
\code{\link{enuToOther}()},
\code{\link{handleFlags,adp-method}},
\code{\link{is.ad2cp}()},
\code{\link{plot,adp-method}},
\code{\link{read.adp.nortek}()},
\code{\link{read.adp.rdi}()},
\code{\link{read.adp.sontek.serial}()},
\code{\link{read.adp.sontek}()},
\code{\link{read.adp}()},
\code{\link{read.aquadoppHR}()},
\code{\link{read.aquadoppProfiler}()},
\code{\link{read.aquadopp}()},
\code{\link{rotateAboutZ}()},
\code{\link{setFlags,adp-method}},
\code{\link{subset,adp-method}},
\code{\link{subtractBottomVelocity}()},
\code{\link{summary,adp-method}},
\code{\link{toEnuAdp}()},
\code{\link{toEnu}()},
\code{\link{velocityStatistics}()},
\code{\link{xyzToEnuAdpAD2CP}()},
\code{\link{xyzToEnuAdp}()},
\code{\link{xyzToEnu}()}

Other things related to ad2cp data: 
\code{\link{ad2cpCodeToName}()},
\code{\link{ad2cpHeaderValue}()},
\code{\link{adpAd2cpFileTrim}()},
\code{\link{is.ad2cp}()}

Other functions that read adp data: 
\code{\link{read.adp.nortek}()},
\code{\link{read.adp.rdi}()},
\code{\link{read.adp.sontek.serial}()},
\code{\link{read.adp.sontek}()},
\code{\link{read.adp}()},
\code{\link{read.aquadoppHR}()},
\code{\link{read.aquadoppProfiler}()},
\code{\link{read.aquadopp}()}
}
\author{
Dan Kelley
}
\concept{functions that read adp data}
\concept{things related to ad2cp data}
\concept{things related to adp data}
