% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.nortek.R
\name{read.adp.ad2cp}
\alias{read.adp.ad2cp}
\title{Read an AD2CP File}
\usage{
read.adp.ad2cp(file, from = 1, to, by = 1, tz = getOption("oceTz"),
  longitude = NA, latitude = NA, orientation, distance, plan, type,
  monitor = FALSE, despike = FALSE, processingLog,
  debug = getOption("oceDebug"), ...)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file
to load.  (For \code{read.adp.sontek.serial}, this is generally a list of
files, which will be concatenated.)}

\item{from}{indication of the first profile to read.  This can be an
integer, the sequence number of the first profile to read, or a POSIXt time
before which profiles should be skipped, or a character string that converts
to a POSIXt time (assuming UTC timezone).  See \dQuote{Examples}, and make
careful note of the use of the \code{tz} argument. If \code{from} is not
supplied, it defaults to 1.}

\item{to}{an optional indication of the last profile to read, in a
format as described for \code{from}.  As a special case, \code{to=0} means
to read the file to the end.  If \code{to} is not supplied, then it defaults
to 0.}

\item{by}{an optional indication of the stride length to use while walking through
the file.  If this is an integer, then \code{by-1} profiles are skipped
between each pair of profiles that is read, e.g. the default \code{by=1}
means to read all the data.  (For RDI files \emph{only}, there are some
extra features to avoid running out of memory; see \dQuote{Memory considerations}.)}

\item{tz}{character string indicating time zone to be assumed in the data.}

\item{longitude}{optional signed number indicating the longitude in degrees
East.}

\item{latitude}{optional signed number indicating the latitude in degrees
North.}

\item{orientation}{Ignored by this function, and provided only for similarity
to other adp-reading functions.}

\item{distance}{Ignored by this function, and provided only for similarity
to other adp-reading functions.}

\item{plan}{Optional integer specifying which 'plan' to focus on (see [1]
for the meaning of 'plan').  If this is not given, it defaults to the most
common plan in the requested subset of the data.}

\item{type}{Optional character value indicating the type of Nortek instrument.
If this is not provided, an attempt is made to infer it
from the file header (if there is one), and \code{"Signature1000"}
is used, otherwise. The importance
of knowing the type is for inferring the slantwise beam angle, which is usd in the
conversion from beam coordinates to xyz coordinates. If \code{type} is
provided, it must be one of \code{"Signature250"}, \code{"Signature500"},
or \code{"Signature1000"}; the first of these has a 20 degree
slant-beam angle, while the others each have 20 degrees (see [2],
section 2 on page 6). Note that \code{\link{oceSetMetadata}}
can be used to alter the slantwise beam angle of an existing object,
and this will alter any later conversion from beam to xyz coordinates.}

\item{monitor}{boolean, set to \code{TRUE} to provide an indication of progress
in reading the file, either by printing a dot for each profile or by writing
a textual progress bar with \code{\link{txtProgressBar}}.}

\item{despike}{Ignored by this function, and provided only for similarity
to other adp-reading functions.}

\item{processingLog}{if provided, the action item to be stored in the log.
(Typically only provided for internal calls; the default that it provides is
better for normal calls by a user.)}

\item{debug}{a flag that turns on debugging.  Set to 1 to get a moderate
amount of debugging information, or to 2 to get more.}

\item{\dots}{additional arguments, passed to called routines.}
}
\value{
An \code{adp} object, i.e. one inheriting from \code{\link{adp-class}}.
The form of this object varies with instrument type. In some cases
navigational data may be included, e.g. \code{read.adp.rdi} can
read NMEA strings (which get stored in an item called \code{nmea}
in the \code{data} slot).
}
\description{
This function is incomplete in several ways, and is still in active
development and testing, as of early 2019.
}
\examples{
\dontrun{
d <- read.adp.ad2cp("~/test.ad2cp", to=100) # or read.oce()
}

}
\references{
1. Nortek AS. “Signature Integration 55|250|500|1000kHz.” Nortek AS, 2017.

2. Nortek AS. “Operations Manual - Signature250, 500 and 1000.” Nortek AS, September 21, 2018.

3. Nortek AS. “Signature Integration 55|250|500|1000kHz.” Nortek AS, 2018. (This revision of
[1] is useful in including new information about instrument orientation. Note that
most of the comments within the \code{read.adp.ad2cp} code refer to [1], which has different
page numbers than [3].)
}
\seealso{
Other things related to \code{adp} data: \code{\link{[[,adp-method}},
  \code{\link{[[<-,adp-method}},
  \code{\link{ad2cpHeaderValue}}, \code{\link{adp-class}},
  \code{\link{adpEnsembleAverage}}, \code{\link{adp}},
  \code{\link{as.adp}}, \code{\link{beamName}},
  \code{\link{beamToXyzAdp}}, \code{\link{beamToXyzAdv}},
  \code{\link{beamToXyz}}, \code{\link{beamUnspreadAdp}},
  \code{\link{binmapAdp}}, \code{\link{enuToOtherAdp}},
  \code{\link{enuToOther}},
  \code{\link{handleFlags,adp-method}},
  \code{\link{is.ad2cp}}, \code{\link{plot,adp-method}},
  \code{\link{read.adp.nortek}},
  \code{\link{read.adp.rdi}},
  \code{\link{read.adp.sontek.serial}},
  \code{\link{read.adp.sontek}}, \code{\link{read.adp}},
  \code{\link{read.aquadoppHR}},
  \code{\link{read.aquadoppProfiler}},
  \code{\link{read.aquadopp}}, \code{\link{rotateAboutZ}},
  \code{\link{setFlags,adp-method}},
  \code{\link{subset,adp-method}},
  \code{\link{summary,adp-method}}, \code{\link{toEnuAdp}},
  \code{\link{toEnu}}, \code{\link{velocityStatistics}},
  \code{\link{xyzToEnuAdp}}, \code{\link{xyzToEnu}}
}
\author{
Dan Kelley
}
\concept{things related to \code{adp} data}
