% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.rdi.R
\name{read.adp.rdi}
\alias{read.adp.rdi}
\title{Read a Teledyne/RDI ADP File}
\usage{
read.adp.rdi(file, from, to, by, tz = getOption("oceTz"), longitude = NA,
  latitude = NA, type = c("workhorse"), monitor = FALSE,
  despike = FALSE, processingLog, testing = FALSE,
  debug = getOption("oceDebug"), ...)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file
to load.  (For \code{read.adp.sontek.serial}, this is generally a list of
files, which will be concatenated.)}

\item{from}{indication of the first profile to read.  This can be an
integer, the sequence number of the first profile to read, or a POSIXt time
before which profiles should be skipped, or a character string that converts
to a POSIXt time (assuming UTC timezone).  See \dQuote{Examples}, and make
careful note of the use of the \code{tz} argument. If \code{from} is not
supplied, it defaults to 1.}

\item{to}{an optional indication of the last profile to read, in a
format as described for \code{from}.  As a special case, \code{to=0} means
to read the file to the end.  If \code{to} is not supplied, then it defaults
to 0.}

\item{by}{an optional indication of the stride length to use while walking through
the file.  If this is an integer, then \code{by-1} profiles are skipped
between each pair of profiles that is read, e.g. the default \code{by=1}
means to read all the data.  (For RDI files \emph{only}, there are some
extra features to avoid running out of memory; see \dQuote{Memory considerations}.)}

\item{tz}{character string indicating time zone to be assumed in the data.}

\item{longitude}{optional signed number indicating the longitude in degrees
East.}

\item{latitude}{optional signed number indicating the latitude in degrees
North.}

\item{type}{character string indicating the type of instrument.}

\item{monitor}{boolean, set to \code{TRUE} to provide an indication (with
numbers and dots) of every profile read.}

\item{despike}{if \code{TRUE}, \code{\link{despike}} will be used to clean
anomalous spikes in heading, etc.}

\item{processingLog}{if provided, the action item to be stored in the log.
(Typically only provided for internal calls; the default that it provides is
better for normal calls by a user.)}

\item{testing}{logical value (IGNORED).}

\item{debug}{a flag that turns on debugging.  Set to 1 to get a moderate
amount of debugging information, or to 2 to get more.}

\item{\dots}{additional arguments, passed to called routines.}
}
\value{
An \code{adp} object, i.e. one inheriting from \code{\link{adp-class}}.
}
\description{
Read a Teledyne/RDI ADCP file (called 'adp' in oce).
}
\details{
As of 2016-09-25, this function has provisional functionality to
read data from the new "SentinelV" series ADCP -- essentially a
combination of a 4 beam workhorse with an additional vertical
centre beam.

If a heading bias had been set with the \code{EB} command during the setup
for the deployment, then a heading bias will have been stored in the file's
header.  This value is stored in the object's metadata as
\code{metadata$heading.bias}.  \strong{Importantly}, this value is
subtracted from the headings stored in the file, and the result of this
subtraction is stored in the objects heading value (in \code{data$heading}).
It should be noted that \code{read.adp.rdi()} was tested for firmware
version 16.30.  For other versions, there may be problems.  For example, the
serial number is not recognized properly for version 16.28.

In Teledyne/RDI ADP data files, velocities are coded to signed 2-byte integers, with a
scale factor being used to convert to velocity in metres per second.  These
two facts control the maximum recordable velocity and the velocity
resolution, values that may be retrieved for an ADP object name \code{d}
with \code{d[["velocityMaximum"]]} and \code{d[["velocityResolution"]]}.
}
\section{Memory considerations}{


For \code{RDI} files only, and only in the case where \code{by} is not specified,
an attempt is made to avoid running out of memory by skipping some profiles
in large input files. This only applies if \code{from} and \code{to} are both
integers; if they are times, none of the rest of this section applies.

The core issue is that the input file and the oce object will not be of equal size, because R
cannot store 2-byte integers natively. The R object will always require more memory
than in the data file (as is the whole point of this 2-byte encoding in RDI files, 
of course). A scale factor can be estimated by ignoring
vector quanties (e.g. time, one per profile) and concentrating on matrix properties
such as velocity, backscatter, and correlation. In this sort of data object,
these three elements are of equal dimensionality.  Velocity 
takes 2 bytes in the input file, while the other two quanties take 1 byte each. In
the oce object, these 4 bytes become 10 bytes (8 for the floating-point velocity, and 1
byte each for the others). Thus the oce object is likely about 10/4 times larger
than the input file. It makes sense to increase this expansion factor to be
conservative and to account for the byte expansion of the vector quantities, so
\code{read.adp.rdi} rounds 10/4 up to 3.

The next consideration is the memory limit in R. The least-performant machines
in typical use appear to be windows systems, which limit R to about 2GB (see
microsoft links pointed to by \code{\link{Memory-limits}}).
Since typical processing will require copying of objects,
e.g. for function call-by-value, \code{read.adp.rdi} uses a safety
factor of 4, setting an upper limit of 5e8 bytes. Using the factor of 3 compression
from R to an RDI file, this corresponds to an input file size of 1.5e9 bytes.
For a round number, \code{read.adp.rdi} translates this to 1e9 bytes.
This is the key to the calculation of a default value used for \code{by},
to be used if that parameter is not supplied in the function call.

The analysis is broken down into two cases. The first case is where \code{from=1} and
\code{to=0}, which also applies if neither \code{from} or \code{to} is given.
If the input file is smaller than 1e9 bytes, then \code{by} defaults to 1.
For larger files, \code{by} is set to the \code{\link{ceiling}} of the
ratio of input file size to 1e9 bytes. The second case is where \code{from} exceeds 1,
and/or \code{to} is nonzero. In this case, the value of \code{by} is calculated
in terms of \code{2e3*(to-from)}, which is an estimate of the number of bytes
to be processed (assuming 2e3 file bytes per profile, a rough estimate for RDI
files). If this value is smaller than 1e9, then \code{by} is set to 1. Otherwise,
\code{by} is set to the \code{\link{ceiling}} of the ratio of this value to 1e9.
}
\author{
Dan Kelley and Clark Richards
}
\references{
1. Teledyne-RDI, 2007. \emph{WorkHorse commands and output data
format.} P/N 957-6156-00 (November 2007).  (Section 5.3 h details the binary
format, e.g. the file should start with the byte \code{0x7f} repeated twice,
and each profile starts with the bytes \code{0x80}, followed by \code{0x00},
followed by the sequence number of the profile, represented as a
little-endian two-byte short integer.  \code{read.adp.rdi()} uses these
sequences to interpret data files.)
2. Teledyne-RDI, 2015. \emph{V Series output data format.} P/N 95D-6022-00 (May 2015).
}
\seealso{
Other things related to \code{adp} data: \code{\link{[[,adp-method}},
  \code{\link{[[<-,adp-method}}, \code{\link{adp-class}},
  \code{\link{adp}}, \code{\link{as.adp}},
  \code{\link{beamName}}, \code{\link{beamToXyzAdp}},
  \code{\link{beamToXyzAdv}}, \code{\link{beamToXyz}},
  \code{\link{beamUnspreadAdp}}, \code{\link{binmapAdp}},
  \code{\link{enuToOtherAdp}}, \code{\link{enuToOther}},
  \code{\link{plot,adp-method}},
  \code{\link{read.adp.nortek}},
  \code{\link{read.adp.sontek.serial}},
  \code{\link{read.adp.sontek}}, \code{\link{read.adp}},
  \code{\link{read.aquadoppHR}},
  \code{\link{read.aquadoppProfiler}},
  \code{\link{read.aquadopp}},
  \code{\link{subset,adp-method}},
  \code{\link{summary,adp-method}}, \code{\link{toEnuAdp}},
  \code{\link{toEnu}}, \code{\link{velocityStatistics}},
  \code{\link{xyzToEnuAdp}}, \code{\link{xyzToEnu}}
}

