% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amsr.R
\name{read.amsr}
\alias{read.amsr}
\title{Read an amsr file}
\usage{
read.amsr(file, debug = getOption("oceDebug"))
}
\arguments{
\item{file}{Sting indicating the name of a compressed file.}

\item{debug}{A debugging flag, integer.}
}
\description{
Read a compressed amsr file, generating an object that inherits from
\code{\link{amsr-class}}.  Note that only compressed files are read in
this version
}
\author{
Dan Kelley and Chantelle Layton
}
\concept{
satellite data
}

