% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremen.R
\name{read.bremen}
\alias{read.bremen}
\title{Read a Bremen File}
\usage{
read.bremen(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file
to load.}
}
\value{
An object of \code{\link{bremen-class}}.
}
\section{Issues}{
 This may be renamed (or removed) without notice.
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{bremen} data: \code{\link{[[,bremen-method}},
  \code{\link{[[<-,bremen-method}},
  \code{\link{bremen-class}},
  \code{\link{plot,bremen-method}},
  \code{\link{summary,bremen-method}}
}

