% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\name{read.coastline.shapefile}
\alias{read.coastline.shapefile}
\title{Read a Coastline File in Shapefile Format}
\usage{
read.coastline.shapefile(
  file,
  lonlim = c(-180, 180),
  latlim = c(-90, 90),
  encoding = NA,
  monitor = FALSE,
  debug = getOption("oceDebug"),
  processingLog
)
}
\arguments{
\item{file}{name of file containing coastline data (a file ending in \code{.shp})
or a zipfile that contains such a file, with a corresponding name.
The second scheme is useful for files downloaded from the NaturalEarth
website (see reference 2).}

\item{lonlim, latlim}{numerical vectors specifying the
west and east edges (and south and north edges) of a focus window.
Coastline polygons that do not intersect the defined box are
skipped, which can be useful in narrowing high-resolution world-scale
data to a local application.}

\item{encoding}{ignored.}

\item{monitor}{Logical indicating whether to print an indication of progress through
the file.}

\item{debug}{set to TRUE to print information about the header, etc.}

\item{processingLog}{if provided, the action item to be stored in the log.
(Typically only provided for internal calls; the default that it provides is
better for normal calls by a user.)}
}
\value{
x a \linkS4class{coastline} object.
}
\description{
Read coastline data stored in the shapefile format (see reference 1).
}
\section{A hack for depth contours}{
 The following demonstrates that this
code is getting close to working with depth contours.  This should be
handled more internally, and a new object for depth contours should be
constructed, of which coastlines could be a subset.
}

\references{
\enumerate{
\item The ``shapefile'' format is described in
\emph{ESRI Shapefile Technical Description}, March 1998, available at
\verb{https://www.esri.com/content/dam/esrisites/sitecore-archive/Files/Pdfs/library/whitepapers/pdfs/shapefile.pdf}
(last checked 2021-03-24).
\item The NaturalEarth website \verb{https://www.naturalearthdata.com/downloads/}
provides coastline datasets in three resolutions, along with similar files
lakes and rivers, for borders, etc. It is highly recommended.
}
}
\seealso{
Other things related to coastline data: 
\code{\link{[[,coastline-method}},
\code{\link{[[<-,coastline-method}},
\code{\link{as.coastline}()},
\code{\link{coastline-class}},
\code{\link{coastlineBest}()},
\code{\link{coastlineCut}()},
\code{\link{coastlineWorld}},
\code{\link{download.coastline}()},
\code{\link{plot,coastline-method}},
\code{\link{read.coastline.openstreetmap}()},
\code{\link{subset,coastline-method}},
\code{\link{summary,coastline-method}}
}
\author{
Dan Kelley
}
\concept{things related to coastline data}
