% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.aml.R
\name{read.ctd.aml}
\alias{read.ctd.aml}
\title{Read an AML CTD file}
\usage{
read.ctd.aml(file, debug = getOption("oceDebug"), processingLog, ...)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to
load.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}

\item{processingLog}{ignored.}

\item{...}{ignored.}
}
\value{
\code{\link[=read.ctd.aml]{read.ctd.aml()}} returns a \linkS4class{ctd} object.
}
\description{
\code{\link[=read.ctd.aml]{read.ctd.aml()}} reads files that hold data acquired with an
AML Oceanographic BaseX2 CTD instrument. In its present form,
the fuction only works for .txt files that were created with
certain settings in the AML software, as explained in
\sQuote{Details} section.
See \sQuote{Alternatives to this function} for hints on dealing
with files that do not match the \code{\link[=read.ctd.aml]{read.ctd.aml()}}
requirements.
}
\details{
The sample file upon which \code{\link[=read.ctd.aml]{read.ctd.aml()}} is based starts
with a line containing\preformatted{Date,Time,Conductivity (mS/cm),Temperature (C),Depth (m),Battery (V)
}

and this must also be true of any file provided to the function.
The entries of this line are the names of columnar data that appear later
in the file, in comma-separated format.
Those columns are assumed to start two lines below
a line starting with\preformatted{Comments:
}

and the lines above that are copied to an item named \code{header} in the
\code{metadata} slot of the returned value.  That slot also
holds \code{Longitude} and \code{Latitude}, inferred from the header, along
with some other items.
}
\section{Alternatives to this function}{

It is not difficult to read files that fail to meet the requirements of
\code{\link[=read.ctd.aml]{read.ctd.aml()}}.  The first step is to examine the file (\emph{without}
modifying it) using a text-based editor.  With this, it should be
easy to recognize a header at the start of the file. Take
note of the Longitude and Latitude, which you will need later,
along with any other information (e.g. serial numbers, etc.) that
you deem to be of interest.

Next, find the line where the columnar data start.  This will be used as the
\code{skip} argument in a call to e.g. \code{\link[=read.csv]{read.csv()}} or \code{\link[=read.delim]{read.delim()}} that you
will use to read the data.  Use \code{header=FALSE} in that function call, so that
the resultant value will be an object with entries named \code{V1}, \code{V2}, etc.  You
must determine the meaning of those columns, from the header or from some
other information.

You will be creating a \linkS4class{ctd} object with \code{\link[=as.ctd]{as.ctd()}}. This function
needs salinity, temperature, and pressure.  For modern computations phrased
in terms of the TEOS-10 equation of state, it also needs longitude and
latitude.  Pressure may be computed with \code{\link[=swPressure]{swPressure()}}, and
salinity with \code{\link[=swSCTp]{swSCTp()}}. Given all these values, the
final step is to use \code{\link[=as.ctd]{as.ctd()}} to create a \linkS4class{ctd} object.
}

\seealso{
Other functions that read ctd data: 
\code{\link{read.ctd.itp}()},
\code{\link{read.ctd.odf}()},
\code{\link{read.ctd.sbe}()},
\code{\link{read.ctd.woce.other}()},
\code{\link{read.ctd.woce}()},
\code{\link{read.ctd}()}
}
\author{
Dan Kelley
}
\concept{functions that read ctd data}
