% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.aml.R
\name{read.ctd.aml}
\alias{read.ctd.aml}
\title{Read an AML CTD file}
\usage{
read.ctd.aml(file, debug = getOption("oceDebug"))
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to
load.  This must be in the "txt" format, not the "csv" format.}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\value{
\code{\link[=read.ctd.aml]{read.ctd.aml()}} returns a \linkS4class{ctd} object.
}
\description{
This function CTD "txt" formatted data from an AML Oceanographic BaseX2 instrument.
It is in a preliminary form, based on inferences made by inspecting
a user-supplied data file; see \dQuote{Details}.
}
\details{
\code{\link[=read.ctd.aml]{read.ctd.aml()}} makes assumptions about the data columns
that might not be true for all files. For example, it assumes that
depth is presented, not pressure, and that the unit is metres.
Indeed, it requires that the name of the depth column (read in as
the first line in the file), is \code{"Depth (m)"}; any other string
will result in an error.  Similarly, conductivity must be provided,
in a column named \code{"Conductivity (mS/cm)"}, in-situ
temperature must be provided in a column named \code{"Temperature (C)"},
observation date (in yyyy-mm-dd format) must be provided in a column
named \code{"Date"}, and time (in hh-mm-ss format, assumed in UTC).
must be provided in a column named \code{"Time"}. If any of these requirements
is not met, an error is reported.

Assumptions are also made about the metadata.  For example, the supplied
test file had two entries for Longitude, and two for Latitude, which
perhaps relate to starting and ending locations, but only the first location
is stored by \code{\link[=read.ctd.aml]{read.ctd.aml()}}.  Also, although files have serial numbers
for multiple items called "Slots", only the one named \code{"SN"} is used
here, and saved as \code{serialNumber} in the metadata.

The header (contents up to and including the \verb{Comments=} line in the file)
is stored as \code{header} in the \code{metadata} slot of the returned value.
The data are assumed to start 2 lines below this \verb{Comments=} line.
}
\seealso{
Other functions that read ctd data: 
\code{\link{read.ctd.itp}()},
\code{\link{read.ctd.odf}()},
\code{\link{read.ctd.sbe}()},
\code{\link{read.ctd.woce.other}()},
\code{\link{read.ctd.woce}()},
\code{\link{read.ctd}()}
}
\author{
Dan Kelley
}
\concept{functions that read ctd data}
