% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{read.ctd}
\alias{read.ctd}
\title{Read a General CTD File}
\usage{
read.ctd(file, type = NULL, columns = NULL, station = NULL, missing.value,
  monitor = FALSE, debug = getOption("oceDebug"), processingLog, ...)
}
\arguments{
\item{file}{A connection or a character string giving the name of the file to
load.  For \code{read.ctd.sbe()} and \code{read.ctd.woce()}, this may be a
wildcard (e.g. \code{"*.cnv"} or \code{"*.csv"}) in which case the return
value is a vector containing CTD objects created by reading the files from
\code{\link{list.files}} with \code{pattern} set to the specified wildcard
pattern.}

\item{type}{If \code{NULL}, then the first line is studied, in order to
determine the file type.  If \code{type="SBE19"}, then a \emph{Seabird 19}, or
similar, CTD format is assumed. If \code{type="WOCE"} then a WOCE-exchange file
is assumed.  If \code{type="ITP"} then an ice-tethered profiler file is
assumed.  If \code{type="ODF"} an ODF file is assumed.  If \code{type="ODV"} an
ascii-ODV file is assumed.}

\item{columns}{An optional \code{\link{list}} that can be used to convert unrecognized
data names to resultant variable names.  This is used only by
\code{\link{read.ctd.sbe}} and \code{\link{read.ctd.odf}}; see
\dQuote{Examples}.}

\item{station}{Optional character string containing an identifying name or
number for the station. This can be useful if the routine cannot determine the
name automatically, or if another name is preferred.}

\item{missing.value}{Optional missing-value flag; data matching this value will
be set to \code{NA} upon reading.}

\item{monitor}{Boolean, set to \code{TRUE} to provide an indication of
progress.  This is useful if \code{filename} is a wildcard.}

\item{debug}{An integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed.}

\item{processingLog}{If provided, the action item to be stored in the log.
This is typically only provided for internal calls; the default that it provides is
better for normal calls by a user.}

\item{...}{additional arguments, passed to called routines.}
}
\value{
An object of \code{\link{ctd-class}}. The details of the contents
depend on the source file. The \code{metadata} slot is particularly 
variable across data formats, because the meta-information provided
in those formats varies widely.
}
\description{
Read a General CTD File
}
\details{
\code{read.ctd()} is a base function that in turn calls specialized functions, e.g.
\code{\link{read.ctd.odf}} for the ODF data used in Fisheries and Oceans (Canada),
\code{\link{read.ctd.woce}} for data in World Ocean Circulation Experiment format,
\code{\link{read.ctd.woce.other}} for a variant of WOCE data,
\code{\link{read.ctd.itp}} for ice-tethered-profiler data, or
\code{\link{read.ctd.sbe}} for Seabird data.
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{cnvName2oceName}}, \code{\link{ctd-class}},
  \code{\link{ctdAddColumn}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdRaw}},
  \code{\link{ctdTrim}}, \code{\link{ctdUpdateHeader}},
  \code{\link{ctd}}, \code{\link{gps-class}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}}, \code{\link{plotProfile}},
  \code{\link{plotScan}}, \code{\link{plotTS}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}},
  \code{\link{woceNames2oceNames}}, \code{\link{write.ctd}}
}

