% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{read.ctd.sbe}
\alias{read.ctd.sbe}
\title{Read an SBE-type CTD file}
\usage{
read.ctd.sbe(file, columns = NULL, station = NULL, missing.value,
  monitor = FALSE, debug = getOption("oceDebug"), processingLog, ...)
}
\arguments{
\item{file}{A connection or a character string giving the name of the file to
load.  For \code{read.ctd.sbe()} and \code{read.ctd.woce()}, this may be a
wildcard (e.g.  \code{"*.cnv"} or \code{"*.csv"}) in which case the return
value is a vector containing CTD objects created by reading the files from
\code{\link{list.files}} with \code{pattern} set to the specified wildcard
pattern.}

\item{columns}{If \code{NULL}, then \code{read.ctd} tries to infer column names
from the header.  For SBE files only, the \code{column} argument can control
the column selection.  It is a list that names data types and the columns
containing them, starting at 1.  The list must include \code{"pressure"},
\code{"temperature"} and \code{"salinity"}, with any other values being ignored
(in this version of the function).}

\item{station}{Optional character string containing an identifying name or
number for the station. This can be useful if the routine cannot determine the
name automatically, or if another name is preferred.}

\item{missing.value}{Optional missing-value flag; data matching this value will
be set to \code{NA} upon reading.}

\item{monitor}{Boolean, set to \code{TRUE} to provide an indication of
progress.  This is useful if \code{filename} is a wildcard.}

\item{debug}{A flag that turns on debugging.  Set to 1 to get a moderate amount
of debugging information, or to 2 to get more.}

\item{processingLog}{If provided, the action item to be stored in the log.
This is typically only provided for internal calls; the default that it provides is
better for normal calls by a user.}

\item{...}{additional arguments, passed to called routines.}
}
\value{
An object of \code{\link{ctd-class}}. The details of the contents
depend on the source file. The \code{metadata} slot is particularly 
variable across data formats, because the meta-information provided
in those formats varies widely.
}
\description{
Read an SBE-type CTD file

The \code{read.ctd} family of functions can read a variety of CTD
data formats. See \dQuote{Details}.
}
\details{
\code{read.ctd.sbe()} reads files stored in Seabird \code{.cnv} format.
}
\author{
Dan Kelley
}
\references{
The Sea-Bird SBE 19plus profiler is described at
\url{http://www.seabird.com/products/spec_sheets/19plusdata.htm}.  Some more
information is given in the Sea-Bird data-processing manaual
\url{http://www.seabird.com/old-manuals/Software_Manuals/SBE_Data_Processing/SBEDataProcessing_7.20g.pdf}.
}
\seealso{
Other functions that handle ctd data: \code{\link{read.ctd.itp}},
  \code{\link{read.ctd.odf}}, \code{\link{read.ctd.odv}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}}
}

