% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.sbe.R
\name{read.ctd.sbe}
\alias{read.ctd.sbe}
\title{Read an Seabird CTD File}
\usage{
read.ctd.sbe(file, columns = NULL, station = NULL, missing.value,
  monitor = FALSE, debug = getOption("oceDebug"), processingLog, ...)
}
\arguments{
\item{file}{A connection or a character string giving the name of the file to
load.  For \code{read.ctd.sbe()} and \code{read.ctd.woce()}, this may be a
wildcard (e.g. \code{"*.cnv"} or \code{"*.csv"}) in which case the return
value is a vector containing CTD objects created by reading the files from
\code{\link{list.files}} with \code{pattern} set to the specified wildcard
pattern.}

\item{columns}{Ignored, and slated for removal in April, 2016;
see \link{oce-defunct}.}

\item{station}{Optional character string containing an identifying name or
number for the station. This can be useful if the routine cannot determine the
name automatically, or if another name is preferred.}

\item{missing.value}{Optional missing-value flag; data matching this value will
be set to \code{NA} upon reading.}

\item{monitor}{Boolean, set to \code{TRUE} to provide an indication of
progress.  This is useful if \code{filename} is a wildcard.}

\item{debug}{An integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed.}

\item{processingLog}{If provided, the action item to be stored in the log.
This is typically only provided for internal calls; the default that it provides is
better for normal calls by a user.}

\item{...}{additional arguments, passed to called routines.}
}
\value{
An object of \code{\link{ctd-class}}. The details of the contents
depend on the source file. The \code{metadata} slot is particularly 
variable across data formats, because the meta-information provided
in those formats varies widely.
}
\description{
Read a Teledyne/Seabird CTD file, i.e. a file with name ending in \code{.cnv}.
}
\details{
\code{read.ctd.sbe} reads files stored in Seabird \code{.cnv} format.
Note that these files can contain multiple sensors for a given field. For example,
the file might contain a column named \code{t090C} for one 
temperature sensor and \code{t190C} for a second. The first will be denoted
\code{temperature} in the \code{data} slot of the return value, and the second
will be denoted \code{temperature1}. This means that the first sensor
will be used in any future processing that accesses \code{temperature}. This
is for convenience of processing, and it does not pose a limitation, because the
data from the second sensor are also available as e.g. \code{x[["temperature1"]]},
where \code{x} is the name of the returned value.  For the details of the 
mapping from \code{.cnv} names to \code{ctd} names, see \code{\link{cnvName2oceName}}.

The original variable names as stored in \code{file} are stored within the \code{data}
slot as attribute named \code{namesOriginal}, and so can be retrieved with e.g.
\code{attr(x@data, "namesOriginal")}, if \code{x} is the object returned by
this function. This feature can be helpful in detailed work.  See the Appendix
VI of [2], noting that the \code{namesOriginal} values are taken from
the "Short Name" column of the table spanning pages 161 through 172.
}
\author{
Dan Kelley
}
\references{
1. The Sea-Bird SBE 19plus profiler is described at
\url{http://www.seabird.com/products/spec_sheets/19plusdata.htm}.  Some more
information is given in the Sea-Bird data-processing manaual
\url{http://www.seabird.com/old-manuals/Software_Manuals/SBE_Data_Processing/SBEDataProcessing_7.20g.pdf}.

2. A SBE data processing manual is at \url{http://www.seabird.com/sites/all/modules/pubdlcnt/pubdlcnt.php?file=http://www.seabird.com/sites/default/files/documents/SBEDataProcessing_7.25.0.pdf&nid=1320}.
}
\seealso{
Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{cnvName2oceName}}, \code{\link{ctd-class}},
  \code{\link{ctdAddColumn}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdRaw}},
  \code{\link{ctdTrim}}, \code{\link{ctdUpdateHeader}},
  \code{\link{ctd}}, \code{\link{gps-class}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}}, \code{\link{plotProfile}},
  \code{\link{plotScan}}, \code{\link{plotTS}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.odv}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}},
  \code{\link{woceNames2oceNames}}, \code{\link{write.ctd}}
}

