% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{read.ctd.woce}
\alias{read.ctd.woce}
\title{Read a WOCE-type CTD file in which the first word is "CTD"}
\usage{
read.ctd.woce(file, columns = NULL, station = NULL, missing.value = -999,
  monitor = FALSE, debug = getOption("oceDebug"), processingLog, ...)
}
\arguments{
\item{file}{A connection or a character string giving the name of the file to
load.  For \code{read.ctd.sbe()} and \code{read.ctd.woce()}, this may be a
wildcard (e.g.  \code{"*.cnv"} or \code{"*.csv"}) in which case the return
value is a vector containing CTD objects created by reading the files from
\code{\link{list.files}} with \code{pattern} set to the specified wildcard
pattern.}

\item{columns}{If \code{NULL}, then \code{read.ctd} tries to infer column names
from the header.  For SBE files only, the \code{column} argument can control
the column selection.  It is a list that names data types and the columns
containing them, starting at 1.  The list must include \code{"pressure"},
\code{"temperature"} and \code{"salinity"}, with any other values being ignored
(in this version of the function).}

\item{station}{Optional character string containing an identifying name or
number for the station. This can be useful if the routine cannot determine the
name automatically, or if another name is preferred.}

\item{missing.value}{Optional missing-value flag; data matching this value will
be set to \code{NA} upon reading.}

\item{monitor}{Boolean, set to \code{TRUE} to provide an indication of
progress.  This is useful if \code{filename} is a wildcard.}

\item{debug}{A flag that turns on debugging.  Set to 1 to get a moderate amount
of debugging information, or to 2 to get more.}

\item{processingLog}{If provided, the action item to be stored in the log.
This is typically only provided for internal calls; the default that it provides is
better for normal calls by a user.}

\item{...}{additional arguments, passed to called routines.}
}
\value{
An object of \code{\link{ctd-class}}. The details of the contents
depend on the source file. The \code{metadata} slot is particularly 
variable across data formats, because the meta-information provided
in those formats varies widely.
}
\description{
Read a WOCE-type CTD file in which the first word is "CTD"

The \code{read.ctd} family of functions can read a variety of CTD
data formats. See \dQuote{Details}.
}
\details{
\code{read.ctd.woce()} reads files stored in the exchange format used
by the World Ocean Circulation Experiment (WOCE), in which the first 4
characters are ``\code{CTD,}''. It also also in a rarer format with
the first 3 characters are \code{CTD}'' followed by a blank or the end
of the line.
}
\author{
Dan Kelley
}
\references{
The WOCE-exchange format is described at
\code{http://woce.nodc.noaa.gov/woce_v3/wocedata_1/whp/exchange/exchange_format_desc.htm},
and a sample file is at
\url{http://woce.nodc.noaa.gov/woce_v3/wocedata_1/whp/exchange/example_ct1.csv}
}
\seealso{
Other functions that handle \code{ctd} data: \code{\link{read.ctd.itp}},
  \code{\link{read.ctd.odf}}, \code{\link{read.ctd.odv}},
  \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}}, \code{\link{read.ctd}}
}

