% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satellite.R
\name{read.g1sst}
\alias{read.g1sst}
\title{Read G1SST satellite data}
\usage{
read.g1sst(filename)
}
\arguments{
\item{filename}{name of a netcdf file containing data from the G1SST satellite}
}
\value{
An object of \code{\link{g1sst-class}}.
}
\description{
This works with netcdf files as provided by the ERDAPP server [1].
}
\examples{
\dontrun{
s <-
"http://coastwatch.pfeg.noaa.gov/erddap/griddap/jplG1SST.nc?SST\%5B(2016-01-02T12:00:00Z)\%5D\%5B(44):(46)\%5D\%5B(-67.5):(-63.5)\%5D"
if (!length(list.files(pattern="^a.nc$")))
    download.file(s, "a.nc")
d <- read.g1sst("a.nc")
plot(d, "SST", col=oceColorsJet)
}
}
\author{
Dan Kelley
}
\references{
1. \url{http://coastwatch.pfeg.noaa.gov/erddap/}
}
\concept{
satellite
}

