% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satellite.R
\name{read.g1sst}
\alias{read.g1sst}
\title{Read G1SST satellite data}
\usage{
read.g1sst(filename)
}
\arguments{
\item{filename}{name of a netcdf file containing data from the G1SST satellite}
}
\value{
An object of \code{\link{g1sst-class}}.
}
\description{
This works with netcdf files as provided by the ERDAPP server [1].
}
\examples{
\dontrun{
# Construct query, making it easier to understand and modify.
day <- "2016-01-02"
lon0 <- -67.5
lon1 <- -63.5
lat0 <- 44
lat1 <- 46
source <- paste("http://coastwatch.pfeg.noaa.gov/erddap/griddap/",
                "jplG1SST.nc?",
                "SST\%5B(", day, "T12:00:00Z)",
                "\%5D\%5B(", lat0, "):(", lat1, ")",
                "\%5D\%5B(", lon0, "):(", lon1, ")",
                "\%5D", sep="")
if (!length(list.files(pattern="^a.nc$")))
    download.file(source, "a.nc")
d <- read.g1sst("a.nc")
plot(d, "SST", col=oceColorsJet)
}
}
\author{
Dan Kelley
}
\references{
1. \url{http://coastwatch.pfeg.noaa.gov/erddap/}
}
\concept{
satellite
}

