% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{read.netcdf}
\alias{read.netcdf}
\title{Read a NetCDF File}
\usage{
read.netcdf(file, ...)
}
\arguments{
\item{file}{the name of a file}

\item{...}{unused}
}
\value{
An object of \code{\link{oce-class}}.
}
\description{
Read a NetCDF File
}
\details{
Read a netcdf file, trying to interpret its contents sensibly.

It is important to note that this is a preliminary version of
this function, and much about it may change without notice.
Indeed, it may be removed entirely.

Below are some features that may be changed.

1. The names of data items are not changed from those in the netcdf
file on the assumption that this will offer the least surprise to
the user.

2. An attempt is made to find some common metadata from global
attributes in the netcdf file. These attributes include
\code{Longitude}, \code{Latitude}, \code{Ship} and \code{Cruise}.
Before they are stored in the metadata, they are converted to
lower case, since that is the oce convention.
}
