% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangaea.R
\name{read.pangaea}
\alias{read.pangaea}
\title{Read a tab-separated file from the PANGEA website [1]}
\usage{
read.pangaea(file, debug = getOption("oceDebug"))
}
\arguments{
\item{file}{a connection or a character string giving the name of the file
to load.}

\item{debug}{An integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed.}
}
\description{
This is a preliminary version of this function, and everything
may change through the Autumn of 2016. Indeed, the function could
be deleted.
}
\author{
Dan Kelley
}
\references{
1. PANGAEA website \url{https://www.pangaea.de}
}

