% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbt.R
\name{read.xbt.noaa2}
\alias{read.xbt.noaa2}
\title{Read an xbt File in UBT (Universal BathyThermograph) Format}
\usage{
read.xbt.noaa2(
  file,
  debug = getOption("oceDebug"),
  missingValue = -9.99,
  encoding = "latin1",
  processingLog
)
}
\arguments{
\item{file}{character value naming a file, or a file connection,
containing the data, with each line corresponding to an XBT
profile.}

\item{debug}{a flag that turns on debugging.  The value indicates the depth
within the call stack to which debugging applies.}

\item{missingValue}{ignored.}

\item{encoding}{a character value that indicates the encoding to be used for
this data file, if it is textual.  The default value for most functions is
\code{"latin1"}, which seems to be suitable for files containing text written in
English and French.}

\item{processingLog}{ignored.}
}
\value{
A list containing \linkS4class{xbt} objects, one per line in \code{file}.
}
\description{
This is preliminary function that might be changed through
the month of February 2025.  Note that, at present, this
function returns a list of \linkS4class{xbt} objects, as opposed
to a single \linkS4class{xbt} object. This is because the data
format holds one profile per line.  Also, please note that
the documentation used to frame the function (Reference 1)
does not accurately describe test files (Reference 2), owing
to a change of year format from the documented two-character
form to a more modern 4-character form.  Other aspects of
the documented format are also at odds with the data (see
issue 2289 on the github site for oce).
}
\references{
\enumerate{
\item https://www.ncei.noaa.gov/data/oceans/nodc/formats/UBT_Universal_Bathythermograph.html
\item https://data.noaa.gov/onestop/collections/details/22c9b8d0-c32b-4824-815b-04ce80078d10
}
}
\seealso{
Other things related to xbt data: 
\code{\link{[[,xbt-method}},
\code{\link{[[<-,xbt-method}},
\code{\link{as.xbt}()},
\code{\link{plot,xbt-method}},
\code{\link{read.xbt}()},
\code{\link{read.xbt.noaa1}()},
\code{\link{subset,xbt-method}},
\code{\link{summary,xbt-method}},
\code{\link{xbt}},
\code{\link{xbt-class}},
\code{\link{xbt.edf}},
\code{\link{xbt2.edf}}
}
\author{
Dan Kelley
}
\concept{things related to xbt data}
