% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename variables according to a specified dictionary}
\usage{
rename(x, dictionary = "ioos.csv", debug = 0)
}
\arguments{
\item{x}{an \linkS4class{oce} object.}

\item{dictionary}{either a string or a data frame.  If a string, then it is
either the name of a built-in vocabulary (e.g. \code{sbe}; see \sQuote{Details})
or the name of a CSV file that defines a dictionary. If it is a data frame,
the it must have unnamed columns containing information as in the file
method; to see an example, type the following in an R console.

\if{html}{\out{<div class="sourceCode">}}\preformatted{system.file("extdata", "dictionary_sbe.csv", package = "oce")
}\if{html}{\out{</div>}}}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}
}
\description{
The is a provisional function, added in September 2024, and likely to change
through the end of that year.
}
\details{
The dictionary format, whether read from a built-in CSV file, or from a
user-supplied CSV file, or as a data frame, is as follows.
\enumerate{
\item A character value that specifies the original name of a variable in the
\code{data} slot of \code{x}. This is used in matching such names against targets.
Matches may be in the form of equality, or string match (in which any \verb{#}
characters in the column are matched to single digits in variable names).
\item The oce-convention name to be used for a match.
\item The unit for the column (only used if the object does not already hold a
unit for this entry).
\item The scale for the column (again, only used if the object does not already
hold a scale).
}

For example, an entry in a CSV file

\if{html}{\out{<div class="sourceCode">}}\preformatted{PSALST##,salinity,,PSS-78
}\if{html}{\out{</div>}}

specifies that a variable named \code{"PSALT"} followed by 2 digits
is to be renamed as \code{"salinity"}, that the unit (if not
already defined within \code{x}) is to be blank, and that
the scale (again, if not already defined) is to be \code{"PSS-78"}.

Note that after the renaming is done, duplicate names are handled
in the standard oce way, e.g. if the object holds three temperature
variables, they will be named \code{temperature}, \code{temperature2} and
\code{temperature3}.
}
\section{History and Plans}{

This function was written in late September, 2024. It is likely
to evolve through the remaining months of 2024, after real-world
testing by the developers.
}

\examples{
library(oce)
# Example 1: made-up data
d <- new("oce")
d <- oceSetData(d, "S", c(30, 31))
d <- oceSetData(d, "T", c(10, 11))
dictText <- "S,salinity,,
T,temperature,degree*C,ITS-90"
dictionary <- read.csv(text = dictText, header = FALSE)
rename(d, dictionary)
#
# Example 2: a CIOOS NetCDF file. Note that this file
# is downloaded and removed at the end; in practice,
# it is likely that the file might be retained locally.
if (requireNamespace("curl")) {
    file <- tempfile(fileext = ".nc") # removed later
    server <- "https://cioosatlantic.ca/erddap/files"
    program <- "bio_atlantic_zone_monitoring_program_ctd"
    subprogram <- "Bedford\%20Basin\%20Monitoring\%20Program"
    year <- "2023"
    cast <- "001"
    url <- paste0(
        server, "/", program, "/", subprogram, "/", year,
        "/CTD_BCD2023667_", cast, "_1_DN.ODF.nc"
    )
    curl::curl_download(url, file)
    d <- read.netcdf(file)
    summary(d)
    dd <- rename(d, "ioos")
    summary(dd)
    unlink(file)
}

}
\author{
Dan Kelley
}
