% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{renameData}
\alias{renameData}
\title{Rename items in the data slot of an oce object (\strong{deprecated})}
\usage{
renameData(x, old = NULL, new = NULL)
}
\arguments{
\item{x}{an \linkS4class{oce} object.}

\item{old}{Vector of strings, containing old names.}

\item{new}{Vector of strings, containing old names.}
}
\description{
This was deprecated in December 2019, because \code{\link[=oceRenameData]{oceRenameData()}} does
a better job and is more consistent with other functions that work
with items in the \code{data} and \code{metadata} slots.
}
