% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repositoryURL}
\alias{repositoryURL}
\title{Calculate a web link to download Oceanographic data}
\usage{
repositoryURL(item, id, window, domain, subdomain)
}
\arguments{
\item{item}{A string indicating the item sought; see \dQuote{Details}.}

\item{id}{A string indicating the ID of the sought item.}

\item{window}{An indication of a window to select data (ignored at present).}

\item{domain}{The base-level domain (more information to be added later).}

\item{subdomain}{The second-level domain (more information to be added later).}
}
\description{
This function is able to construct universal resource identifiers (URIs)
that can be used with \code{\link{download.file}} to download data from
several common Oceanographic repositories.   The goal of this function is to
spare users the task of finding such URLs for casual use or one-off
applications.  There is no intention of trying  to reproduce the web tools
provided with many repositories, nor to provide an exhaustive interface to
a wide suite of data types (see \dQuote{Contributing}).
}
\details{
The following values for \code{item} are handled. \itemize{
\item \code{item="argo"}. The data are by default sought from domain
\code{ftp://usgodae.org}, and the procedure is to construct
a test URL as \code{ftp://usgodae.org/pub/outgoing/argo//dac/} and
then to probe that to find the list of available 'dac' (data 
archiving centre) subdirectories. Then each of these is searched
to try to find a subdirectory with name equal to \code{id}. The
first matching subdirectory is then opened and a test is made
for the existence of a file with name constructed by pasting 
\code{"_prof.nc"} onto \code{id}. If this works, then the URL
is returned; otherwise an error message is given. Note: early
tests with \code{http://data.nodc.noaa.gov} showed that it stores 
its argo files with \emph{lower-case} names, which is inconsistent
with argo documentation [1] and incompatible with \code{\link{read.argo}}.
So, don't use that server!
}
}
\section{History}{

This function was drafted in late January, 2016, and it is likely to
change, perhaps in incompatible ways, through February.
}

\section{Contributing}{

This function will be most useful if users contribute to it, both by 
pointing out changes to URLs and by suggesting additions for useful
datasets. The most useful comments will be concrete; merely suggesting
that an interface be provided to such-and-such server is too vague
to help much.
}
\examples{
file <- repositoryURL(item="argo", id="6900388")
\dontrun{
download.file(file, "argo.nc")
argo <- read.oce("argo.nc")
summary(argo)
plot(argo)
}

}
\author{
Dan Kelley
}
\references{
\itemize{
\item \strong{1.} The \code{argo} netCDF format is described at
\code{http://archimer.ifremer.fr/doc/00187/29825/40575.pdf}; 
as of January 28, 2016, Section 2.2.4 of this document
indicates that variable names are to be in upper case,
which is true for the \code{nodc.noaa.gov} server, so data provided
by this server cannot be read with \code{\link{read.argo}}.
}
}
\seealso{
\code{\link{download.file}} for downloading the files found by \code{repositoryURI}
}

