% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repositoryURL}
\alias{repositoryURL}
\title{Calculate a web link to download Oceanographic data}
\usage{
repositoryURL(item, id, window, domain, subdomain)
}
\arguments{
\item{item}{A string indicating the item sought; see \dQuote{Details}.}

\item{id}{A string indicating the ID of the sought item.}

\item{window}{A list specifying a data window for a snapshot; the contents are
\code{basin}, a string naming the ocean basin (must be \code{"Atlantic"},
\code{"Pacific"} or \code{"Indian"}) and a string named \code{time},
specifying a date in the format \code{yyyymmdd}.}

\item{domain}{The base-level domain (more information to be added later).}

\item{subdomain}{The second-level domain (more information to be added later).}
}
\description{
This function is able to construct universal resource identifiers (URIs)
that can be used with \code{\link{download.file}} to download data from
several common Oceanographic repositories.   The goal of this function is to
spare users the task of finding such URLs for casual use or one-off
applications.  There is no intention of trying  to reproduce the web tools
provided with many repositories, nor to provide an exhaustive interface to
a wide suite of data types (see \dQuote{Contributing}).
}
\details{
The following values for \code{item} are handled.

\enumerate{
\item \code{item="argo"} looks up argo data, referenced either by 
float id or by basin plus time.

\itemize{

\item \emph{id mode.} The return value is constructed
as \code{[domain]/pub/outgoing/argo/dac/[subdomain]/[id]/[id]_prof.nc},
where square brackets indicate the values of the named arguments
to \code{repositoryURL}. This has only been checked for domain
\code{domain="ftp://usgodae.org"} and \code{subdomain="bodc"},
so these are the defaults. The value of the relevant subdomain may
be known to users familiar with argo data; others can discover the
data archiving center for a given float by inspecting the subdirectories
of the FTP site \code{ftp://usgodae.org/pub/outgoing/argo/dac}
or by accessing
\code{ftp://usgodae.org/pub/outgoing/argo/ar_index_global_meta.txt}
and searching for the float identifer.
\item \emph{basin/time mode.}  See example 2.

}

Caution: tests made in January 2016 revealed that 
\code{http://data.nodc.noaa.gov} stores argo files with
\emph{lower-case} names, which is inconsistent
with argo documentation [1] and incompatible with \code{\link{read.argo}}.

\item \code{item="hydrography"} looks up hydrographic cruise data.
At present, the only option is \code{window=list(cruise=EXPOCODE)},
where \code{EXPOCODE} is a string holding an expedition code.  See example 3.

}
}
\section{History}{

This function was drafted in late January, 2016, and its argument
list is likely to change through the early months of 2016. Some
of this will be to accommodate new data types, but changes to 
website structures will also precipitate changes to how this
function works.
}

\section{Contributing}{

This function will be most useful if users contribute to it, both by 
pointing out changes to URLs and by suggesting additions for useful
datasets. The most useful comments will be concrete; merely suggesting
that an interface be provided to such-and-such server is too vague
to help much.
}
\examples{
# 1. Historical data from individual float
file <- repositoryURL(item="argo", id="6900388")
\dontrun{
download.file(file, "argo.nc")
argo <- read.oce("argo.nc")
summary(argo)
plot(argo)
mtext(file, side=3, line=0)
}

# 2. Plot map of Atlantic floats on New Year's Day, 2016.
file <- repositoryURL(item="argo",
                      window=list(basin="atlantic", time="20130101"))
\dontrun{
download.file(file, "nyd.nc")
nyd <- read.oce("nyd.nc")
summary(nydj)
plot(nyd)
mtext(file, side=3, line=0)
}

# 3. Get a URL to a zipfile that expands to a directory holding files
#    for data from cruise 74JC20150110.
file <- repositoryURL(item="hydrography", window=list(cruise="74JC20150110"))
download.file(file, "cchdo_74JC20150110.zip")
# expect  http://cchdo.ucsd.edu/cruise/74JC20150110?download=dataset

}
\author{
Dan Kelley
}
\references{
1. The \code{argo} netCDF format is described at
\code{http://archimer.ifremer.fr/doc/00187/29825/40575.pdf}; 
as of January 28, 2016, Section 2.2.4 of this document
indicates that variable names are to be in upper case,
which is true for the \code{nodc.noaa.gov} server, so data provided
by this server cannot be read with \code{\link{read.argo}}.
}
\seealso{
\code{\link{download.file}} for downloading the files found by \code{repositoryURI}
}

