% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsk.R
\docType{data}
\name{rsk}
\alias{rsk}
\title{Sample rsk Dataset}
\description{
A sample \code{rsk} object derived from a Concerto CTD manufactured by RBR Ltd.
}
\details{
The data were obtained September 2015, off the west coast
    of Greenland, by Matt Rutherford and Nicole Trenholm of the
    Ocean Research Project, in collaboration with RBR and with the
    NASA Oceans Melting Greenland project.
}

