% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satellite.R
\docType{class}
\name{satellite-class}
\alias{satellite-class}
\title{Class to hold satellite data}
\description{
This class holds satellite data of various types, including
\code{\link{amsr-class}} and \code{\link{g1sst-class}}.
}
\author{
Dan Kelley and Chantelle Layton
}
\concept{
satellite
}

