% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.R
\docType{class}
\name{satellite-class}
\alias{satellite-class}
\title{Class to Store Satellite Data}
\description{
This class holds satellite data of various types, including
\linkS4class{amsr} and \linkS4class{g1sst}.
}
\seealso{
Other things related to satellite data: \code{\link{g1sst-class}},
  \code{\link{plot,satellite-method}},
  \code{\link{read.g1sst}},
  \code{\link{summary,satellite-method}}
}
\author{
Dan Kelley and Chantelle Layton
}
\concept{satellite}
\concept{things related to satellite data}
