\name{sealevel-class}

\docType{class}

\alias{sealevel-class}

\title{Class to store sealevel data}

\description{Class to store sealevel data, e.g. from a tide gauge, with
    standard slots \code{metadata}, \code{data} and \code{processingLog}.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

