\name{sealevel}
\docType{data}
\alias{sealevel}
\title{Sample sea-level data set}
\description{This sample sea-level dataset is the 2003 record from Halifax Harbour in Nova Scotia, Canada.
Note that Hurricane Juan hit Halifax on September 28th of 2003, causing a strong storm surge.}  
\usage{data(sealevel)}
\source{The data were downloaded from MEDS and then read with \code{\link{read.sealevel}}.}
\references{The MEDS archive is at \url{http://www.meds-sdmm.dfo-mpo.gc.ca/meds/Databases/TWL/TWL_inventory_e.htm}.
An entry to the literature about Hurricane Juan is at
\url{http://www.atl.ec.gc.ca/weather/hurricane/juan/}.}
\author{Dan Kelley}
\keyword{misc}
