\name{section-class}

\docType{class}

\alias{section-class}

\title{Class to store hydrographic section data}

\description{Class to store hydrographic section data, with standard slots
    \code{metadata}, \code{data} and \code{processingLog}.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

