% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{setFlags}
\alias{setFlags}
\title{Set data-quality flags within a oce object}
\usage{
setFlags(object, name = NULL, i = NULL, value = NULL, default = NULL,
  debug = 0)
}
\arguments{
\item{object}{An oce object.}

\item{name}{Character string indicating the name of the variable to be flagged. If
this variable is not contained in the object's \code{data} slot, an error is reported.}

\item{i}{Indication of where to insert the flags; see \dQuote{Details}.}

\item{value}{The value to be inserted in the flag.}

\item{default}{The default (good) value of the flag. This is used only if
the object does not yet have yet have a entry for \code{name} flags. In that case,
storage is set up for the flag and it is filled with the \code{default} value, after
which \code{setFlags} returns to the task of setting flag values at indicated locations.}

\item{debug}{Integer set to 0 for quiet action or to 1 for some debugging.}
}
\value{
An object with flags set as indicated.
}
\description{
This function changes specified entries in the data-quality
flags of a oce object, which are stored within
a list named \code{flags} that resides in the \code{metadta}
slot. If the object already has a flag set up for \code{name},
then only the specified entries are altered. If not, the flag
entry is first created and its entries set to \code{default},
after which the specified entries are changed to \code{value}. The
specification is made with \code{i}, the form of which
varies between classes; see \dQuote{Details} for the
particular case of \code{\link{oce-class}} objects.
}
\details{
This generic function is overridden by specialized functions for some object classes.
}
\section{Development note}{

This function was added in late April, 2018, and is likely to undergo
changes until the end of June of that year.  Use with caution.
}

\seealso{
Other functions relating to data-quality flags: \code{\link{handleFlags,adp-method}},
  \code{\link{handleFlags,argo-method}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{handleFlags,section-method}},
  \code{\link{handleFlags}},
  \code{\link{setFlags,adp-method}},
  \code{\link{setFlags,ctd-method}},
  \code{\link{setFlags,oce-method}}
}
