% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{shiftLongitude}
\alias{shiftLongitude}
\title{Shift longitude to range -180 to 180, if any element exceeds 180}
\usage{
shiftLongitude(longitudes)
}
\arguments{
\item{longitudes}{a numericl vector of longitudes}
}
\value{
vector of longitudes, shifted to the desired range.
}
\description{
This is a utility function used by \code{\link{mapGrid}}. It works
simply by subtracting 180 from each longitude, if any longitude
in the vector exceeds 180.
}
\seealso{
\code{\link{matrixShiftLongitude}} and \code{\link{standardizeLongitude}}.
}

