% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{standardDepths}
\alias{standardDepths}
\title{Standard Oceanographic Depths}
\usage{
standardDepths()
}
\value{
A vector of depths, c(0, 10, ...).
}
\description{
This returns so-called standard depths 0m, 10m, etc. below the sea surface.
}
\author{
Dan Kelley
}
