% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\docType{methods}
\name{[[,adp-method}
\alias{[[,adp-method}
\title{Extract Something From an adp Object}
\usage{
\S4method{[[}{adp}(x, i, j, ...)
}
\arguments{
\item{x}{An adp object, i.e. one inheriting from \code{\link{adp-class}}.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
In addition to the usual extraction of elements by name, some shortcuts
are also provided, e.g. \code{u1} retrieves \code{v[,1]}, and similarly
for the other velocity components. The \code{a} and \code{q}
data can be retrived in \code{\link{raw}} form or numeric
form; see examples.
}
\examples{
data(adp)
head(adp[["v"]][,,1])
}
\seealso{
Other functions that access oce data and metadata: \code{\link{[[,adv-method}},
  \code{\link{[[,argo-method}},
  \code{\link{[[,ladp-method}},
  \code{\link{[[,section-method}}

Other functions that handle adp data: \code{\link{[[<-,adp-method}},
  \code{\link{subset,adp-method}}
}

