% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argo.R
\docType{methods}
\name{[[,argo-method}
\alias{[[,argo-method}
\title{Extract something from an argo object}
\usage{
\S4method{[[}{argo}(x, i, j, ...)
}
\arguments{
\item{x}{An argo object, i.e. one inheriting from \code{\link{argo-class}}.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
Extract something from an argo object
}
\seealso{
Other functions that access oce data and metadata: \code{\link{[[,adp-method}},
  \code{\link{[[,adv-method}}, \code{\link{[[,ctd-method}},
  \code{\link{[[,ladp-method}},
  \code{\link{[[,section-method}}

Other functions that deal with argo data: \code{\link{[[<-,argo-method}},
  \code{\link{argoGrid}}, \code{\link{as.argo}},
  \code{\link{plot,argo-method}}, \code{\link{read.argo}},
  \code{\link{subset,argo-method}},
  \code{\link{summary,argo-method}}
}

