% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{[[,ctd-method}
\alias{[[,ctd-method}
\title{Extract Something From a CTD Object}
\usage{
\S4method{[[}{ctd}(x, i, j, ...)
}
\arguments{
\item{x}{A ctd object, i.e. one inheriting from \code{\link{ctd-class}}.}

\item{i}{The item to extract.}

\item{j}{Optional additional information on the \code{i} item.}

\item{...}{Optional additional information (ignored).}
}
\description{
Extract Something From a CTD Object
}
\examples{
data(ctd)
head(ctd[["temperature"]])

}
\author{
Dan Kelley
}
\seealso{
Other functions that access oce data and metadata: \code{\link{[[,adp-method}},
  \code{\link{[[,adv-method}}, \code{\link{[[,argo-method}}

Other functions that handle CTD data: \code{\link{as.ctd}},
  \code{\link{ctd-class}}, \code{\link{ctdAddColumn}},
  \code{\link{ctdDecimate}}, \code{\link{ctdFindProfiles}},
  \code{\link{ctdTrim}}, \code{\link{ctdUpdateHeader}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.odv}}, \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}}, \code{\link{write.ctd}}
}

