% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adp.R
\docType{methods}
\name{[[<-,adp-method}
\alias{[[<-,adp-method}
\title{Change Something Within an adp Object}
\usage{
\S4method{[[}{adp}(x, i, j) <- value
}
\arguments{
\item{x}{An adp object}

\item{i}{The item to insert}

\item{j}{Optional additional information on the \code{i} item.}

\item{value}{The value to be inserted into \code{x}.}

\item{...}{Optional additional information (ignored).}
}
\description{
In addition to the usual insertion of elements by name, note
that e.g. \code{pitch} gets stored into \code{pitchSlow}.
}

