% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adp.R
\docType{methods}
\name{[[<-,adp-method}
\alias{[[<-,adp-method}
\title{Replace Parts of an \code{adp} Object}
\usage{
\S4method{[[}{adp}(x, i, j) <- value
}
\arguments{
\item{x}{An adp object}

\item{i}{The item to insert}

\item{j}{Optional additional information on the \code{i} item.}

\item{value}{The value to be inserted into \code{x}.}

\item{...}{Optional additional information (ignored).}
}
\description{
In addition to the usual insertion of elements by name, note
that e.g. \code{pitch} gets stored into \code{pitchSlow}.
}
\seealso{
Other things related to \code{adp} data: \code{\link{[[,adp-method}},
  \code{\link{adp-class}}, \code{\link{adp}},
  \code{\link{as.adp}}, \code{\link{subset,adp-method}},
  \code{\link{xyzToEnuAdp}}
}

