% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.R
\docType{methods}
\name{[[<-,adv-method}
\alias{[[<-,adv-method}
\title{Replace Parts of an \code{adv} Object}
\usage{
\S4method{[[}{adv}(x, i, j) <- value
}
\arguments{
\item{x}{An adv object}

\item{i}{The item to insert}

\item{j}{Optional additional information on the \code{i} item.}

\item{value}{The value to be inserted into \code{x}.}

\item{...}{Optional additional information (ignored).}
}
\description{
In addition to the usual insertion of elements by name, note
that e.g. \code{pitch} gets stored into \code{pitchSlow}.
}
\seealso{
Other functions that handle \code{adv} data: \code{\link{[[,adv-method}},
  \code{\link{plot,adv-method}},
  \code{\link{subset,adv-method}},
  \code{\link{summary,adv-method}}
}

