% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{[[<-,ctd-method}
\alias{[[<-,ctd-method}
\title{Replace Parts of a \code{ctd} Object}
\usage{
\S4method{[[}{ctd}(x, i, j) <- value
}
\arguments{
\item{x}{A ctd object}

\item{i}{The item to insert}

\item{j}{Optional additional information on the \code{i} item.}

\item{value}{The value to be inserted into \code{x}.}

\item{...}{Optional additional information (ignored).}
}
\description{
Replace Parts of a \code{ctd} Object
}
\seealso{
Other functions that alter oce data and metadata: \code{\link{[[<-,adp-method}}

Other functions that handle \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{as.ctd}}, \code{\link{ctdAddColumn}},
  \code{\link{ctdDecimate}}, \code{\link{ctdFindProfiles}},
  \code{\link{ctdTrim}}, \code{\link{ctdUpdateHeader}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}}, \code{\link{write.ctd}}
}

