% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coastline.R
\docType{methods}
\name{subset,coastline-method}
\alias{subset,coastline-method}
\title{Subset a Coastline Object}
\usage{
\S4method{subset}{coastline}(x, subset, ...)
}
\arguments{
\item{x}{A \code{coastline} object, i.e. one inheriting from \code{\link{coastline-class}}.}

\item{subset}{An expression indicating how to subset \code{x}. See \dQuote{Details}.}

\item{...}{Ignored.}
}
\value{
A \code{coastline} object.
}
\description{
Subsets a coastline object according to limiting values for longitude
and latitude. The \CRANpkg{raster} package must be installed for this
to work, because it relies on the \code{\link[raster]{intersect}} function
from that package.

As illustrated in the \dQuote{Examples}, \code{subset} must be an
expression that indicates limits on \code{latitude} and
\code{longitude}. The individual elements are provided in R notation,
not mathematical notation (i.e. \code{30<latitude<60} would not work).
Ampersands must be used to combine the limits.  The simplest way
to understand this is to copy the example directly, and then modify
the stated limits. Note that \code{>} comparison is not permitted,
and that \code{<} is converted to \code{<=} in the calculation.
Similarly, \code{&&} is converted to \code{&}. Spaces in the
expression are ignored. For convenience, \code{longitude} and
and \code{latitude} may be abbreviated as \code{lon} and \code{lat},
as in the \dQuote{Examples}.
}
\examples{
\donttest{
data(coastlineWorld)
cl <- subset(coastlineWorld, -70<lon & lon<-40 & 30<lat & lat<60)
plot(cl)
}
}
\seealso{
Other things related to \code{coastline} data: \code{\link{[[,coastline-method}},
  \code{\link{[[<-,coastline-method}},
  \code{\link{as.coastline}},
  \code{\link{coastline-class}},
  \code{\link{coastlineBest}}, \code{\link{coastlineCut}},
  \code{\link{coastlineWorld}},
  \code{\link{download.coastline}},
  \code{\link{plot,coastline-method}},
  \code{\link{read.coastline.openstreetmap}},
  \code{\link{read.coastline.shapefile}},
  \code{\link{summary,coastline-method}}

Other functions that subset \code{oce} objects: \code{\link{subset,adp-method}},
  \code{\link{subset,adv-method}},
  \code{\link{subset,amsr-method}},
  \code{\link{subset,argo-method}},
  \code{\link{subset,cm-method}},
  \code{\link{subset,ctd-method}},
  \code{\link{subset,echosounder-method}},
  \code{\link{subset,lobo-method}},
  \code{\link{subset,met-method}},
  \code{\link{subset,oce-method}},
  \code{\link{subset,odf-method}},
  \code{\link{subset,rsk-method}},
  \code{\link{subset,sealevel-method}},
  \code{\link{subset,section-method}},
  \code{\link{subset,topo-method}}
}
\author{
Dan Kelley
}
\concept{functions that subset \code{oce} objects}
\concept{things related to \code{coastline} data}
