% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobo.R
\docType{methods}
\name{subset,lobo-method}
\alias{subset,lobo-method}
\title{Subset an lobo object}
\usage{
\S4method{subset}{lobo}(x, subset, ...)
}
\arguments{
\item{x}{a \code{lobo} object.}

\item{subset}{a condition to be applied to the \code{data} portion of
\code{x}.  See \sQuote{Details}.}

\item{\dots}{ignored.}
}
\value{
A new \code{lobo} object.
}
\description{
Subset an lobo object, in a way that is somewhat
analogous to \code{\link{subset.data.frame}}.
}
\author{
Dan Kelley
}
\seealso{
Other functions that handle \code{lobo} data: \code{\link{[[,lobo-method}},
  \code{\link{[[<-,lobo-method}}, \code{\link{as.lobo}},
  \code{\link{plot,lobo-method}},
  \code{\link{summary,lobo-method}}
}

