% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.R
\docType{methods}
\name{summary,adv-method}
\alias{summary,adv-method}
\title{Summarize an ADV object}
\usage{
\S4method{summary}{adv}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"adv"}, usually, a result of a call to
\code{\link{read.adv}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summarize data in an \code{adv} object.
}
\examples{
library(oce)
data(adv)
summary(adv)

}
\author{
Dan Kelley
}
\seealso{
Other functions that handle \code{adv} data: \code{\link{[[,adv-method}},
  \code{\link{[[<-,adv-method}},
  \code{\link{plot,adv-method}},
  \code{\link{subset,adv-method}},
  \code{\link{xyzToEnuAdv}}
}

