% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amsr.R
\docType{methods}
\name{summary,amsr-method}
\alias{summary,amsr-method}
\alias{summary.amsr}
\title{Summarize an amsr Object}
\usage{
\S4method{summary}{amsr}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Ignored.}
}
\description{
Although the data are stored in \code{\link{raw}} form, the summary
presents results in physical units.
}
\author{
Dan Kelley
}
\seealso{
Other functions.dealing.with.satellite.data: \code{\link{[[,amsr-method}};
  \code{\link{[[,landsat-method}};
  \code{\link{amsr-class}}; \code{\link{g1sst-class}};
  \code{\link{landsat-class}},
  \code{\link{satellite-class}},
  \code{\link{satellite-class}}; \code{\link{landsatAdd}};
  \code{\link{landsatTrim}};
  \code{\link{plot,amsr-method}}, \code{\link{plot.amsr}};
  \code{\link{plot,landsat-method}},
  \code{\link{plot.landsat}};
  \code{\link{plot,satellite-method}},
  \code{\link{plot.satellite}}; \code{\link{read.amsr}};
  \code{\link{read.g1sst}}; \code{\link{read.landsat}};
  \code{\link{summary,landsat-method}},
  \code{\link{summary.landsat}};
  \code{\link{summary,satellite-method}},
  \code{\link{summary.g1sst}},
  \code{\link{summary.satellite}}
}
\concept{
satellite
}

