% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\name{summary,amsr-method}
\alias{summary,amsr-method}
\title{Summarize an amsr Object}
\usage{
\S4method{summary}{amsr}(object, ...)
}
\arguments{
\item{object}{an \linkS4class{amsr} object.}

\item{...}{ignored.}
}
\description{
Although the data are stored in \code{\link[=raw]{raw()}} form, the summary
presents results in physical units.
}
\seealso{
Other things related to amsr data: 
\code{\link{[[,amsr-method}},
\code{\link{[[<-,amsr-method}},
\code{\link{amsr-class}},
\code{\link{amsr}},
\code{\link{composite,amsr-method}},
\code{\link{download.amsr}()},
\code{\link{plot,amsr-method}},
\code{\link{read.amsr}()},
\code{\link{subset,amsr-method}}
}
\author{
Dan Kelley
}
\concept{things related to amsr data}
