% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\docType{methods}
\name{summary,amsr-method}
\alias{summary,amsr-method}
\alias{summary.amsr}
\title{Summarize an amsr object}
\usage{
\S4method{summary}{amsr}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Ignored.}
}
\description{
Although the data are stored in \code{\link{raw}} form, the summary
presents results in physical units.
}
\author{
Dan Kelley
}
\seealso{
Other functions dealing with satellite data: \code{\link{[[,landsat-method}},
  \code{\link{amsr-class}}, \code{\link{g1sst-class}},
  \code{\link{landsatAdd}}, \code{\link{landsatTrim}},
  \code{\link{plot,amsr-method}},
  \code{\link{plot,landsat-method}},
  \code{\link{plot,satellite-method}},
  \code{\link{read.amsr}}, \code{\link{read.g1sst}},
  \code{\link{read.landsat}},
  \code{\link{satellite-class}},
  \code{\link{summary,landsat-method}},
  \code{\link{summary,satellite-method}}
}
\concept{
satellite
}

