% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/amsr.R
\docType{methods}
\name{summary,amsr-method}
\alias{summary,amsr-method}
\alias{summary.amsr}
\title{Summarize an amsr object}
\usage{
\S4method{summary}{amsr}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Ignored.}
}
\description{
Although the data are stored in \code{\link{raw}} form, the summary
presents results in physical units.
}
\author{
Dan Kelley
}
\concept{
satellite data
}

