% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsr.R
\docType{methods}
\name{summary,amsr-method}
\alias{summary,amsr-method}
\title{Summarize an AMSR Object}
\usage{
\S4method{summary}{amsr}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Ignored.}
}
\description{
Although the data are stored in \code{\link{raw}} form, the summary
presents results in physical units.
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{amsr} data: \code{\link{[[<-,amsr-method}},
  \code{\link{amsr-class}}, \code{\link{plot,amsr-method}},
  \code{\link{read.amsr}}, \code{\link{subset,amsr-method}}
}
\concept{
satellite
}

