% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremen.R
\docType{methods}
\name{summary,bremen-method}
\alias{summary,bremen-method}
\title{Summarize a Bremen Object}
\usage{
\S4method{summary}{bremen}(object, ...)
}
\arguments{
\item{object}{a \linkS4class{bremen} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Pertinent summary information is presented, including the station name,
sampling location, data ranges, etc.
}
\seealso{
Other things related to bremen data: \code{\link{[[,bremen-method}},
  \code{\link{[[<-,bremen-method}},
  \code{\link{bremen-class}},
  \code{\link{plot,bremen-method}},
  \code{\link{read.bremen}}
}
\author{
Dan Kelley
}
\concept{things related to bremen data}
