% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm.R
\name{summary,cm-method}
\alias{summary,cm-method}
\title{Summarize a CM Object}
\usage{
\S4method{summary}{cm}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{cm} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in a \code{cm} object, presenting such information
as the station name, sampling location, data ranges, etc.
}
\examples{
library(oce)
data(cm)
summary(cm)

}
\seealso{
The documentation for the \linkS4class{cm} class explains the structure
of \code{cm} objects, and also outlines the other functions dealing with them.

Other things related to cm data: 
\code{\link{[[,cm-method}},
\code{\link{[[<-,cm-method}},
\code{\link{as.cm}()},
\code{\link{cm-class}},
\code{\link{cm}},
\code{\link{plot,cm-method}},
\code{\link{read.cm}()},
\code{\link{rotateAboutZ}()},
\code{\link{subset,cm-method}}
}
\author{
Dan Kelley
}
\concept{things related to cm data}
