% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\docType{methods}
\name{summary,ctd-method}
\alias{summary,ctd-method}
\alias{summary.ctd}
\title{Summarize a \code{ctd} Object}
\usage{
\S4method{summary}{ctd}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"ctd"}, usually, a result of a call to
\code{\link{read.ctd}}, \code{\link{read.oce}}, or \code{\link{as.ctd}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in a \code{ctd} object, presenting such information
as the station name, sampling location, data ranges, etc.
}
\examples{
library(oce)
data(ctd)
summary(ctd)

}
\author{
Dan Kelley
}
\seealso{
The documentation for \code{\link{ctd-class}} explains the structure
of CTD objects, and also outlines the other functions dealing with them.

Other functions that handle \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{ctdAddColumn}}, \code{\link{ctdDecimate}},
  \code{\link{ctdFindProfiles}}, \code{\link{ctdTrim}},
  \code{\link{ctdUpdateHeader}},
  \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}},
  \code{\link{subset,ctd-method}}, \code{\link{write.ctd}}
}

