% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echosounder.R
\docType{methods}
\name{summary,echosounder-method}
\alias{summary,echosounder-method}
\title{Summarize an Echosounder Object

Summarizes some of the data in an \code{echosounder} object.}
\usage{
\S4method{summary}{echosounder}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"echosounder"}, usually, a result of
a call to \code{\link{read.echosounder}}, \code{\link{read.oce}}, or
\code{\link{as.echosounder}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{echosounder} data: \code{\link{[[,echosounder-method}},
  \code{\link{[[<-,echosounder-method}},
  \code{\link{as.echosounder}},
  \code{\link{echosounder-class}},
  \code{\link{echosounder}}, \code{\link{findBottom}},
  \code{\link{plot,echosounder-method}},
  \code{\link{read.echosounder}},
  \code{\link{subset,echosounder-method}}
}

