% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\docType{methods}
\name{summary,gps-method}
\alias{summary,gps-method}
\title{Summarize a gps object}
\usage{
\S4method{summary}{gps}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"gps"}}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Summarize a \code{gps} object, i.e. one inheriting from \code{\link{gps-class}}.
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{gps} data: \code{\link{[[,gps-method}},
  \code{\link{[[<-,gps-method}}, \code{\link{as.gps}},
  \code{\link{gps-class}}, \code{\link{plot,gps-method}},
  \code{\link{read.gps}}
}

