% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landsat.R
\docType{methods}
\name{summary,landsat-method}
\alias{summary,landsat-method}
\alias{summary.landsat}
\title{Summarize a landsat Object}
\usage{
\S4method{summary}{landsat}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{landsat-class}}, usually a result of a
call to \code{\link{read.landsat}}.}

\item{...}{Ignored.}
}
\description{
Provides a summary of a some information about an object of
\code{\link{landsat-class}}.
}
\author{
Dan Kelley
}
\seealso{
Other functions.dealing.with.satellite.data: \code{\link{[[,amsr-method}};
  \code{\link{[[,landsat-method}};
  \code{\link{amsr-class}}; \code{\link{g1sst-class}};
  \code{\link{landsat-class}},
  \code{\link{satellite-class}},
  \code{\link{satellite-class}}; \code{\link{landsatAdd}};
  \code{\link{landsatTrim}};
  \code{\link{plot,amsr-method}}, \code{\link{plot.amsr}};
  \code{\link{plot,landsat-method}},
  \code{\link{plot.landsat}};
  \code{\link{plot,satellite-method}},
  \code{\link{plot.satellite}}; \code{\link{read.amsr}};
  \code{\link{read.g1sst}}; \code{\link{read.landsat}};
  \code{\link{summary,amsr-method}},
  \code{\link{summary.amsr}};
  \code{\link{summary,satellite-method}},
  \code{\link{summary.g1sst}},
  \code{\link{summary.satellite}}
}
\concept{
satellite
}

