% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landsat.R
\docType{methods}
\name{summary,landsat-method}
\alias{summary,landsat-method}
\alias{summary.landsat}
\title{Summarize a landsat object.}
\usage{
\S4method{summary}{landsat}(object, ...)
}
\arguments{
\item{object}{An object of \code{\link{landsat-class}}, usually a result of a
call to \code{\link{read.landsat}}.}

\item{...}{Ignored.}
}
\description{
Provides a summary of a some information about an object of
\code{\link{landsat-class}}.
}
\author{
Dan Kelley
}
\concept{
satellite
}

