% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat.R
\docType{methods}
\name{summary,landsat-method}
\alias{summary,landsat-method}
\title{Summarize a landsat Object}
\usage{
\S4method{summary}{landsat}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{landsat} object.}

\item{...}{Ignored.}
}
\description{
Provides a summary of a some information about a \linkS4class{landsat} object.
}
\seealso{
Other things related to landsat data: \code{\link{[[,landsat-method}},
  \code{\link{[[<-,landsat-method}},
  \code{\link{landsat-class}}, \code{\link{landsatAdd}},
  \code{\link{landsatTrim}}, \code{\link{landsat}},
  \code{\link{plot,landsat-method}},
  \code{\link{read.landsat}}
}
\author{
Dan Kelley
}
\concept{satellite}
\concept{things related to landsat data}
