% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisst.R
\docType{methods}
\name{summary,lisst-method}
\alias{summary,lisst-method}
\title{Summarize a \code{lisst} Object}
\usage{
\S4method{summary}{lisst}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lisst}, usually, a result of a call to
\code{\link{read.lisst}} or \code{\link{as.lisst}}.}

\item{...}{Ignored.}
}
\description{
Summarizes some of the data in a \code{lisst} object, presenting such information
as the station name, sampling location, data ranges, etc.
}
\examples{
library(oce)
data(lisst)
summary(lisst)

}
\author{
Dan Kelley
}
\seealso{
Other functions that handle \code{lisst} data: \code{\link{[[,lisst-method}},
  \code{\link{[[<-,lisst-method}}, \code{\link{as.lisst}},
  \code{\link{plot,lisst-method}}, \code{\link{read.lisst}}
}

