% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\docType{methods}
\name{summary,met-method}
\alias{summary,met-method}
\title{Summarize a met object}
\usage{
\S4method{summary}{met}(object, ...)
}
\arguments{
\item{object}{A \code{met} object, i.e. one inheriting from \code{\link{met-class}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in a \code{met} object.
}
\details{
Pertinent summary information is presented, including the sampling location,
data ranges, etc.
}
\author{
Dan Kelley
}
\seealso{
Other things related to \code{met} data: \code{\link{as.met}},
  \code{\link{met-class}}, \code{\link{met}},
  \code{\link{plot,met-method}}, \code{\link{read.met}},
  \code{\link{subset,met-method}}
}

