% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/satellite.R
\docType{methods}
\name{summary,satellite-method}
\alias{summarize.g1sst}
\alias{summarize.satellite}
\alias{summary,satellite-method}
\title{Summarize a satellite object}
\usage{
\S4method{summary}{satellite}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Ignored.}
}
\description{
Summarize a satellite object
}
\author{
Dan Kelley
}
\concept{
satellite
}

