% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.R
\docType{methods}
\name{summary,satellite-method}
\alias{summary,satellite-method}
\alias{summary.g1sst}
\alias{summary.satellite}
\title{Summarize a satellite object}
\usage{
\S4method{summary}{satellite}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Ignored.}
}
\description{
Summarize a satellite object
}
\author{
Dan Kelley
}
\seealso{
Other functions that handle satellite data: \code{\link{plot,satellite-method}},
  \code{\link{satellite-class}}
}
\concept{
satellite
}

