% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellite.R
\docType{methods}
\name{summary,satellite-method}
\alias{summary,satellite-method}
\title{Summarize a satellite object}
\usage{
\S4method{summary}{satellite}(object, ...)
}
\arguments{
\item{object}{The object to be summarized.}

\item{...}{Ignored.}
}
\description{
Summarize a satellite object
}
\seealso{
Other things related to satellite data: \code{\link{g1sst-class}},
  \code{\link{plot,satellite-method}},
  \code{\link{read.g1sst}}, \code{\link{satellite-class}}
}
\author{
Dan Kelley
}
\concept{satellite}
\concept{things related to satellite data}
