% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\docType{methods}
\name{summary,windrose-method}
\alias{summary,windrose-method}
\alias{summary.windrose}
\title{Summarize a \code{windrose} object}
\usage{
\S4method{summary}{windrose}(object, ...)
}
\arguments{
\item{object}{An \code{windrose} object, i.e. inheriting from \code{\link{windrose-class}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes some of the data in a \code{windrose} object.
}
\author{
Dan Kelley
}
\seealso{
The documentation for \code{\link{windrose-class}} explains the structure
of \code{windrose} objects, and also outlines the other functions dealing with them.

Other functions that handle \code{windrose} data: \code{\link{[[,windrose-method}},
  \code{\link{[[<-,windrose-method}},
  \code{\link{as.windrose}},
  \code{\link{plot,windrose-method}}
}

