\name{summary.bremen}

\alias{summary.bremen}
\alias{summary,bremen,missing-method}
\alias{summary,bremen-method}

\title{Summarize a bremen object}

\description{Summarize a \code{bremen} object}

\usage{\S4method{summary}{bremen}(object, \dots)}

\arguments{
  \item{object}{an object of class \code{"bremen"}, as read by
  \code{\link{read.bremen}}.}
  \item{\dots}{further arguments passed to or from other methods.}
}

\details{Pertinent summary information is presented.}

\value{\code{summary.bremen} returns an object of class \code{"summary.bremen"},
  containing pertinent information about the \code{bremen} object and its
  processing.}

\seealso{The documentation for \code{\link{bremen-class}} explains the structure of
    bremen objects, and also outlines the other functions dealing with them.}

\author{Dan Kelley}

\keyword{misc}
