\name{summary.rbrtdr}

\alias{summary.rbrtdr}

\title{Summarize an RBR temperature-depth data object}

\description{Summarize an RBR temperature-depth data object}

\usage{\method{summary}{rbrtdr}(object, \dots)}

\arguments{
  \item{object}{A \code{rbrtdr} object, e.g. as read by \code{\link{read.rbrtdr}}.}
  \item{\dots}{passed to children.}
}

\details{Pertinent summary information is presented, including the
		     start and stop times, as well as statistics of
		     temperature and pressure.}

\value{None}

\seealso{A \code{rbrtdr} object may be read with \code{\link{read.rbrtdr}}
and plotted with \code{\link{read.rbrtdr}}.}

\examples{
\dontrun{
library(oce)
...
}}

\references{(fill in later)}

\author{Dan Kelley}

\keyword{misc}
