% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw.R
\name{swSigma}
\alias{swSigma}
\title{Seawater density anomaly}
\usage{
swSigma(salinity, temperature = NULL, pressure = NULL, longitude = 300,
  latitude = 30, eos = getOption("oceEOS", default = "gsw"))
}
\arguments{
\item{salinity}{either practical salinity (in which case \code{temperature}
and \code{pressure} must be provided) \strong{or} an \code{oce} object, in
which case \code{salinity}, \code{temperature} (in the ITS-90 scale; see
next item), etc. are inferred from the object.}

\item{temperature}{\emph{in-situ} temperature [\eqn{^\circ}{deg}C], defined
on the ITS-90 scale.  This scale is used by GSW-style calculation (as
requested by setting \code{eos="gsw"}), and is the value contained within
\code{ctd} objects (and probably most other objects created with data
acquired in the past decade or two). Since the UNESCO-style calculation is
based on IPTS-68, the temperature is converted within the present function,
using \code{\link{T68fromT90}}.}

\item{pressure}{pressure [dbar]}

\item{longitude}{longitude of observation (only used if \code{eos="gsw"};
see \sQuote{Details}).}

\item{latitude}{latitude of observation (only used if \code{eos="gsw"}; see
\sQuote{Details}).}

\item{eos}{equation of state, either \code{"unesco"} [1,2] or \code{"gsw"}
[3,4].}
}
\value{
Density anomaly [kg/m\eqn{^3}{^3}], defined as \code{\link{swRho}} -
1000 kg/m\eqn{^3}{^3}.
}
\description{
Compute \eqn{\sigma_\theta}{sigma}, the density of seawater, minus 1000
kg/m\eqn{^3}{^3}.
}
\examples{
library(oce)
swSigma(35, 13, 1000, eos="gsw")                   # 30.82397
swSigma(35, T90fromT68(13), 1000, eos="unesco")    # 30.8183

}
\author{
Dan Kelley
}
\references{
See citations provided in the \code{\link{swRho}} documentation.
}
\seealso{
Other functions that calculate seawater properties: \code{\link{swAbsoluteSalinity}},
  \code{\link{swAlphaOverBeta}}, \code{\link{swAlpha}},
  \code{\link{swBeta}}, \code{\link{swCSTp}},
  \code{\link{swConservativeTemperature}},
  \code{\link{swDepth}}, \code{\link{swDynamicHeight}},
  \code{\link{swLapseRate}}, \code{\link{swN2}},
  \code{\link{swPressure}}, \code{\link{swRho}},
  \code{\link{swRrho}}, \code{\link{swSCTp}},
  \code{\link{swSTrho}}, \code{\link{swSigma0}},
  \code{\link{swSigma1}}, \code{\link{swSigma2}},
  \code{\link{swSigma3}}, \code{\link{swSigma4}},
  \code{\link{swSigmaTheta}}, \code{\link{swSigmaT}},
  \code{\link{swSoundAbsorption}},
  \code{\link{swSoundSpeed}}, \code{\link{swSpecificHeat}},
  \code{\link{swSpice}}, \code{\link{swTFreeze}},
  \code{\link{swTSrho}},
  \code{\link{swThermalConductivity}},
  \code{\link{swTheta}}, \code{\link{swViscosity}},
  \code{\link{swZ}}
}

