% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw.R
\name{swSpecificHeat}
\alias{swSpecificHeat}
\title{Seawater specific heat
Source= http://sam.ucsd.edu/sio210/propseawater/ppsw_fortran/ppsw.f
check value: cpsw = 3849.500 j/(kg deg. c) for s = 40 (ipss-78),}
\usage{
swSpecificHeat(salinity, temperature = NULL, pressure = 0,
  longitude = 300, latitude = 30, eos = getOption("oceEOS", default =
  "gsw"))
}
\arguments{
\item{salinity}{either practical salinity (in which case \code{temperature}
and \code{pressure} must be provided) \strong{or} an \code{oce} object (in
which case \code{salinity}, etc. are inferred from the object).}

\item{temperature}{\emph{in-situ} temperature [\eqn{^\circ}{deg}C], defined
on the ITS-90 scale.}

\item{pressure}{seawater pressure [dbar]}

\item{longitude}{longitude of observation (only used if \code{eos="gsw"};
see \sQuote{Details}).}

\item{latitude}{latitude of observation (only used if \code{eos="gsw"}; see
\sQuote{Details}).}

\item{eos}{equation of state, either \code{"unesco"} or \code{"gsw"}.}
}
\value{
Specific heat \eqn{J kg^{-1}\,^\circ C^{-1}}{J/(kg degC)}
}
\description{
Compute specific heat of seawater.
}
\details{
If the first argument is a \code{ctd} object, then salinity, etc, are
extracted from it, and used for the calculation.
}
\examples{
swSpecificHeat(40, T90fromT68(40), 10000, eos="unesco") # 3949.499

}
\references{
Millero et. al., J. Geophys. Res. 78 (1973), 4499-4507

Millero et. al., UNESCO report 38 (1981), 99-188.
}
\seealso{
Other functions that calculate seawater properties: \code{\link{T68fromT90}},
  \code{\link{T90fromT48}}, \code{\link{T90fromT68}},
  \code{\link{swAbsoluteSalinity}},
  \code{\link{swAlphaOverBeta}}, \code{\link{swAlpha}},
  \code{\link{swBeta}}, \code{\link{swCSTp}},
  \code{\link{swConservativeTemperature}},
  \code{\link{swDepth}}, \code{\link{swDynamicHeight}},
  \code{\link{swLapseRate}}, \code{\link{swN2}},
  \code{\link{swPressure}}, \code{\link{swRho}},
  \code{\link{swRrho}}, \code{\link{swSCTp}},
  \code{\link{swSTrho}}, \code{\link{swSigma0}},
  \code{\link{swSigma1}}, \code{\link{swSigma2}},
  \code{\link{swSigma3}}, \code{\link{swSigma4}},
  \code{\link{swSigmaTheta}}, \code{\link{swSigmaT}},
  \code{\link{swSigma}}, \code{\link{swSoundAbsorption}},
  \code{\link{swSoundSpeed}}, \code{\link{swSpice}},
  \code{\link{swTFreeze}}, \code{\link{swTSrho}},
  \code{\link{swThermalConductivity}},
  \code{\link{swTheta}}, \code{\link{swViscosity}},
  \code{\link{swZ}}
}
\author{
Dan Kelley
}
