% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oce.R
\name{tail.oce}
\alias{tail.oce}
\title{Extract the end of an \code{oce} object}
\usage{
\method{tail}{oce}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{oce} object.}

\item{n}{Number of elements to extract.}

\item{...}{ignored}
}
\description{
This only works for \code{adp} and \code{adv} objects.
}
\seealso{
\code{\link{head.oce}}, which yields the start of an \code{oce} object.
}

