% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{threenum}
\alias{threenum}
\title{Calculate min, mean, and max values}
\usage{
threenum(x)
}
\arguments{
\item{x}{a vector or matrix of numerical values.}
}
\value{
A character vector of four values: the minimum, the mean, the
maximum, and an indication of the number of data.
}
\description{
This is a faster cousin of the standard \code{\link{fivenum}} function,
used in generic \code{summary} functions for \code{oce} objects.
}
\examples{
library(oce)
threenum(1:10)
}
\author{
Dan Kelley
}
