\name{tidem-class}

\docType{class}

\alias{tidem-class}
\alias{[[,tidem-method}

\title{Class to store tidal-constituent models}

\description{Class to store tidal-constituent models, with standard slots
    \code{metadata}, \code{data} and \code{processingLog}.}

\author{Dan Kelley}
\keyword{classes}
\keyword{oce}

