% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidem.R
\name{tidemAstron}
\alias{tidemAstron}
\title{Do ephemeris calculations for tidem}
\usage{
tidemAstron(t)
}
\arguments{
\item{t}{time in \code{POSIXct} format.  (It is \strong{very} important to
use \code{tz="GMT"} in constructing \code{t}.)}
}
\value{
A \code{\link[base]{data.frame}} containing
\item{list("astro")}{(fill in later)} \item{list("ader")}{(fill in later)}
}
\description{
Do ephemeris calculations for tidem.
}
\details{
Based directly on \code{t_astron}, from the \code{T_TIDE} package.
}
\examples{
tidemAstron(as.POSIXct("2008-01-22 18:50:24"))
}
\author{
Dan Kelley, based directly on \code{t_astron} from the \code{T_TIDE}
package.
}
\seealso{
\code{\link{tidem}}, which uses this.

Other things related to \code{tidem} data: \code{\link{[[,tidem-method}},
  \code{\link{[[<-,tidem-method}},
  \code{\link{plot,tidem-method}},
  \code{\link{predict.tidem}},
  \code{\link{summary,tidem-method}},
  \code{\link{tidedata}}, \code{\link{tidem-class}},
  \code{\link{tidemVuf}}, \code{\link{tidem}}
}

