% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{titleCase}
\alias{titleCase}
\title{Capitalize first letter of each of a vector of words}
\usage{
titleCase(w)
}
\arguments{
\item{w}{vector of character strings}
}
\value{
vector of strings patterned on \code{w} but with first letter
in upper case and others in lower case
}
\description{
This is used in making labels for data names in some ctd functions
}
