% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv.R
\name{toEnuAdv}
\alias{toEnuAdv}
\title{Convert an ADV Object to ENU Coordinates}
\usage{
toEnuAdv(x, declination = 0, debug = getOption("oceDebug"))
}
\arguments{
\item{x}{An \code{adv} object, i.e. one inheriting from \code{\link{adv-class}}.}

\item{declination}{magnetic declination to be added to the heading, to get
ENU with N as "true" north.}

\item{debug}{An integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed.}
}
\description{
Convert an ADV Object to ENU Coordinates
}
\author{
Dan Kelley
}
\references{
\url{http://www.nortek-as.com/lib/forum-attachments/coordinate-transformation}
}
\seealso{
See \code{\link{read.adv}} for notes on functions relating to
\code{"adv"} objects.  Also, see \code{\link{beamToXyzAdv}} and
\code{\link{xyzToEnuAdv}}.

Other things related to \code{adv} data: \code{\link{[[,adv-method}},
  \code{\link{[[<-,adv-method}}, \code{\link{adv-class}},
  \code{\link{adv}}, \code{\link{beamName}},
  \code{\link{beamToXyz}}, \code{\link{enuToOtherAdv}},
  \code{\link{enuToOther}}, \code{\link{plot,adv-method}},
  \code{\link{read.adv.nortek}},
  \code{\link{read.adv.sontek.adr}},
  \code{\link{read.adv.sontek.serial}},
  \code{\link{read.adv.sontek.text}},
  \code{\link{read.adv}}, \code{\link{subset,adv-method}},
  \code{\link{summary,adv-method}}, \code{\link{toEnu}},
  \code{\link{velocityStatistics}},
  \code{\link{xyzToEnuAdv}}, \code{\link{xyzToEnu}}
}

