% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\docType{class}
\name{topo-class}
\alias{topo-class}
\title{Class to store topographic data}
\description{
Topographic data may be read with \code{\link{read.topo}} or assembled with
\code{\link{as.topo}}.  Plotting are handled with \code{\link{plot,topo-method}}
and summaries with \code{\link{summary,topo-method}}. Data retrieval may be
done with \code{\link{[[,topo-method}} and replacement with
\code{\link{[[<-,topo-method}}.  The key data, stored within the data slot,
are: \code{longititude}, \code{latitude}, and \code{z}.
}
\author{
Dan Kelley
}
\seealso{
Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{coastline-class}}, \code{\link{ctd-class}},
  \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{met-class}}, \code{\link{odf-class}},
  \code{\link{rsk-class}}, \code{\link{sealevel-class}},
  \code{\link{section-class}}, \code{\link{windrose-class}}

Other things related to \code{topo} data: \code{\link{[[,topo-method}},
  \code{\link{[[<-,topo-method}}, \code{\link{as.topo}},
  \code{\link{plot,topo-method}}, \code{\link{read.topo}},
  \code{\link{subset,topo-method}},
  \code{\link{summary,topo-method}},
  \code{\link{topoInterpolate}}, \code{\link{topoWorld}}
}

