\name{topo.maritimes}

\docType{data}

\alias{topo.maritimes}

\title{Sample topographic dataset, showing Maritime region of Canada.}

\description{Topography of the Gulf of St Lawrence region, including
  portions of the Canadian provinces known as the Maritimes.}

\usage{
data(topo.maritimes)
}

\source{The dataset was downloaded in early 2008 from the NOAA website
  \url{http://www.ngdc.noaa.gov/mgg/gdas/gd_designagrid.html}, with 
  a longitude box from 70W to 55W and a latitude box from 45.5N to 53N.}

\references{\url{http://www.ngdc.noaa.gov/mgg/gdas/gd_designagrid.html}}

\seealso{Files may be read with \code{\link{read.topo}} and
  summarized with \code{\link{summary.topo}}}

\author{Dan Kelley}

\keyword{misc}
