% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trimString}
\alias{trimString}
\title{Remove leading and trailing whitespace from strings}
\usage{
trimString(s)
}
\arguments{
\item{s}{vector of character strings}
}
\value{
a new vector formed by trimming leading and trailing whitespace
from the elements of \code{s}.
}
\description{
Remove leading and trailing whitespace from strings
}
