% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unduplicateNames}
\alias{unduplicateNames}
\title{Rename duplicated items (used in reading CTD files) [local function]}
\usage{
unduplicateNames(names)
}
\arguments{
\item{names}{Vector of strings with variable names.}
}
\value{
names Vector of strings with numbered variable names.
}
\description{
Rename duplicated items (used in reading CTD files) [local function]
}

