% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unduplicateNames}
\alias{unduplicateNames}
\title{Rename duplicated items (used in reading CTD files)}
\usage{
unduplicateNames(names)
}
\arguments{
\item{names}{Vector of strings with variable names.}
}
\value{
names Vector of strings with numbered variable names.
}
\description{
Rename items to avoid name collision, by appending a \code{2} to
the second occurrence of a name, etc.
}
\examples{
unduplicateNames(c("a", "b", "a", "c", "b"))
}
\seealso{
used by \code{\link{read.ctd.sbe}}.
}

