% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidem.R
\name{webtide}
\alias{webtide}
\title{Get a Tidal Prediction from a WebTide Database}
\source{
The WebTide software may be downloaded for free at the
Department of Fisheries and Oceans (Canada) website, which in February 2016
was
\code{http://www.bio.gc.ca/science/research-recherche/ocean/webtide/index-en.php},
although this site seems not to be particularly static.
}
\usage{
webtide(action = c("map", "predict"), longitude, latitude, node, time,
  basedir = getOption("webtide"), region = "nwatl", plot = TRUE, tformat,
  debug = getOption("oceDebug"), ...)
}
\arguments{
\item{action}{An indication of the action, either \code{action="map"} to
draw a map or \code{action="predict"} to get a prediction; see
\sQuote{Details}.}

\item{longitude, latitude}{optional location at which prediction is required (ignored if
\code{node} is given).}

\item{node}{optional integer specifying which node to look up; if this is given,
then neither \code{latitude} nor \code{longitude} may be given.}

\item{time}{times at which prediction is to be made.  If not supplied, this
will be the week starting at the present time, incrementing by 15 minutes.}

\item{basedir}{directory containing the \code{WebTide} application.}

\item{region}{database region, given as a directory name in the WebTide
directory.  For example, \code{h3o} is for Halifax Harbour, \code{nwatl} is
for the northwest Atlantic, and \code{sshelf} is for the Scotian Shelf and
Gulf of Maine.}

\item{plot}{boolean indicating whether to plot.}

\item{tformat}{optional argument passed to \code{\link{oce.plot.ts}}, for
plot types that call that function.  (See \code{\link{strptime}} for the
format used.)}

\item{debug}{an integer specifying whether debugging information is
to be printed during the processing. This is a general parameter that
is used by many \code{oce} functions. Generally, setting \code{debug=0}
turns off the printing, while higher values suggest that more information
be printed. If one function calls another, it usually reduces the value of
\code{debug} first, so that a user can often obtain deeper debugging
by specifying higher \code{debug} values.}

\item{\dots}{optional arguments passed to plotting functions. A common
example is to set \code{xlim} and \code{ylim}, to focus a map region.}
}
\value{
If \code{action="map"} the return value is a
list containing the index of the nearest node, along with the
\code{latitude} and \code{longitude} of that node.  If
\code{plot} is \code{FALSE}, this value is returned invisibly.

If \code{action="predict"}, the return value is a list containing a vector
of times (\code{time}), as well as vectors of the predicted \code{elevation}
in metres and the predicted horizontal components of velocity, \code{u} and
\code{v}, along with the \code{node} number, and the \code{basedir} and
\code{region} as supplied to this function.
If \code{plot} is \code{FALSE}, this value is returned invisibly.
}
\description{
Get a tidal prediction from a WebTide database. There are two distinct cases.

\emph{Case 1:} \code{action="map"}. In this case, if
\code{plot} is \code{FALSE}, a list is returned, containing
all the \code{node}s in the selected database, along with all
the \code{latitude}s and \code{longitude}s. This value is
also returned (silently) if \code{plot} is true, but in that case,
a plot is drawn to indicate the node locations. If \code{latitude} and
\code{longitude} are given, then the node nearest that spot is indicated on
the map; otherwise, if \code{node} is given, then the location of that
node is indicated. There is also a special case: if \code{node} is negative
and \code{interactive()} is \code{TRUE},
then \code{\link{locator}} is called, and the node nearest the spot
where the user clicks the mouse is indicated in the plot and in the
return value.

\code{Case 2:} \code{action="predict"}. If \code{plot} is \code{FALSE},
then a list is returned, indicating \code{time}, predicted
\code{elevation}, velocity components \code{u} and \code{v},
\code{node} number, the name of the \code{basedir}, and
the \code{region}. If \code{plot} is \code{TRUE}, this list is returned
silently, and time-series plots are drawn for elevation, u, and v.

Naturally, \code{webtide} will not work unless WebTide has been installed on
the computer.
}
\section{Caution}{

WebTide is not an open-source application, so the present function was
designed based on little more than guesses about the WebTide file structure.
Users should be on the lookout for odd results.
}

\examples{
\dontrun{
library(oce)
lon <- -69.61
lat <- 48.14
prediction <- webtide("predict", longitude=lon, latitude=lat)
mtext(sprintf("prediction at \%fN \%fE", lat, lon), line=0.75, side=3)
}
}
\author{
Dan Kelley
}
