\name{wind}
\docType{data}
\alias{wind}
\title{Wind dataset}
\description{Wind data.}
\usage{data(wind)}

\format{Columns \code{wind$x} and \code{wind$y} are location, while
  \code{wind$z} is the wind speed, in m/s.}

\source{The data are taken from Figure 5 of Koch et al. (1983).}

\references{S. E.  Koch and M.  DesJardins and P. J. Kocin, 1983.  ``An
interactive Barnes objective map anlaysis scheme for use with satellite
and conventional data,'' \emph{J.  Climate Appl.  Met.}, vol 22,
p. 1487-1503.}

\seealso{See \code{\link{interpBarnes}}, which uses the present dataset as a
    example.}

\author{Dan Kelley}

\keyword{datasets}
