% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\docType{data}
\name{wind}
\alias{wind}
\title{Wind dataset}
\description{
Wind data inferred from Figure 5 of Koch et al. (1983), provided to
illustrate the \code{\link[=interpBarnes]{interpBarnes()}} function.
Columns \code{wind$x} and \code{wind$y} are location, while \code{wind$z} is
the wind speed, in m/s.
}
\references{
S. E.  Koch and M.  DesJardins and P. J. Kocin, 1983.  ``An
interactive Barnes objective map analysis scheme for use with satellite and
conventional data,'' \emph{J.  Climate Appl.  Met.}, vol 22, p. 1487-1503.
}
\seealso{
Other datasets provided with oce: 
\code{\link{adp}},
\code{\link{adv}},
\code{\link{argo}},
\code{\link{cm}},
\code{\link{coastlineWorld}},
\code{\link{ctdRaw}},
\code{\link{ctd}},
\code{\link{echosounder}},
\code{\link{landsat}},
\code{\link{lisst}},
\code{\link{lobo}},
\code{\link{met}},
\code{\link{ocecolors}},
\code{\link{rsk}},
\code{\link{sealevelTuktoyaktuk}},
\code{\link{sealevel}},
\code{\link{section}},
\code{\link{topoWorld}},
\code{\link{xbt}}
}
\concept{datasets provided with oce}
