% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose.R
\docType{class}
\name{windrose-class}
\alias{windrose-class}
\title{Class to store windrose data}
\description{
Windrose objects store statistical information about winds, mainly for
plotting as "wind rose" plots (see \code{\link{plot,windrose-method}}.
There is no reading method, because there is no standard way to store
wind data in files; instead, \code{\link{as.windrose}} is provided
to construct \code{windrose} objects.  Data elements may be retrieved
with \code{\link{[[,windrose-method}} or replaced with
\code{\link{[[<-,windrose-method}}. Data summaries are provided with
\code{\link{summary,windrose-method}}.
}
\seealso{
Other classes provided by \code{oce}: \code{\link{adp-class}},
  \code{\link{adv-class}}, \code{\link{argo-class}},
  \code{\link{bremen-class}}, \code{\link{cm-class}},
  \code{\link{coastline-class}}, \code{\link{ctd-class}},
  \code{\link{echosounder-class}},
  \code{\link{lisst-class}}, \code{\link{lobo-class}},
  \code{\link{met-class}}, \code{\link{odf-class}},
  \code{\link{rsk-class}}, \code{\link{sealevel-class}},
  \code{\link{sealevel}}, \code{\link{section-class}},
  \code{\link{topo-class}}

Other things related to \code{windrose} data: \code{\link{[[,windrose-method}},
  \code{\link{[[<-,windrose-method}},
  \code{\link{as.windrose}},
  \code{\link{plot,windrose-method}},
  \code{\link{summary,windrose-method}}
}

