% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.woce.R
\name{woceNames2oceNames}
\alias{woceNames2oceNames}
\title{Translate WOCE Data Names to Oce Data Names}
\usage{
woceNames2oceNames(names)
}
\arguments{
\item{names}{vector of strings holding WOCE-style names.}
}
\value{
vector of strings holding \code{oce}-style names.
}
\description{
Translate WOCE-style names to \code{oce} names, using \code{\link{gsub}}
to match patterns. For example, the pattern \code{"CTDOXY.*"} is taken
to mean \code{oxygen}.
}
\references{
Several online sources list WOCE names. An example is
\url{https://geo.h2o.ucsd.edu/documentation/manuals/pdf/90_1/chap4.pdf}
}
\seealso{
Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{cnvName2oceName}}, \code{\link{ctd-class}},
  \code{\link{ctdDecimate}}, \code{\link{ctdFindProfiles}},
  \code{\link{ctdRaw}}, \code{\link{ctdTrim}},
  \code{\link{ctd}}, \code{\link{handleFlags,ctd-method}},
  \code{\link{plot,ctd-method}}, \code{\link{plotProfile}},
  \code{\link{plotScan}}, \code{\link{plotTS}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}}, \code{\link{write.ctd}}

Other functions that interpret variable names and units from headers: \code{\link{ODFNames2oceNames}},
  \code{\link{cnvName2oceName}},
  \code{\link{unitFromStringRsk}},
  \code{\link{unitFromString}}
}
\author{
Dan Kelley
}
