% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.woce.R
\name{woceUnit2oceUnit}
\alias{woceUnit2oceUnit}
\title{Translate WOCE units to oce units}
\usage{
woceUnit2oceUnit(woceUnit)
}
\arguments{
\item{woceUnit}{string holding a WOCE unit}
}
\value{
expression in oce unit form
}
\description{
Translate WOCE-style units to \code{oce} units.
}
\seealso{
Other things related to \code{ctd} data: \code{\link{[[,ctd-method}},
  \code{\link{[[<-,ctd-method}}, \code{\link{as.ctd}},
  \code{\link{cnvName2oceName}}, \code{\link{ctd-class}},
  \code{\link{ctdDecimate}}, \code{\link{ctdFindProfiles}},
  \code{\link{ctdRaw}}, \code{\link{ctdTrim}},
  \code{\link{ctd}}, \code{\link{handleFlags,ctd-method}},
  \code{\link{oceNames2whpNames}},
  \code{\link{oceUnits2whpUnits}},
  \code{\link{plot,ctd-method}}, \code{\link{plotProfile}},
  \code{\link{plotScan}}, \code{\link{plotTS}},
  \code{\link{read.ctd.itp}}, \code{\link{read.ctd.odf}},
  \code{\link{read.ctd.sbe}},
  \code{\link{read.ctd.woce.other}},
  \code{\link{read.ctd.woce}}, \code{\link{read.ctd}},
  \code{\link{subset,ctd-method}},
  \code{\link{summary,ctd-method}},
  \code{\link{woceNames2oceNames}}, \code{\link{write.ctd}}

Other functions that interpret variable names and units from headers: \code{\link{ODFNames2oceNames}},
  \code{\link{cnvName2oceName}},
  \code{\link{oceNames2whpNames}},
  \code{\link{oceUnits2whpUnits}},
  \code{\link{unitFromStringRsk}},
  \code{\link{unitFromString}},
  \code{\link{woceNames2oceNames}}
}
\author{
Dan Kelley
}
