% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{xbt.edf}
\alias{xbt.edf}
\title{Sample xbt dataset}
\description{
Sample xbt dataset
}
\examples{
\dontrun{
xbt <- read.oce(system.file("extdata", "xbt.edf", package="oce"))
}

}
\seealso{
Other raw datasets: 
\code{\link{CTD_BCD2014666_008_1_DN.ODF.gz}},
\code{\link{adp_rdi.000}},
\code{\link{ctd.cnv}},
\code{\link{d200321-001.ctd}},
\code{\link{d201211_0011.cnv}},
\code{\link{test_met_csv1.csv}},
\code{\link{test_met_csv2.csv}},
\code{\link{test_met_xml2.xml}}

Other things related to xbt data: 
\code{\link{[[<-,xbt-method}},
\code{\link{as.xbt}()},
\code{\link{plot,xbt-method}},
\code{\link{read.xbt.noaa1}()},
\code{\link{read.xbt}()},
\code{\link{subset,xbt-method}},
\code{\link{summary,xbt-method}},
\code{\link{xbt-class}}
}
\concept{raw datasets}
\concept{things related to xbt data}
