% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbt.R
\docType{data}
\name{xbt}
\alias{xbt}
\title{An XBT Object}
\usage{
data(xbt)
}
\description{
The result of using \code{\link[=read.xbt]{read.xbt()}} on a Sippican file created by extracting the near-surface
fraction of the sample provided in Section 5.5.6 of reference 1.
}
\examples{
library(oce)
data(xbt)
summary(xbt)
plot(xbt)

}
\references{
\enumerate{
\item Sippican, Inc. “Bathythermograph Data Acquisition System: Installation, Operation and Maintenance
Manual (P/N 308195, Rev. A),” 2003.
https://pages.uoregon.edu/drt/MGL0910_Science_Report/attachments/MK21_ISA_Manual_Rev_A.pdf.
\item Cheng, Lijing, John Abraham, Gustavo Goni, Timothy Boyer, Susan Wijffels, Rebecca
Cowley, Viktor Gouretski, et al. “XBT Science: Assessment of Instrumental Biases and Errors.”
Bulletin of the American Meteorological Society 97, no. 6 (June 2016): 924–33.
https://doi.org/10.1175/BAMS-D-15-00031.1.
}
}
\seealso{
Other datasets provided with oce: \code{\link{adp}},
  \code{\link{adv}}, \code{\link{argo}}, \code{\link{cm}},
  \code{\link{coastlineWorld}}, \code{\link{ctdRaw}},
  \code{\link{ctd}}, \code{\link{echosounder}},
  \code{\link{landsat}}, \code{\link{lisst}},
  \code{\link{lobo}}, \code{\link{met}},
  \code{\link{ocecolors}}, \code{\link{rsk}},
  \code{\link{sealevelTuktoyaktuk}},
  \code{\link{sealevel}}, \code{\link{section}},
  \code{\link{topoWorld}}, \code{\link{wind}}

Other things related to xbtdata: \code{\link{[[,xbt-method}}
}
\concept{datasets provided with oce}
\concept{things related to xbtdata}
